### 2023-10-10 #5936 @editorialbot
[PRE REVIEW]: hp3D: A Scalable MPI/OpenMP hp-Adaptive Finite Element Software Library for Complex Multiphysics Applications
**Submitting author:** @stefanhenneking (<a href="http://orcid.org/0000-0003-2177-8519">Stefan Henneking</a>)
**Repository:** https://github.com/Oden-EAG/hp3d
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0-beta
