### 2023-10-10 #5936 @stefanhenneking
Hi @kyleniemeyer, thank you for getting the pre-review started and for considering our paper at JOSS. Perhaps, @jedbrown is interested as an editor if available -- he'd suggested submitting to JOSS in the first place.

Yes, I will take care of the missing DOIs. For reviewers, are former contributors/collaborators of the software who are not co-authors of the paper eligible for reviewing?
