+++ paper/paper.bib
@@ -1,92 +1,103 @@
 
 @article{henneking2021fiber,
-	title = {Model and computational advancements to full vectorial {Maxwell} model for studying fiber amplifiers},
-	author = {Henneking, Stefan and Grosek, Jacob and Demkowicz, Leszek},
-	journal = {Comput. Math. Appl.},
-	volume = {85},
-	pages = {30--41},
-	year = {2021},
-	publisher = {Elsevier}}
+  title = {Model and computational advancements to full vectorial {Maxwell} model for studying fiber amplifiers},
+  author = {Henneking, Stefan and Grosek, Jacob and Demkowicz, Leszek},
+  journal = {Comput. Math. Appl.},
+  volume = {85},
+  pages = {30--41},
+  year = {2021},
+  publisher = {Elsevier},
+  doi = {10.1016/j.camwa.2021.01.006}}
 
 @article{henneking2021pollution,
-	Title = {A numerical study of the pollution error and {DPG} adaptivity for long waveguide simulations},
-	Author = {Henneking, Stefan and Demkowicz, Leszek},
-	Journal = {Comput. Math. Appl.},
-	Publisher = {Elsevier},
-	Volume = {95},
-	Pages = {85--100},
-	Year = {2021}}
+  Title = {A numerical study of the pollution error and {DPG} adaptivity for long waveguide simulations},
+  Author = {Henneking, Stefan and Demkowicz, Leszek},
+  Journal = {Comput. Math. Appl.},
+  Publisher = {Elsevier},
+  Volume = {95},
+  Pages = {85--100},
+  Year = {2021},
+  doi = {10.1016/j.camwa.2020.03.024}}
 
 @phdthesis{henneking2021phd,
-	author = {Henneking, Stefan},
-	school = {The University of Texas at Austin},
-	title = {A scalable $hp$-adaptive finite element software with applications in fiber optics},
-	year = {2021}}
+  author = {Henneking, Stefan},
+  school = {The University of Texas at Austin},
+  title = {A scalable $hp$-adaptive finite element software with applications in fiber optics},
+  year = {2021},
+  doi = {10.26153/tsw/13716}}
 
 @inproceedings{henneking2022parallel,
-	author = {Henneking, Stefan and Grosek, Jacob and Demkowicz, Leszek},
-	title = {Parallel Simulations of High-Power Optical Fiber Amplifiers},
-	booktitle = {Spectral and High Order Methods for Partial Differential Equations ICOSAHOM 2020+1},
-	year = {2022},
-	publisher = {Springer},
-	pages = {349--360}}
+  author = {Henneking, Stefan and Grosek, Jacob and Demkowicz, Leszek},
+  title = {Parallel Simulations of High-Power Optical Fiber Amplifiers},
+  booktitle = {Spectral and High Order Methods for Partial Differential Equations ICOSAHOM 2020+1},
+  year = {2022},
+  publisher = {Springer},
+  pages = {349--360},
+  doi = {10.1007/978-3-031-20432-6_22}}
 
 @article{badger2023scalable,
-	title = {Scalable {DPG} Multigrid Solver for {Helmholtz} Problems: A Study on Convergence},
-	author = {Badger, Jacob and Henneking, Stefan and Petrides, Socratis and Demkowicz, Leszek},
-	journal = {Comput. Math. Appl.},
-	volume = {148},
-	pages = {81--92},
-	year = {2023},
-	publisher = {Elsevier}}
+  title = {Scalable {DPG} Multigrid Solver for {Helmholtz} Problems: A Study on Convergence},
[…]
