+++ paper/paper.md
@@ -44,7 +44,7 @@ The $hp\mathrm{3D}$ finite element (FE) library is a tool for computational mode
 
 # Statement of need
 
-$hp\mathrm{3D}$ combines a list of unique features and algorithms setting it apart from other publicly available finite element libraries. The software supports hybrid meshes combining elements of "all shapes:" hexahedra, tetrahedra, prisms, and pyramids. The internal Geometric Modeling Package (GMP) provides support for \emph{exact geometry} elements and \emph{isoparametric} elements. Exact geometry elements directly use the parametrizations provided by GMP resulting in computations with no geometry error; isoparametric elements approximate the geometry maps with polynomials spanning the element space of shape functions. $hp\mathrm{3D}$'s shape functions package provides compatible discretization of energy spaces forming the $H^1-H(\text{curl})-H(\text{div})-L^2$ exact sequence [@fuentes2015shape; @demkowicz2020fem]. Additionally, the $hp\mathrm{3D}$ FE code sets itself apart from other advanced FE libraries (e.g., MFEM [@anderson2021mfem] or deal.II [@bangerth2007deal]) by focusing on \emph{$hp$-adaptive} solutions; in particular, the code supports \emph{anisotropic refinements} in both element size $h$ and polynomial order $p$. Such $hp$-adaptive methods are the most efficient way to converge to difficult solutions, e.g., resolving solutions with boundary layers, adapting toward geometric singularities, etc. [@chakraborty2023hp]. $hp\mathrm{3D}$ features a number of unique algorithms, including \emph{constrained approximation} routines for assembling elements with hanging nodes [@demkowicz1989-hp1; @demkowicz1989-hp2; @demkowicz1989-hp3], and \emph{projection-based interpolation} for computation of nodal constraints [@demkowicz2008interp]. Besides discretization with the standard Bubnov-Galerkin FE method, the $hp\mathrm{3D}$ library also supports discretization with the discontinuous Petrov-Galerkin (DPG) method [@demkowicz2017dpg]. The $hp\mathrm{3D}$ software leverages hybrid MPI/OpenMP parallelism to run efficiently on large-scale computing facilities [@henneking2021phd; @badger2023scalable].
[…]
