### 2023-10-12 #5936 @jedbrown
@likask @jeremiah-corrado @peterrum @kinnala @amartinhuertas :wave: Hi, would any of you be available to review this submission for JOSS?
