### 2023-10-28 #5946 @peterrum
@stefanhenneking  I went through the checklist and read the paper.

I didn't have any problems with installation after I have installed flex, bison, and libx11-dev; however, there are a bunch of warnings. I have run `/problems/POISSON/GALERKIN/`; it runs and solves the system; however, I had problems visualizing the output. The internal visualization tool `graphb` seems to be a bit old fashion; when selection `Paraview`, it apparently writes a file (`Dumping to Paraview...`) but I cannot find it. Could the authors point out where the output is (I have also tries the environment variable `PARAVIEW_DIR` as stated in the documentation). May I ask you to print the path to screen?

The text of the manuscript is overall clear and well written. I only have a few points:

> three-dimensional [...] problems

I found the limitation to 3D a bit irritating when reading the text. In the documentation, it says there is a 2D version, however, it is not public. Could you add the same comment in the manuscript. I hope both codes share functionalities?

> runs efficiently

Do you have any numbers or papers that verify this claim?

> state-of-the-art supercomputers.

I have tried the code on CPU. Is it also working on GPUs? If not, could you be more precise by stating `on various CPU-based compute architectures`.

> Additionally, the ℎp3D FE code sets itself apart from other advanced FE libraries (e.g., MFEM (Anderson et al., 2021) or deal.II (Bangerth et al., 2007)) by focusing on ℎ𝑝-adaptive solutions

I would suggest to relax this statement about other libraries. For instance, deal.II supports parallel matrix-based and matrix-free hp-adaptivity (with threading via TBB and MPI; see Bangerth, Kayser-Herold 2009 and Fehling, Bangerth 2023). "anisotropic refinement" and additional ways to apply constrains seems to be features that sets hp3D apart.
