### 2023-10-28 #5946 @likask
@stefanhenneking I have yet to test the code; however, I was able to install the code. The process will be much easier if you would add HP3D to spack—an easy job to do, which will make life easier.
