### 2023-11-02 #5946 @stefanhenneking
@peterrum @likask Thank you both very much for agreeing to review our submission and for your feedback on the paper and the code. We really appreciate it!

@peterrum thanks for pointing out the issue with the ParaView path. I'll add the output path to be printed on the screen by default. As you intuited correctly, `PARAVIEW_DIR` is the relative path to the output directory but if you ran one of the examples this path is set at runtime from the program arguments given in `run.sh` (or if not given, it uses the default argument set in `set_environment` of the corresponding problem implementation). I'll try to clarify that point or make it more intuitive with the next update of the user manual as well.

Yes, we do have a 2D version that shares most of the functionality with the 3D version. So far the 2D version hasn't been made public because it'd need some refactoring and additional documentation first. We would like to either make that 2D version public in the future or perhaps extend the 3D version to support 2D computations directly. I will add a comment along those lines in the paper.

Currently, only CPU-based architectures are supported. Thanks for bringing that up, I will change the sentence in the paper accordingly. The efficiency of the code is documented for fairly large problems (~O(10^9) dofs) in a few places - two papers are cited a bit further down in the paper in the "Statement of need" section.

I'll also relax the statement about other libraries in the way you suggested.

@likask I'm glad to hear you were able to install the library. Thank you for your suggestion to add hp3D to spack. I don't have any prior experience with that but I'll look into adding that to simplify installing hp3D and its dependencies.
