### 2023-11-06 #5946 @stefanhenneking
@likask @peterrum Thank you both again very much for reviewing our submission.

@jedbrown Thank you very much as well for serving as the editor for our submission (and also for encouraging us to submit to JOSS in the first place). What are the next steps?
