### 2023-11-12 #5946 @jedbrown
Hi, thanks for your work here. The paper reads well. I have two suggestions:
1. I would strongly encourage enabling continuous integration (GitHub Actions is pretty easy) so that at least some test are run automatically in pull requests. This has the advantage of creating a tested workflow from a base system to working build. Our reviewers here have a lot of experience with building/linking, but many other users and potential contributors will not and it's easy for documentation to miss necessary steps when it isn't tested.
2. The main point of friction feels like installation. The README doesn't currently document versions. Ideally, `apt install petsc-dev` (and a few more packages, or a Docker image with the latest release) would be enough for the build to work out of the box. If `PETSC_DIR` and `PETSC_ARCH` are set, the build should be able to use them. It looks like a regular build of PETSc (not installed to a prefix) will not work since it's currently hard-coded that there must be only one include path. (This restriction is especially clumsy if one is hacking on PETSc while testing it from hp3D.) You could probably make an `m_options` that gets all its information from `pkg-config` and even make that the default, which should make it possible to build in many environments without editing a file. `$PETSC_DIR/share/petsc/Makefile.user` shows how to get everything from pkg-config while staying symmetric with respect to other packages.
