### 2023-11-15 #5946 @stefanhenneking
Thanks for your suggestions, @jedbrown. I completely agree that both CI and simplifying installation are important missing pieces. I will try to work on it as soon as possible and link the corresponding Issues / PRs to this thread here.
