### 2023-11-15 Oden-EAG/hp3d#158 @stefanhenneking
Simplify installation of hp3D
See comments in JOSS paper review here:
https://github.com/openjournals/joss-reviews/issues/5946
