+++ trunk/src/graphics/x11/base/lsvisidb.F
@@ -1,3 +1,5 @@
+#if HP3D_USE_X11
+
 c----------------------------------------------------------------------
 c
 c   routine name       - lsvisidb
@@ -253,3 +255,5 @@ c
 c
 c
       end subroutine lsvisidb
+
+#endif
