### 2024-01-10 4e3b6c3
BCond Enconding modified to take only H1, Hdiv and Hcurl variables while setting up Dirichlet Bcoundary conditions
