+++ trunk/src/modules/physics.F90
@@ -37,12 +37,14 @@ module physics
   !
   !  ...number of entries in index (total number of components)
   integer, save :: NRINDEX
+   !  ...number of entries in index (total number of components)
+  integer, save :: NRINDEX_HEV
   !
   !  ...max number of component indices
   !     currently, 31 is the maximum number of components supported
   !     because the Dirichlet BC flags are binary-encoded per component
   !     into node%bcond which is an integer value
-  integer, parameter :: MAX_NRINDEX = 31
+  integer, parameter :: MAX_NRINDEX_HEV = 31
   !
   !----------------------------------------------------------------------
   !
