+++ trunk/problems/MAXWELL/ULTRAWEAK_DPG/celem.F90
@@ -39,7 +39,7 @@ subroutine celem(Mdle,Idec, Nrdofs,Nrdofm,Nrdofc,Nodm,NdofmH,NdofmE,NdofmV,Ndofm
       integer,                      intent(out) :: Nrnodm
       VTYPE  ,                      intent(out) :: Bload(*),Astif(*)
 !
-      integer, dimension(NRINDEX) :: nbcond
+      integer, dimension(NRINDEX_HEV) :: nbcond
 !
 !---------------------------------------------------------------------------------------------------------------
 !
@@ -50,7 +50,7 @@ subroutine celem(Mdle,Idec, Nrdofs,Nrdofm,Nrdofc,Nodm,NdofmH,NdofmE,NdofmV,Ndofm
 !
 !..remove bubble DOFs from trace components
    nbcond = 0; nbcond(1) = 1
-   call encod(nbcond,2,NRINDEX, NODES(Mdle)%bcond)
+   call encod(nbcond,2,NRINDEX_HEV, NODES(Mdle)%bcond)
 !
 !..redirect to the system routine
    call celem_system(Mdle,Idec,                            &
