### 2024-01-11 41fa19c
rolling back changes to run.sh and removing commented code from get_index.F90
