+++ trunk/problems/LASER/UW_COUPLED/elem/elem_maxwell_gram_hexa.F90
@@ -190,12 +190,9 @@
 !
 !..determine z-coordinate inside the element
    select case(etype)
-      case('mdlb')
+      case(MDLB)
          maxz = maxval(xnod(3,1:8))
          minz = minval(xnod(3,1:8))
-      case('mdlp')
-         maxz = maxval(xnod(3,1:6))
-         minz = minval(xnod(3,1:6))
       case default
          write(*,*) 'elem_maxwell_gram_hexa: unexpected etype=',etype,'. stop.'
          stop
