+++ trunk/problems/LASER/UW_COUPLED/elem/elem_maxwell.F90
@@ -254,10 +254,10 @@ subroutine elem_maxwell(Mdle,Fld_flag,                &
 !
 !..determine z-coordinate inside the element
    select case(etype)
-      case('mdlb')
+      case(MDLB)
          maxz = maxval(xnod(3,1:8))
          minz = minval(xnod(3,1:8))
-      case('mdlp')
+      case(MDLP)
          maxz = maxval(xnod(3,1:6))
          minz = minval(xnod(3,1:6))
       case default
