+++ trunk/problems/LASER/UW_COUPLED/elem/elem_maxwell_fi_pris.F90
@@ -354,10 +354,10 @@ subroutine elem_maxwell_fi_pris(Mdle,Fld_flag,                &
 !
 !..determine z-coordinate inside the element
    select case(etype)
-      case('mdlb')
+      case(MDLB)
          maxz = maxval(xnod(3,1:8))
          minz = minval(xnod(3,1:8))
-      case('mdlp')
+      case(MDLP)
          maxz = maxval(xnod(3,1:6))
          minz = minval(xnod(3,1:6))
       case default
