+++ trunk/problems/LASER/UW_COUPLED/exec_job_nl.F90
@@ -291,7 +291,7 @@ subroutine exec_job_nl
          call get_L2NormAttr(prevAttr, fieldNormQ)
       endif
       if (RANK.eq.ROOT) write(*,4240) '   L2NormDiff = ', L2NormDiff
-      if (RANK.eq.ROOT) write(*,4240) '   FieldNormQ = ', FieldNormQ
+      if (RANK.eq.ROOT) write(*,4240) '   FieldNormQ = ', fieldNormQ
  4240 format(A,F10.4)
 !
 !  ...copy current solution components of all fields into previous solution
@@ -327,7 +327,7 @@ subroutine exec_job_nl
 !
 !..display stats
    if (RANK.eq.ROOT) then
-      write(*,*) 'L2NormDiff/fieldNormQ:'
+      write(*,*) 'L2NormDiff/FieldNormQ:'
       do j=1,i
          write(*,4241) L2NormDiffIter(j)
       enddo
