+++ trunk/problems/MAXWELL/GALERKIN/exec_job.F90
@@ -103,7 +103,6 @@ subroutine exec_job
 !
 !  ...solve problem with par_mumps (MPI MUMPS)
       call MPI_BARRIER (MPI_COMM_WORLD, ierr)
-      !if(RANK .eq. ROOT) write(*,200) '6. calling PETSc (MPI) solver...'
       if(RANK .eq. ROOT) write(*,200) '6. calling solver...'
       call MPI_BARRIER (MPI_COMM_WORLD, ierr); start_time = MPI_Wtime()
       call par_mumps_sc('G')
