### 2024-01-12 b0e0370
adding peak temp comp
