+++ trunk/problems/LASER/COMMON_FILES/thermal.F90
@@ -7,10 +7,11 @@
 !
 !-------------------------------------------------------------------------------
 !
-!     latest revision:  - June 2021
+!     latest revision:  - Jan 2024
 !
-!     purpose:          - routine sets up computation of average temperature
-!                         in the fiber at certain z-locations
+!     purpose:          - routine sets up computation of average and peak
+!                         temperature in the fiber at certain z-locations
+!                         (peak temperature assumed at R=0 fiber axis)
 !
 !     arguments:
 !        inout:
@@ -32,7 +33,7 @@ subroutine get_avgTemp(NumPts,FileIter)
    integer, intent(in)    :: FileIter
    integer, intent(inout) :: NumPts
 !
-   real(8), allocatable :: zValues(:),coreTemp(:)
+   real(8), allocatable :: zValues(:),coreTemp(:),peakTemp(:)
 !
    real(8) :: a,b
    integer :: i
@@ -52,7 +53,7 @@ subroutine get_avgTemp(NumPts,FileIter)
 !
    if ((.not. DISTRIBUTED .or. HOST_MESH) .and. RANK .ne. ROOT) goto 99
 !
-   allocate(zValues(NumPts),coreTemp(NumPts))
+   allocate(zValues(NumPts),coreTemp(NumPts),peakTemp(NumPts))
 !
 !..distributing sample points uniformly
    if (RANK .eq. ROOT) then
@@ -69,33 +70,47 @@ subroutine get_avgTemp(NumPts,FileIter)
    zValues = zValues*PI*(113.d0/355.d0)
 !
    if (RANK .eq. ROOT) write(*,*) ' get_avgTemp: computing core temperature values..'
-   call comp_avgTemp(zValues,NumPts, coreTemp)
+   call comp_avgTemp(zValues,NumPts, coreTemp,peakTemp)
 !
-!..only ROOT proc has valid coreTemp values
+!..only ROOT proc has valid coreTemp and peakTemp values
    if (RANK .ne. ROOT) goto 90
 !
    if (FileIter .eq. -1) then
-      write(*,*) ' get_avgTemp: printing core temperature values..'
+      write(*,*) ' get_avgTemp: printing average core temperature values..'
       do i = 1,NumPts
          write(*,2020) coreTemp(i)
  2020    format('    ',es12.5)
       enddo
+!
+      write(*,*) ' get_avgTemp: printing peak temperature values..'
+      do i = 1,NumPts
+         write(*,2020) peakTemp(i)
+      enddo
    endif
 !
    if (FileIter .ge. 0) then
-      write(*,*) ' get_avgTemp: printing core temperature values to file..'
+      write(*,*) ' get_avgTemp: printing average core temperature values to file..'
       fmt = '(I5.5)'
       write (suffix,fmt) FileIter
-      filename=trim(OUTPUT_DIR)//'temp/temp_'//trim(suffix)//'.dat'
+      filename=trim(OUTPUT_DIR)//'temp/avg_temp_'//trim(suffix)//'.dat'
       open(UNIT=9,FILE=filename,FORM="FORMATTED",STATUS="REPLACE",ACTION="WRITE")
       do i = 1,NumPts
          write(UNIT=9, FMT="(es12.5)") coreTemp(i)
       enddo
       close(UNIT=9)
+!
+      write(*,*) ' get_avgTemp: printing peak temperature values to file..'
+      write (suffix,fmt) FileIter
+      filename=trim(OUTPUT_DIR)//'temp/peak_temp_'//trim(suffix)//'.dat'
+      open(UNIT=9,FILE=filename,FORM="FORMATTED",STATUS="REPLACE",ACTION="WRITE")
+      do i = 1,NumPts
+         write(UNIT=9, FMT="(es12.5)") peakTemp(i)
+      enddo
+      close(UNIT=9)
[…]
