### 2024-01-15 24b0222
minor style changes
