### 2024-01-16 b583dbd
removing index case 7 and minor style changes
