### 2024-01-16 16b8655
compatibility of problems with node%bcond index change
