+++ trunk/problems/LASER/COMMON_FILES/commonRoutines.F90
@@ -24,7 +24,7 @@
 !
 !   arguments:
 !     in:
-!              Icomp   - Physics attribute component number (1,..,NRINDEX)
+!              Icomp   - Physics attribute component number (1,..,NRINDEX_HEV)
 !              Nflag   - A custom BC flag (2-9); e.g., impedance BC flag
 !
 !-------------------------------------------------------------------------------
@@ -55,12 +55,17 @@ subroutine propagate_flag(Icomp,Nflag)
    integer :: nodesl(27),norientl(27),nface_nodes(9)
 !
 !..element face BC flags, decoded BC flag for a node
-   integer :: ibc(6,NRINDEX),nodflag(NRINDEX)
+   integer :: ibc(6,NRINDEX),nodflag(NRINDEX_HEV)
 !
 !-------------------------------------------------------------------------------
 !
    if (IBCFLAG .ne. 3) then
       write(*,*) 'propagate_flag called for IBCFLAG.ne.3, returning...'
+      return
+   endif
+!
+   if ((Icomp.lt.1) .or. (Icomp.gt.NRINDEX_HEV))
+      write(*,*) 'propagate_flag: invalid Icomp = ', Icomp
       return
    endif
 !
@@ -118,13 +123,13 @@ subroutine propagate_flag(Icomp,Nflag)
 !$OMP DO
    do nod=1,NRNODS
       if (NODES(nod)%visit.eq.0) cycle
-      call decod(NODES(nod)%bcond,2,NRINDEX, nodflag)
+      call decod(NODES(nod)%bcond,2,NRINDEX_HEV, nodflag)
       if (NODES(nod)%visit.eq.-Nflag) then
          nodflag(Icomp) = 0
       elseif (NODES(nod)%visit.eq.Nflag) then
          nodflag(Icomp) = 1
       endif
-      call encod(nodflag,2,NRINDEX, NODES(nod)%bcond)
+      call encod(nodflag,2,NRINDEX_HEV, NODES(nod)%bcond)
 !
    enddo
 !$OMP END DO
