### 2024-01-16 30b5707
simplifying loop
