+++ trunk/problems/MAXWELL/ULTRAWEAK_DPG/common/commonRoutines.F90
@@ -35,7 +35,7 @@
          return
       endif
 !
-      if ((Icomp.lt.1) .or. (Icomp.gt.NRINDEX_HEV))
+      if ((Icomp.lt.1) .or. (Icomp.gt.NRINDEX_HEV)) then
          write(*,*) 'propagate_flag: invalid Icomp = ', Icomp
          return
       endif
