### 2024-01-23 42c48d1
Node bcond enconding modified with new parameter `INDEX_HEV` (#162)
Co-authored-by: stefanhenneking <stefan.henneking@gmail.com>
