+++ trunk/src/hpinterp/edge/dhpedgeE.F90
@@ -82,7 +82,7 @@ subroutine dhpedgeE(Mdle,Iflag,No,Etav,Ntype,Icase,Bcond,&
 !
 ! decoded case and BC flag for the face node
   integer, dimension(NR_PHYSA)          :: ncase
-  integer, dimension(NRINDEX)           :: ibcnd
+  integer, dimension(NRINDEX_HEV)       :: ibcnd
 !
 ! work space for linear solvers
   integer                               :: naE,info
