### 2024-01-27 #5946 @stefanhenneking
Hi @jedbrown, thanks for the reminder. I tried out a few things that didn't quite work out the way I'd hoped and I didn't end up pushing those to the public repo. But this is on my todo list and I hope to have it done over the next several weeks. I'll ping you when I have an update!
