+++ .github/workflows/linux.yml
@@ -30,10 +30,11 @@ jobs:
       - run: echo "This job's status is ${{ job.status }}."
 
       # Get MPI
-      - name: get MPI (Linux)
-        run: sudo apt-get install mpich libmpich-dev
+      #- name: get MPI (Linux)
+      #  run: sudo apt-get install mpich libmpich-dev
 
-      - run: echo "Finished installing MPI."
+      #- run: echo "Finished installing MPI."
 
       # Get PETSc
-
+      - name: build-petsc
+        run: spack info petsc
