+++ .github/workflows/linux.yml
@@ -30,10 +30,10 @@ jobs:
       - run: echo "This job's status is ${{ job.status }}."
 
       # Get MPI
-      #- name: get MPI (Linux)
-      #  run: sudo apt-get install mpich libmpich-dev
+      - name: get MPI (Linux)
+        run: sudo apt-get install mpich libmpich-dev
 
-      #- run: echo "Finished installing MPI."
+      - run: echo "Finished installing MPI."
 
       # Get PETSc
       #- name: build-petsc
@@ -45,8 +45,11 @@ jobs:
           echo "Getting latest PETSc release version"
           git clone -b release https://gitlab.com/petsc/petsc.git petsc
           echo "Configuring PETSc"
+          cd petsc
+          ./configure --with-cc=mpicc --with-cxx=mpicxx --with-mpiexec=mpirun \
+          --with-scalar-type=real --PETSC_ARCH=arch-ubuntu-real
 
           echo "Building PETSc"
-          # make -j
+          make -j
 
           # echo "Installing PETSc"
