### 2024-02-17 b5973a5
petsc configure options
