+++ .github/workflows/linux.yml
@@ -19,7 +19,7 @@ jobs:
       - run: echo "This job is now running on a ${{ runner.os }} server hosted by GitHub!"
       - run: echo "The name of your branch is ${{ github.ref }} and your repository is ${{ github.repository }}."
 
-      # Interrupt workflow already running on the same branch.
+      # Interrupt workflow already running on the same branch
       - name: Cancel Previous Runs
         uses: styfle/cancel-workflow-action@0.12.1
         with:
