+++ .github/workflows/makedepend.yml
@@ -1,60 +1,0 @@
-name: Linux makedepend
-run-name: hp3d-linux-makedepend
-
-on: [push]
-jobs:
-  linux:
-    runs-on: ubuntu-latest
-    steps:
-      - run: echo "The job was automatically triggered by a ${{ github.event_name }} event."
-      - run: echo "This job is now running on a ${{ runner.os }} server hosted by GitHub!"
-      - run: echo "The name of your branch is ${{ github.ref }} and your repository is ${{ github.repository }}."
-
-      # Interrupt workflow already running on the same branch.
-      - name: Cancel Previous Runs
-        uses: styfle/cancel-workflow-action@0.12.1
-        with:
-          access_token: ${{ github.token }}
-
-      # Check out hp3d repository
-      - name: Check out hp3d
-        uses: actions/checkout@v4
-
-      - run: echo "The ${{ github.repository }} repository has been cloned to the runner."
-      - run: echo "The workflow is now ready to test your code on the runner."
-
-      # List files in the repository
-      - name: List files in the repository
-        run: ls ${{ github.workspace }}
-
-      # Get MPI
-      - name: Get MPI
-        run: |
-          sudo apt-get install mpich libmpich-dev
-
-      - name: Check MPI
-        run: |
-          mpicc --version
-          mpicxx --version
-          mpif90 --version
-          mpirun --version
-
-      # Build hp3D
-      - name: Build hp3D
-        run: |
-          echo "Configuring hp3D"
-          cd trunk
-          cp m_options_files/workflows/m_options_ubuntu_real ./m_options
-
-          echo "Building hp3D"
-          make -j
-
-      # Run hp3D tests
-      - name: Run hp3D tests
-        run: |
-          echo "Running hp3D tests"
-          cd trunk
-          make test
-
-      - run: echo "This job's status is ${{ job.status }}."
-
