+++ .github/workflows/linux.yml
@@ -25,9 +25,10 @@ jobs:
       # Print runner info to console
       - name: Print runner info
         run: |
-          echo "The job was automatically triggered by a ${{ github.event_name }} event."
-          echo "This job is now running on a ${{ runner.os }} server hosted by GitHub."
-          echo "The branch name is ${{ github.ref }} repository is ${{ github.repository }}."
+          echo "This job was triggered by a ${{ github.event_name }} event."
+          echo "This job is running on a ${{ runner.os }} server hosted by GitHub."
+          echo "The repository name is ${{ github.repository }}."
+          echo "The branch name is ${{ github.ref }}."
 
       # Interrupt workflow already running on the same branch
       - name: Cancel previous runs
@@ -36,7 +37,7 @@ jobs:
           access_token: ${{ github.token }}
 
       # Check out hp3d repository
-      - name: Check out hp3d
+      - name: Check out hp3d repository
         uses: actions/checkout@v4
 
       - run: echo "${{ github.repository }} repository has been cloned to the runner."
@@ -50,6 +51,7 @@ jobs:
         run: |
           sudo apt-get install mpich libmpich-dev
 
+      # Check MPI
       - name: Check MPI
         run: |
           mpicc --version
@@ -57,7 +59,7 @@ jobs:
           mpif90 --version
           mpirun --version
 
-      # Get PETSc
+      # Build PETSc
       - name: Build PETSc
         run: |
           echo "Getting latest PETSc release version"
