+++ .github/workflows/github-actions-demo.yml
@@ -1,6 +1,6 @@
 name: GitHub Actions Demo
 run-name: ${{ github.actor }} is testing out GitHub Actions 🚀
-on: workflow_dispatch
+on: [push]
 jobs:
   Explore-GitHub-Actions:
     runs-on: ubuntu-latest
