+++ .github/workflows/linux.yml
@@ -7,6 +7,7 @@ on:
       - master
       - release
       - next
+  pull_request:
   workflow_dispatch:
 
 jobs:
