+++ .github/workflows/builds-and-tests.yml
@@ -52,11 +52,12 @@ jobs:
         if: matrix.os == 'ubuntu-latest'
         run: sudo apt-get install mpich libmpich-dev
 
-      # - name: Get MPI (MacOS)
-      #   if: matrix.os == 'macos-latest'
-      #   run: |
-      #     export HOMEBREW_NO_INSTALL_CLEANUP=1
-      #     brew install open-mpi
+      - name: Get MPI (MacOS)
+        if: matrix.os == 'macos-latest'
+        run: |
+          export HOMEBREW_NO_INSTALL_CLEANUP=1
+          brew install gfortran
+      #brew install open-mpi
 
       # Check MPI
       - name: Check MPI (Linux)
