+++ .github/workflows/builds-and-tests.yml
@@ -56,7 +56,8 @@ jobs:
         if: matrix.os == 'macos-latest'
         run: |
           export HOMEBREW_NO_INSTALL_CLEANUP=1
-          brew reinstall gfortran
+          ln -s /usr/local/bin/gfortran-13 /usr/local/bin/gfortran
+          gfortran --version
           brew install open-mpi
 
       # Check MPI
@@ -68,8 +69,7 @@ jobs:
           mpirun --version
 
       # Build PETSc
-      - name: Build PETSc (Linux)
-        if: matrix.os == 'ubuntu-latest'
+      - name: Build PETSc
         run: |
           echo "Getting latest PETSc release version"
           git clone -b release https://gitlab.com/petsc/petsc.git petsc
@@ -80,40 +80,6 @@ jobs:
           --with-cc=mpicc --COPTFLAGS="-g -O" \
           --with-cxx=mpicxx --CXXOPTFLAGS="-g -O" \
           --with-fc=mpif90 --FOPTFLAGS="-g -O" \
-          --with-mpiexec=mpirun \
-          --download-fblaslapack=yes \
-          --download-scalapack=yes \
-          --download-mumps=yes \
-          --download-metis=yes \
-          --download-parmetis=yes \
-          --download-ptscotch=yes \
-          --download-zoltan=yes \
-          --download-hdf5=yes \
-          --with-hdf5-fortran-bindings=1 \
-          --with-shared-libraries=0 \
-          --with-debugging=0 \
-          --with-scalar-type=${{ matrix.mode }} \
-          --PETSC_ARCH=arch-${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
-
-          echo "Building PETSc"
-          make -j
-
-          echo "Checking PETSc build"
-          make check
-
-      # Build PETSc
-      - name: Build PETSc (MacOS)
-        if: matrix.os == 'macos-latest'
-        run: |
-          echo "Getting latest PETSc release version"
-          git clone -b release https://gitlab.com/petsc/petsc.git petsc
-          
-          echo "Configuring PETSc"
-          cd petsc
-          ./configure \
-          --with-cc=mpicc --COPTFLAGS="-g -O" \
-          --with-cxx=mpicxx --CXXOPTFLAGS="-g -O" \
-          --with-fc=mpifort --FOPTFLAGS="-g -O" \
           --with-mpiexec=mpirun \
           --download-fblaslapack=yes \
           --download-scalapack=yes \
