+++ .github/workflows/builds-and-tests.yml
@@ -47,19 +47,33 @@ jobs:
       - name: List files in the repository
         run: ls ${{ github.workspace }}
 
-      # Get MPI
-      - name: Get MPI (Linux)
-        if: matrix.os == 'ubuntu-latest'
-        run: sudo apt-get install mpich libmpich-dev
+      # Symlink gfortran (MacOS)
+      # (alternatively, use "brew reinstall gfortran")
+      - name: Symlink gfortran (MacOS)
+        if: matrix.os == 'macos-latest'
+        run: |
+          ln -s /usr/local/bin/gfortran-13 /usr/local/bin/gfortran
+          gfortran --version
 
-      # Alternatively, use "brew reinstall gfortran"
+      # Get Bison and Flex (MacOS)
+      - name: Get Bison and Flex (MacOS)
+        if: matrix.os == 'macos-latest'
+        run: 
+          brew install bison flex
+          sudo ln -s /usr/local/opt/bison/bin/bison /usr/local/bin/bison
+          sudo ln -s /usr/local/opt/flex/bin/flex /usr/local/bin/flex
+
+      # Get MPI (MacOS)
       - name: Get MPI (MacOS)
         if: matrix.os == 'macos-latest'
         run: |
           export HOMEBREW_NO_INSTALL_CLEANUP=1
-          ln -s /usr/local/bin/gfortran-13 /usr/local/bin/gfortran
-          gfortran --version
           brew install open-mpi
+
+      # Get MPI (Linux)
+      - name: Get MPI (Linux)
+        if: matrix.os == 'ubuntu-latest'
+        run: sudo apt-get install mpich libmpich-dev
 
       # Check MPI
       - name: Check MPI
@@ -68,11 +82,6 @@ jobs:
           mpicxx --version
           mpif90 --version
           mpirun --version
-
-      # Get Bison
-      - name: Get Bison (MacOS)
-        if: matrix.os == 'macos-latest'
-        run: brew install bison
 
       # Build PETSc
       - name: Build PETSc
