+++ .github/workflows/builds-and-tests.yml
@@ -69,7 +69,7 @@ jobs:
       - name: Get X11 (MacOS)
         if: matrix.os == 'macos-latest'
         run: |
-          brew cask install xquartz
+          brew install xquartz --cask
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
