+++ .github/workflows/builds-and-tests.yml
@@ -66,16 +66,18 @@ jobs:
           sudo ln -s /usr/local/opt/flex/bin/flex /usr/local/bin/flex
 
       # Get X11 (MacOS)
+      # (alternatively, use "brew install xquartz --cask")
       - name: Get X11 (MacOS)
         if: matrix.os == 'macos-latest'
-        run: |
-          brew install xquartz --cask
+        run:
+          echo $PATH 
+          brew install libx11
+          echo $PATH
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
         if: matrix.os == 'macos-latest'
-        run: |
-          brew install open-mpi
+        run: brew install open-mpi
 
       # Get MPI (Linux)
       - name: Get MPI (Linux)
