+++ .github/workflows/builds-and-tests.yml
@@ -76,6 +76,10 @@ jobs:
           echo $PATH 
           brew reinstall libx11
           echo $PATH
+          ls /usr/local/Cellar/libx11
+          ls /usr/local/Cellar/libx11/1.8.7
+          find /usr/local/Cellar -name "libx11*"
+          find /usr/local/opt -name "libx11*"
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
