+++ .github/workflows/builds-and-tests.yml
@@ -87,6 +87,8 @@ jobs:
           ls /usr/local/opt/libx11/include
           echo "find /usr/local/opt/libx11 -name libX11*"
           find /usr/local/opt/libx11 -name "libX11*"
+          echo "ls /usr/local/opt/libx11/X11"
+          ls /usr/local/opt/libx11/X11
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
