+++ .github/workflows/builds-and-tests.yml
@@ -83,12 +83,12 @@ jobs:
           ls /usr/local/opt/libx11
           echo "ls /usr/local/opt/libx11/lib"
           ls /usr/local/opt/libx11/lib
+          echo "ls /usr/local/opt/libx11/lib/X11"
+          ls /usr/local/opt/libx11/lib/X11
           echo "ls /usr/local/opt/libx11/include"
           ls /usr/local/opt/libx11/include
-          echo "find /usr/local/opt/libx11 -name libX11*"
-          find /usr/local/opt/libx11 -name "libX11*"
-          echo "ls /usr/local/opt/libx11/X11"
-          ls /usr/local/opt/libx11/X11
+          echo "ls /usr/local/opt/libx11/include/X11"
+          ls /usr/local/opt/libx11/include/X11
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
