+++ trunk/test/makefile
@@ -7,6 +7,9 @@ info :
 
 include ../m_options
 
+HP3D_MPIRUN ?= mpirun
+HP3D_MPI_NP ?= 2
+
 OBJ_PATH     = _obj_
 EXEC_PATH    = _exec_
 MODULE_PATH  = $(OBJ_PATH)
@@ -172,7 +175,7 @@ all : seq par
 .PHONY : run
 run : one
 	@printf "\n Running test: "$(TEST)"\n"
-	@pass=1; mpirun -np 1 ./$(EXEC_PATH)/$(TEST) || pass=0; \
+	@pass=1; $(HP3D_MPIRUN) -np 1 ./$(EXEC_PATH)/$(TEST) || pass=0; \
 	if [ $$((pass)) -eq 1 ]; then \
 		printf "          test_"$(TEST)" PASSED\n\n"; \
 		true; \
@@ -186,7 +189,7 @@ run : one
 .PHONY : run-par
 run-par : one
 	@printf "\n Running test: "$(TEST)"\n"
-	@pass=1; mpirun -np 4 ./$(EXEC_PATH)/$(TEST) || pass=0; \
+	@pass=1; $(HP3D_MPIRUN) -np $(HP3D_MPI_NP) ./$(EXEC_PATH)/$(TEST) || pass=0; \
 	if [ $$((pass)) -eq 1 ]; then \
 		printf "          test_"$(TEST)" PASSED\n\n"; \
 		true; \
