+++ .github/workflows/builds-and-tests.yml
@@ -53,26 +53,26 @@ jobs:
 
       # Symlink gfortran (MacOS)
       # (alternatively, use "brew reinstall gfortran")
-      # - name: Symlink gfortran (MacOS)
-      #   if: matrix.os == 'macos-latest'
-      #   run: |
-      #     echo "ls /usr/local/bin/gfort*"
-      #     ls /usr/local/bin/gfort*
-      #     sudo ln -s /usr/local/bin/gfortran-12 /usr/local/bin/gfortran
-      #     gfortran --version
-
-      # Install gfortran (MacOS)
-      - name: Install gfortran (MacOS)
+      - name: Symlink gfortran (MacOS)
         if: matrix.os == 'macos-latest'
         run: |
-          export HOMEBREW_NO_INSTALL_CLEANUP=1
           echo "ls /usr/local/bin/gfort*"
           ls /usr/local/bin/gfort*
-          brew reinstall gfortran@11
-          echo "ls /usr/local/bin/gfort*"
-          ls /usr/local/bin/gfort*
-          echo "gfortran --version"
+          sudo ln -s /usr/local/bin/gfortran-11 /usr/local/bin/gfortran
           gfortran --version
+
+      # Install gfortran (MacOS)
+      # - name: Install gfortran (MacOS)
+      #   if: matrix.os == 'macos-latest'
+      #   run: |
+      #     export HOMEBREW_NO_INSTALL_CLEANUP=1
+      #     echo "ls /usr/local/bin/gfort*"
+      #     ls /usr/local/bin/gfort*
+      #     brew reinstall gfortran
+      #     echo "ls /usr/local/bin/gfort*"
+      #     ls /usr/local/bin/gfort*
+      #     echo "gfortran --version"
+      #     gfortran --version
 
       # Set deployment target (MacOS)
       # $(sw_vers -productVersion)
