+++ .github/workflows/builds-and-tests.yml
@@ -14,7 +14,7 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [macos-latest]
+        os: [macos-13]
         mode: [real]
         target: [dbg]
         # os: [ubuntu-latest, macos-latest]
@@ -54,7 +54,7 @@ jobs:
       # Symlink gfortran (MacOS)
       # (alternatively, use "brew reinstall gfortran")
       - name: Symlink gfortran (MacOS)
-        if: matrix.os == 'macos-latest'
+        if: runner.os == 'macOS'
         run: |
           echo "ls /usr/local/bin/gfort*"
           ls /usr/local/bin/gfort*
@@ -63,7 +63,7 @@ jobs:
 
       # Install gfortran (MacOS)
       # - name: Install gfortran (MacOS)
-      #   if: matrix.os == 'macos-latest'
+      #   if: runner.os == 'macOS'
       #   run: |
       #     export HOMEBREW_NO_INSTALL_CLEANUP=1
       #     echo "ls /usr/local/bin/gfort*"
@@ -77,7 +77,7 @@ jobs:
       # Set deployment target (MacOS)
       # $(sw_vers -productVersion)
       # - name: Set deployment target (MacOS)
-      #   if: matrix.os == 'macos-latest'
+      #   if: runner.os == 'macOS'
       #   run: export MACOSX_DEPLOYMENT_TARGET=12.0
 
       # Get X11 (MacOS)
@@ -86,13 +86,13 @@ jobs:
       # ls /usr/local/Cellar/libx11/1.8.7
       # find /usr/local/Cellar -name "libx11*"
       # - name: Get X11 (MacOS)
-      #   if: matrix.os == 'macos-latest'
+      #   if: runner.os == 'macOS'
       #   run: |
       #     export PATH="/usr/local/opt/libx11:${PATH}"
 
       # Get Bison and Flex (MacOS)
       - name: Get Bison and Flex (MacOS)
-        if: matrix.os == 'macos-latest'
+        if: runner.os == 'macOS'
         run: |
           export HOMEBREW_NO_INSTALL_CLEANUP=1
           brew install bison flex
@@ -101,7 +101,7 @@ jobs:
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
-        if: matrix.os == 'macos-latest'
+        if: runner.os == 'macOS'
         run: brew install open-mpi
 
       # Get MPI (Linux)
