### 2024-02-19 9deb307
with brew scalapack
