+++ .github/workflows/builds-and-tests.yml
@@ -14,10 +14,10 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [macos-latest]
+        os: [ubuntu-latest, macos-latest]
         mode: [real]
         target: [dbg]
-        # os: [ubuntu-latest, macos-latest]
+        # os: [ubuntu-latest, macos-latest, macos-13]
         # mode: [real, complex]
         # target: [dbg, opt]
 
@@ -79,18 +79,8 @@ jobs:
       - name: Set deployment target (MacOS)
         if: runner.os == 'macOS'
         run: |
-          export MACOSX_DEPLOYMENT_TARGET=12.0
-          export CMAKE_OSX_DEPLOYMENT_TARGET=12.0
-
-      # Get X11 (MacOS)
-      # (alternatively, use "brew reinstall libx11"
-      #                  or "brew install xquartz --cask")
-      # ls /usr/local/Cellar/libx11/1.8.7
-      # find /usr/local/Cellar -name "libx11*"
-      # - name: Get X11 (MacOS)
-      #   if: runner.os == 'macOS'
-      #   run: |
-      #     export PATH="/usr/local/opt/libx11:${PATH}"
+          export MACOSX_DEPLOYMENT_TARGET=$(sw_vers -productVersion)
+          export CMAKE_OSX_DEPLOYMENT_TARGET=$(sw_vers -productVersion)
 
       # Get Bison and Flex (MacOS)
       - name: Get Bison and Flex (MacOS)
@@ -108,7 +98,7 @@ jobs:
 
       # Get MPI (Linux)
       - name: Get MPI (Linux)
-        if: matrix.os == 'ubuntu-latest'
+        if: runner.os == 'Linux'
         run: sudo apt-get install mpich libmpich-dev
 
       # Check MPI
