+++ .github/workflows/builds-and-tests.yml
@@ -14,7 +14,7 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-13]
         mode: [real]
         target: [dbg]
         # os: [ubuntu-latest, macos-latest, macos-13]
@@ -56,31 +56,17 @@ jobs:
       - name: Symlink gfortran (MacOS)
         if: runner.os == 'macOS'
         run: |
-          echo "ls /usr/local/bin/gfort*"
-          ls /usr/local/bin/gfort*
           sudo ln -s /usr/local/bin/gfortran-13 /usr/local/bin/gfortran
           gfortran --version
-
-      # Install gfortran (MacOS)
-      # - name: Install gfortran (MacOS)
-      #   if: runner.os == 'macOS'
-      #   run: |
-      #     export HOMEBREW_NO_INSTALL_CLEANUP=1
-      #     echo "ls /usr/local/bin/gfort*"
-      #     ls /usr/local/bin/gfort*
-      #     brew reinstall gfortran
-      #     echo "ls /usr/local/bin/gfort*"
-      #     ls /usr/local/bin/gfort*
-      #     echo "gfortran --version"
-      #     gfortran --version
 
       # Set deployment target (MacOS)
       # $(sw_vers -productVersion)
       - name: Set deployment target (MacOS)
         if: runner.os == 'macOS'
         run: |
-          export MACOSX_DEPLOYMENT_TARGET=$(sw_vers -productVersion)
-          export CMAKE_OSX_DEPLOYMENT_TARGET=$(sw_vers -productVersion)
+          echo "$(sw_vers -productVersion)"
+          export MACOSX_DEPLOYMENT_TARGET=13.0
+          export CMAKE_OSX_DEPLOYMENT_TARGET=13.0
 
       # Get Bison and Flex (MacOS)
       - name: Get Bison and Flex (MacOS)
