+++ .github/workflows/builds-and-tests.yml
@@ -14,7 +14,7 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-13]
+        os: [ubuntu-latest, macos-latest]
         mode: [real]
         target: [dbg]
         # os: [ubuntu-latest, macos-latest, macos-13]
@@ -76,11 +76,15 @@ jobs:
           brew install bison flex
           sudo ln -s /usr/local/opt/bison/bin/bison /usr/local/bin/bison
           sudo ln -s /usr/local/opt/flex/bin/flex /usr/local/bin/flex
+          cmake --version
+          brew reinstall cmake
+          cmake --version
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
         if: runner.os == 'macOS'
-        run: brew install open-mpi
+        run: brew install mpich libmpich-dev
+#open-mpi
 
       # Get MPI (Linux)
       - name: Get MPI (Linux)
