### 2024-02-19 52682c8
cmake w/ open-mpi
