+++ .github/workflows/builds-and-tests.yml
@@ -80,7 +80,8 @@ jobs:
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
         if: runner.os == 'macOS'
-        run: brew install open-mpi
+        run: brew install mpich 
+        #open-mpi
 
       # Get MPI (Linux)
       - name: Get MPI (Linux)
