+++ .github/workflows/builds-and-tests.yml
@@ -14,7 +14,7 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [macos-latest]
         mode: [real]
         target: [dbg]
         # os: [ubuntu-latest, macos-latest, macos-13]
