+++ .github/workflows/builds-and-tests.yml
@@ -14,7 +14,7 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [macos-latest]
+        os: [ubuntu-latest, macos-latest]
         mode: [real]
         target: [dbg]
         # os: [ubuntu-latest, macos-latest, macos-13]
@@ -36,6 +36,7 @@ jobs:
           echo "This job is running on a ${{ runner.os }} server hosted by GitHub."
           echo "The repository name is ${{ github.repository }}."
           echo "The branch name is ${{ github.ref }}."
+          ld -v
 
       # Interrupt workflow already running on the same branch
       # - name: Cancel previous runs
@@ -104,9 +105,9 @@ jobs:
           echo "Configuring PETSc"
           cd petsc
           ./configure \
-          --with-cc=mpicc --COPTFLAGS="-g -O -mmacosx-version-min=12.0" \
-          --with-cxx=mpicxx --CXXOPTFLAGS="-g -O -mmacosx-version-min=12.0" \
-          --with-fc=mpif90 --FOPTFLAGS="-g -O -mmacosx-version-min=12.0" \
+          --with-cc=mpicc --COPTFLAGS="-g -O" \
+          --with-cxx=mpicxx --CXXOPTFLAGS="-g -O" \
+          --with-fc=mpif90 --FOPTFLAGS="-g -O" \
           --with-mpiexec=mpirun \
           --download-scalapack=yes \
           --download-fblaslapack=yes \
