+++ .github/workflows/builds-and-tests.yml
@@ -14,7 +14,7 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest, macos-13]
+        os: [ubuntu-latest, macos-13]
         mode: [real]
         target: [dbg]
         # os: [ubuntu-latest, macos-latest, macos-13]
@@ -27,8 +27,8 @@ jobs:
 
     env:
       CONFIG: ${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
-      MACOSX_DEPLOYMENT_TARGET: 12.7
-      CMAKE_OSX_DEPLOYMENT_TARGET: 12.7
+      MACOSX_DEPLOYMENT_TARGET: 13.0
+      CMAKE_OSX_DEPLOYMENT_TARGET: 13.0
 
     steps:
       # Print runner info to console
