+++ .github/workflows/builds-and-tests.yml
@@ -27,8 +27,8 @@ jobs:
 
     env:
       CONFIG: ${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
-      MACOSX_DEPLOYMENT_TARGET: 12.7
-      CMAKE_OSX_DEPLOYMENT_TARGET: 12.7
+      MACOSX_DEPLOYMENT_TARGET: 12.0
+      CMAKE_OSX_DEPLOYMENT_TARGET: 12.0
 
     steps:
       # Print runner info to console
