+++ .github/workflows/builds-and-tests.yml
@@ -27,8 +27,9 @@ jobs:
 
     env:
       CONFIG: ${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
-      MACOSX_DEPLOYMENT_TARGET: 12.0
-      CMAKE_OSX_DEPLOYMENT_TARGET: 12.0
+      MACOSX_DEPLOYMENT_TARGET: '12.0'
+      CMAKE_OSX_DEPLOYMENT_TARGET: '12.0'
+      HOMEBREW_NO_INSTALL_CLEANUP: 1
 
     steps:
       # Print runner info to console
@@ -75,7 +76,6 @@ jobs:
       - name: Get Bison and Flex (MacOS)
         if: runner.os == 'macOS'
         run: |
-          export HOMEBREW_NO_INSTALL_CLEANUP=1
           brew install bison flex
           sudo ln -s /usr/local/opt/bison/bin/bison /usr/local/bin/bison
           sudo ln -s /usr/local/opt/flex/bin/flex /usr/local/bin/flex
@@ -83,9 +83,7 @@ jobs:
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
         if: runner.os == 'macOS'
-        run: |
-          export HOMEBREW_NO_INSTALL_CLEANUP=1
-          brew install open-mpi
+        run: brew install open-mpi
 
       # Get MPI (Linux)
       - name: Get MPI (Linux)
