### 2024-02-19 07b7714
enable all tests
