+++ .github/workflows/builds-and-tests.yml
@@ -17,6 +17,13 @@ jobs:
         os: [ubuntu-latest, macos-12, macos-13]
         mode: [real, complex]
         target: [dbg, opt]
+      include:
+        - os: macos-12
+          deployment: '12.0'
+        - os: macos-13
+          deployment: '13.0'
+        - os: ubuntu-latest
+          deployment: ''
 
     name: ${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
 
@@ -24,8 +31,8 @@ jobs:
 
     env:
       CONFIG: ${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
-      MACOSX_DEPLOYMENT_TARGET: ${{ sw_vers -productVersion }}
-      CMAKE_OSX_DEPLOYMENT_TARGET: ${{ sw_vers -productVersion }}
+      MACOSX_DEPLOYMENT_TARGET: ${{ matrix.deployment }}
+      CMAKE_OSX_DEPLOYMENT_TARGET: ${{ matrix.deployment }}
       HOMEBREW_NO_INSTALL_CLEANUP: 1
 
     steps:
