+++ .github/workflows/builds-and-tests.yml
@@ -17,14 +17,13 @@ jobs:
         os: [ubuntu-latest, macos-12, macos-13]
         mode: [real, complex]
         target: [dbg, opt]
-        deployment: ['']
-      include:
-        - os: macos-12
-          deployment: '12.0'
-        - os: macos-13
-          deployment: '13.0'
-        - os: ubuntu-latest
-          deployment: ''
+        include:
+          - os: ubuntu-latest
+            deployment: ''
+          - os: macos-12
+            deployment: '12.0'
+          - os: macos-13
+            deployment: '13.0'
 
     name: ${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
 
