+++ .github/workflows/builds-and-tests.yml
@@ -21,9 +21,9 @@ jobs:
           - os: ubuntu-latest
             deployment: ''
           - os: macos-12
-            deployment: '12.0'
+            deployment: '12.7'
           - os: macos-13
-            deployment: '13.0'
+            deployment: '13.6'
 
     name: ${{ matrix.os }}-${{ matrix.mode }}-${{ matrix.target }}
 
