+++ .github/workflows/builds-and-tests.yml
@@ -45,10 +45,10 @@ jobs:
           echo "The branch name is ${{ github.ref }}."
 
       # Interrupt workflow already running on the same branch
-      # - name: Cancel previous runs
-      #   uses: styfle/cancel-workflow-action@0.12.1
-      #   with:
-      #     access_token: ${{ github.token }}
+      - name: Cancel previous runs
+        uses: styfle/cancel-workflow-action@0.12.1
+        with:
+          access_token: ${{ github.token }}
 
       # Check out hp3d repository
       - name: Check out hp3d repository
