+++ trunk/makefile
@@ -212,15 +212,27 @@ endif
 
 check: lib
 	@echo "Checking the hp3D library."
-	@(cd test ; $(MAKE) one run ; cd ..);
-	@(cd problems/MPI_TEST/TEST ; $(MAKE) ; cd ../..);
+	@ERR=0; \
+	if ! $(MAKE) -C test one run; then ERR=1; fi; \
+	if ! $(MAKE) -C problems/MPI_TEST/TEST; then ERR=1; fi; \
+	if [ 0 -eq $${ERR} ]; then \
+		echo "'make check' passed."; \
+	else \
+		echo "'make check' failed."; \
+		exit 1; \
+	fi ;
 
 #-------------------------------------------------
 .PHONY : test
 
 test: lib
 	@echo "Testing the hp3D library."
-	@(cd test ; $(MAKE) all ; cd ..);
+	@if $(MAKE) -C test all; then \
+		echo "'make test' passed."; \
+	else \
+		echo "'make test' failed."; \
+		exit 1; \
+	fi;
 
 #-------------------------------------------------
 .PHONY : all
