+++ .github/workflows/builds-and-tests.yml
@@ -14,7 +14,8 @@ jobs:
   builds-and-tests:
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-12, macos-13]
+        #os: [ubuntu-latest, macos-12, macos-13]
+        os: [ubuntu-latest]
         mode: [real, complex]
         target: [dbg, opt]
         include:
@@ -72,6 +73,7 @@ jobs:
           brew install bison flex
           sudo ln -s /usr/local/opt/bison/bin/bison /usr/local/bin/bison
           sudo ln -s /usr/local/opt/flex/bin/flex /usr/local/bin/flex
+          brew install pkg-config
 
       # Get MPI (MacOS)
       - name: Get MPI (MacOS)
