+++ paper/paper.bib
@@ -60,10 +60,10 @@
   doi = {10.48550/arXiv.2207.12211}}
 
 @book{hpbook3,
-  Author = {Henneking, Stefan and Demkowicz, Leszek and Petrides, Socratis and Badger, Jacob and Fuentes, Federico and Keith, Brendan and Gatto, Paolo},
+  Author = {Henneking, Stefan and Demkowicz, Leszek and Petrides, Socratis and Fuentes, Federico and Keith, Brendan and Gatto, Paolo},
   Publisher = {In preparation},
   Title = {{Computing with $hp$ Finite Elements. III. Parallel $hp$3D Code}},
-  Year = {2023}}
+  Year = {2024}}
 
 @book{hpbook2,
   Author = {Demkowicz, Leszek and Kurtz, Jason and Pardo, David and Paszy\'{n}ski, Maciej and Rachowicz, Waldemar and Zdunek, Adam},
