### 2024-02-20 2d51cbb
Simplifying installation (#160)
* Added INSTALL file with version compatibilities.

* Added HP3D_USE_X11 flag option.

* Added default m_options files using PETSc config.
