+++ trunk/src/graphics/x11/util/arctan.F
@@ -1,3 +1,5 @@
+#if HP3D_USE_X11
+
 c----------------------------------------------------------------------
 c
 c   routine name       - arctan
@@ -45,3 +47,5 @@ c
       endif
 c
       end subroutine arctan
+
+#endif
