+++ trunk/src/graphics/x11/util/mat3prod.F
@@ -1,3 +1,5 @@
+#if HP3D_USE_X11
+
 c
 c----------------------------------------------------------------------
 c
@@ -40,3 +42,4 @@ c
 c
       end subroutine
 
+#endif
