### 2024-02-20 #5946 @stefanhenneking
Hi @jedbrown, I merged #159 and #160 that added CI via Github Actions and new `m_options` files that uses PETSc's config for compiler settings, library paths, and include paths based on `PETSC_DIR` and `PETSC_ARCH`.
