+++ trunk/src/solver/par_mumps/par_nested.F90
@@ -61,7 +61,9 @@ subroutine par_nested(mtype)
    use mpi_param , only: RANK,ROOT,NUM_PROCS
    use MPI       , only: MPI_SUM,MPI_MIN,MPI_MAX,MPI_IN_PLACE, &
                          MPI_INTEGER,MPI_INTEGER8,             &
-                         MPI_REAL8,MPI_COMPLEX16,MPI_Wtime
+                         MPI_REAL8,MPI_COMPLEX16,MPI_Wtime,    &
+                         MPI_BCAST,MPI_REDUCE,MPI_ALLREDUCE
+
    use mkl_spblas
 !
    implicit none
