+++ trunk/src/mpi/par_aux.F90
@@ -18,7 +18,8 @@ subroutine partition_fiber(subd_next)
    use data_structure3D
    use mpi_param, only: RANK,ROOT,NUM_PROCS
    use MPI,       only: MPI_COMM_WORLD,MPI_INTEGER,MPI_REAL8,  &
-                        MPI_MIN,MPI_MAX,MPI_IN_PLACE
+                        MPI_MIN,MPI_MAX,MPI_IN_PLACE,          &
+                        MPI_BCAST,MPI_REDUCE,MPI_ALLREDUCE
 !
    implicit none
 !
@@ -98,8 +99,8 @@ subroutine repartition_fiber(Subd_next)
 !
    use data_structure3D
    use mpi_param, only: RANK,ROOT,NUM_PROCS
-   use MPI,       only: MPI_COMM_WORLD,MPI_INTEGER,MPI_REAL8,  &
-                        MPI_SUM,MPI_IN_PLACE
+   use MPI,       only: MPI_COMM_WORLD,MPI_INTEGER,MPI_REAL8,MPI_SUM,  &
+                        MPI_IN_PLACE,MPI_BCAST,MPI_REDUCE,MPI_ALLREDUCE
    use stc,       only: stc_get_nrdof
 !
    implicit none
