+++ trunk/src/graphics/paraview/vector2vtk.F90
@@ -19,8 +19,7 @@ subroutine vector2vtk(Sname,Sfile,Snick,Idx, Ic)
    use physics
    use upscale
    use paraview
-   use MPI              , only: MPI_COMM_WORLD,MPI_SUM,MPI_INTEGER,MPI_Wtime, &
-                                MPI_BCAST,MPI_REDUCE,MPI_ALLREDUCE
+   use MPI
    use mpi_param        , only: RANK,ROOT,NUM_PROCS
    use par_mesh         , only: DISTRIBUTED,HOST_MESH
 !
