+++ trunk/src/mpi/par_verify.F90
@@ -16,8 +16,7 @@ subroutine par_verify()
    use environment, only: QUIET_MODE
    use mpi_param  , only: RANK,ROOT
    use par_mesh   , only: DISTRIBUTED
-   use MPI        , only: MPI_COMM_WORLD,MPI_INTEGER,MPI_MIN,MPI_BCAST, &
-                          MPI_BCAST,MPI_REDUCE,MPI_ALLREDUCE
+   use MPI
 !
    implicit none
 !
