+++ trunk/src/modules/bitvisit.F90
@@ -194,9 +194,8 @@ module bitvisit
 !--------------------------------------------------------------------
    subroutine reduce_visit
 !
+      use MPI
       use mpi_param,       only: RANK,ROOT
-      use MPI,             only: MPI_INTEGER, MPI_BOR, MPI_IN_PLACE, &
-                                 MPI_COMM_WORLD
       use par_mesh,        only: DISTRIBUTED
 !
       implicit none
@@ -207,7 +206,7 @@ module bitvisit
 !
       if (.not. DISTRIBUTED) return
 !
-      call MPI_Allreduce(MPI_IN_PLACE,visitation, nword,MPI_INTEGER,MPI_BOR,MPI_COMM_WORLD,ierr)
+      call MPI_ALLREDUCE(MPI_IN_PLACE,visitation, nword,MPI_INTEGER,MPI_BOR,MPI_COMM_WORLD,ierr)
 !
    end subroutine reduce_visit
 !
