+++ trunk/src/solver/petsc/petsc_solve.F90
@@ -47,10 +47,7 @@ subroutine petsc_solve(mtype)
    use par_mesh   , only:  DISTRIBUTED,HOST_MESH
    use mpi_param  , only:  RANK,ROOT,NUM_PROCS
    use mpi_wrapper, only:  mpi_w_handle_err
-   use MPI        , only:  MPI_SUM,MPI_MIN,MPI_MAX,MPI_IN_PLACE,  &
-                           MPI_INTEGER,MPI_INTEGER8,              &
-                           MPI_REAL8,MPI_COMPLEX16,               &
-                           MPI_COMM_WORLD,MPI_STATUS_SIZE,MPI_Wtime
+   use MPI
    use petscksp
    use petsc_w_ksp, only:  petsc_ksp_start,petsc_ksp_destroy,     &
                            petsc_ksp,petsc_ksp_type,              &
