+++ trunk/test/poly_maxw.F90
@@ -7,8 +7,6 @@ program test_poly_maxw
    use control
    use data_structure3D
    use environment
-   use mpi
-   use mpi_param
    use mpi_wrapper
    use par_mesh
    use stc
