+++ trunk/test/hexa_adapt_h_aniso.F90
@@ -7,8 +7,6 @@ program test_hexa_adapt_h_aniso
 !
    use data_structure3D
    use environment
-   use mpi
-   use mpi_param
    use mpi_wrapper
    use par_mesh
    use zoltan_wrapper
