+++ trunk/test/conv_maxw.F90
@@ -8,8 +8,6 @@ program test_conv_maxw
    use control
    use data_structure3D
    use environment
-   use mpi
-   use mpi_param
    use mpi_wrapper
    use par_mesh
    use stc
