+++ trunk/src/hpinterp/update_Ddof.F90
@@ -14,8 +14,7 @@ subroutine update_Ddof
    use data_structure3D
    use environment, only: QUIET_MODE
    use par_mesh   , only: DISTRIBUTED
-   use MPI
-   use mpi_param  , only: RANK,ROOT,NUM_PROCS
+   use mpi_wrapper
    use par_mesh   , only: DISTRIBUTED,HOST_MESH
 !
    implicit none
@@ -485,8 +484,7 @@ subroutine update_Ddof_omp
    use data_structure3D
    use environment, only: QUIET_MODE
    use par_mesh   , only: DISTRIBUTED
-   use MPI
-   use mpi_param  , only: RANK,ROOT,NUM_PROCS
+   use mpi_wrapper
    use par_mesh   , only: DISTRIBUTED,HOST_MESH
 !
    implicit none
