+++ trunk/problems/LASER/COMMON_FILES/thermal.F90
@@ -24,8 +24,7 @@ subroutine get_avgTemp(NumPts,FileIter)
 !
    use commonParam
    use laserParam
-   use mpi_param, only: RANK,ROOT
-   use MPI      , only: MPI_COMM_WORLD
+   use mpi_wrapper
    use par_mesh , only: DISTRIBUTED,HOST_MESH
 !
    implicit none
@@ -145,8 +144,7 @@ subroutine comp_avgTemp(ZValues,NumPts, CoreTemp,PeakTemp)
    use data_structure3D
    use control    , only : GEOM_TOL
    use environment, only : QUIET_MODE
-   use mpi_param  , only : RANK,ROOT
-   use MPI        , only : MPI_COMM_WORLD,MPI_IN_PLACE,MPI_REAL8,MPI_SUM,MPI_Wtime
+   use mpi_wrapper
    use par_mesh   , only : DISTRIBUTED,HOST_MESH
 !
    implicit none
