+++ trunk/problems/POISSON/ULTRAWEAK_DPG/main.F90
@@ -21,10 +21,7 @@ program main
    use assembly
    use assembly_sc, only: IPRINT_TIME
    use stc        , only: STORE_STC,HERM_STC
-!
-   use MPI        , only: MPI_COMM_WORLD
-   use mpi_param  , only: ROOT,RANK,NUM_PROCS
-   use mpi_wrapper, only: mpi_w_init,mpi_w_finalize
+   use mpi_wrapper
 !
    implicit none
 !
@@ -37,7 +34,7 @@ program main
 #endif
 !
 !..timer
-   real(8) :: MPI_Wtime,start_time,end_time
+   real(8) :: start_time,end_time
 !
 !----------------------------------------------------------------------
 !
@@ -138,9 +135,7 @@ subroutine master_main()
    use common_prob_data
    use data_structure3D
    use GMP
-!
-   use MPI           , only: MPI_COMM_WORLD,MPI_INTEGER
-   use mpi_param     , only: ROOT,RANK,NUM_PROCS
+   use mpi_wrapper
    use par_mesh      , only: DISTRIBUTED,HOST_MESH
    use zoltan_wrapper, only: zoltan_w_set_lb
 !
@@ -339,9 +334,7 @@ subroutine worker_main()
    use common_prob_data
    use data_structure3D
    use GMP
-!
-   use MPI           , only: MPI_COMM_WORLD,MPI_INTEGER
-   use mpi_param     , only: ROOT,RANK
+   use mpi_wrapper
    use par_mesh      , only: DISTRIBUTED
    use zoltan_wrapper, only: zoltan_w_set_lb
 !
