+++ trunk/problems/LASER/COMMON_FILES/power.F90
@@ -27,8 +27,7 @@ subroutine get_power(Fld,NumPts,FileIter)
 !
    use commonParam
    use laserParam
-   use mpi_param, only: RANK,ROOT
-   use MPI      , only: MPI_COMM_WORLD,MPI_IN_PLACE,MPI_REAL8,MPI_SUM
+   use mpi_wrapper
    use par_mesh , only: DISTRIBUTED,HOST_MESH
 !
    implicit none
@@ -767,8 +766,7 @@ subroutine compute_power(ZValues,Num_zpts,Fld, Power,DiffPower,CorePower,CladPow
    use data_structure3D
    use control    , only : GEOM_TOL
    use environment, only : QUIET_MODE
-   use mpi_param  , only : RANK,ROOT
-   use MPI        , only : MPI_COMM_WORLD,MPI_Wtime
+   use mpi_wrapper
    use par_mesh   , only : DISTRIBUTED
 !
    implicit none
