+++ trunk/problems/MAXWELL/GALERKIN/common/exact_error.F90
@@ -16,8 +16,7 @@ subroutine exact_error
    use environment
    use assembly_sc, only: NRDOF_TOT,NRDOF_CON
    use par_mesh   , only: DISTRIBUTED,HOST_MESH
-   use mpi_param  , only: ROOT,RANK
-   use MPI        , only: MPI_SUM,MPI_COMM_WORLD,MPI_REAL8
+   use mpi_wrapper
 !
    implicit none
 !
