### 2024-02-20 3aed5ad
mpi_wrapper for omp solvers
