+++ trunk/src/modules/par_ghost.F90
@@ -442,8 +442,8 @@ contains
 !
    use data_structure3D
    use bitvisit
-   use par_mesh,           only: DISTRIBUTED, HOST_MESH
    use mpi_param
+   use par_mesh,           only: DISTRIBUTED, HOST_MESH
 !
    implicit none
 !
@@ -574,8 +574,8 @@ contains
 !
    use data_structure3D
    use bitvisit
-   use par_mesh,           only: DISTRIBUTED, HOST_MESH
    use mpi_param
+   use par_mesh,           only: DISTRIBUTED, HOST_MESH
 !
    implicit none
 !
