+++ trunk/src/modules/mpi_wrapper.F90
@@ -3,7 +3,7 @@
 !
 !----------------------------------------------------------------------
 !     module:              mpi_wrapper
-!     last modified:       July 2019
+!     last modified:       Feb 2024
 !----------------------------------------------------------------------
 module mpi_wrapper
 !
