### 2024-02-21 20c7d45
enable mpif08 and fix a few issues
