+++ trunk/src/solver/petsc/petsc_solve.F90
@@ -45,7 +45,8 @@ subroutine petsc_solve(mtype)
                            stc_alloc,stc_dealloc,stc_get_nrdof
    use parameters , only:  NRRHS
    use par_mesh   , only:  DISTRIBUTED,HOST_MESH
-   use mpi_wrapper
+   use MPI
+   use mpi_param
    use petscksp
    use petsc_w_ksp, only:  petsc_ksp_start,petsc_ksp_destroy,     &
                            petsc_ksp,petsc_ksp_type,              &
@@ -121,7 +122,7 @@ subroutine petsc_solve(mtype)
    integer, allocatable :: TEMP_BUF(:)
 !
 !..dummy variables
-   VTYPE :: zvoid
+   VTYPE :: zvoid1(1),zvoid2(1)
 !
 !..workspace for celem
    integer, dimension(MAXNODM) :: nodm,ndofmH,ndofmE,ndofmV,ndofmQ
@@ -172,7 +173,7 @@ subroutine petsc_solve(mtype)
 !
    if (IPRINT_TIME .eq. 1) then
       call MPI_BARRIER(MPI_COMM_WORLD, ierr)
-      call mpi_w_handle_err(ierr,'MPI_BARRIER')
+      !call mpi_w_handle_err(ierr,'MPI_BARRIER')
       if (RANK .eq. ROOT) write(*,1001)
  1001 format(' STEP 1 started : Get assembly info')
       start_time = MPI_Wtime()
@@ -193,7 +194,7 @@ subroutine petsc_solve(mtype)
 !..compute global node ownership
    count = NRNODS
    call MPI_ALLREDUCE(MPI_IN_PLACE,NOD_OWN,count,MPI_INTEGER,MPI_MIN,MPI_COMM_WORLD, ierr)
-   call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
+   !call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
 !
    allocate(MAXDOFS(NR_PHYSA))
    MAXDOFS = 0; MAXDOFM = 0
@@ -231,10 +232,10 @@ subroutine petsc_solve(mtype)
 !  ...get information from celem
       if (ISTC_FLAG) then
          call celem_systemI(iel,mdle,idec, nrdofs,nrdofm,nrdofc,nodm,  &
-            ndofmH,ndofmE,ndofmV,ndofmQ,nrnodm,zvoid,zvoid)
+            ndofmH,ndofmE,ndofmV,ndofmQ,nrnodm,zvoid1,zvoid2)
       else
          call celem(mdle,idec, nrdofs,nrdofm,nrdofc,nodm,  &
-            ndofmH,ndofmE,ndofmV,ndofmQ,nrnodm,zvoid,zvoid)
+            ndofmH,ndofmE,ndofmV,ndofmQ,nrnodm,zvoid1,zvoid2)
       endif
 !
 !  ...nrdofc = number of modified element dof after compression
@@ -325,7 +326,7 @@ subroutine petsc_solve(mtype)
    nrdof_subd(RANK+1) = nrdof
    count = NUM_PROCS
    call MPI_ALLREDUCE(MPI_IN_PLACE,nrdof_subd,count,MPI_INTEGER,MPI_MAX,MPI_COMM_WORLD, ierr)
-   call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
+   !call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
 !
 !..compute non-zero entries per row, accounting for own subdomain interaction
    allocate(petsc_dnz(nrdof_subd(RANK+1))); petsc_dnz = 0
[…]
