+++ trunk/src/solver/par_mumps/par_mumps.F90
@@ -40,14 +40,15 @@ subroutine mumps_destroy_par
    call mumps_destroy(mumps_par)
 end subroutine mumps_destroy_par
 !
-subroutine mumps_start(mumps,mpi_comm)
+subroutine mumps_start(mumps,mumps_comm)
 !
 #if C_MODE
    type (ZMUMPS_STRUC), intent(inout) :: mumps
 #else
    type (DMUMPS_STRUC), intent(inout) :: mumps
 #endif
-   integer, intent(in) :: mpi_comm
+!
+   type(MPI_Comm), intent(in) :: mumps_comm
 !
 !..NOTE:
 !..If you plan to run MUMPS sequentially from a parallel MPI application,
@@ -55,7 +56,7 @@ subroutine mumps_start(mumps,mpi_comm)
 !  communicator containing a single processor to the MUMPS library
 !
 !..Define a communicator for the package.
-   mumps%COMM = mpi_comm
+   mumps%COMM = mumps_comm%MPI_VAL
 !
 !..PAR
 !     0 : host is not involved in factorization/solve phases
@@ -194,7 +195,7 @@ subroutine mumps_start_subd
 !..this routine initiates sparse solver for interior of subdomain
 !
 !..Define a communicator for the package.
-   mumps_bub%COMM = MPI_COMM_SELF
+   mumps_bub%COMM = MPI_COMM_SELF%MPI_VAL
 !
 !..PAR
 !     0 : host is not involved in factorization/solve phases

[…]
