+++ trunk/src/solver/petsc/petsc_solve.F90
@@ -45,8 +45,7 @@ subroutine petsc_solve(mtype)
                            stc_alloc,stc_dealloc,stc_get_nrdof
    use parameters , only:  NRRHS
    use par_mesh   , only:  DISTRIBUTED,HOST_MESH
-   use MPI
-   use mpi_param
+   use mpif90_wrapper
    use petscksp
    use petsc_w_ksp, only:  petsc_ksp_start,petsc_ksp_destroy,     &
                            petsc_ksp,petsc_ksp_type,              &
@@ -173,7 +172,7 @@ subroutine petsc_solve(mtype)
 !
    if (IPRINT_TIME .eq. 1) then
       call MPI_BARRIER(MPI_COMM_WORLD, ierr)
-      !call mpi_w_handle_err(ierr,'MPI_BARRIER')
+      call mpi_w_handle_err(ierr,'MPI_BARRIER')
       if (RANK .eq. ROOT) write(*,1001)
  1001 format(' STEP 1 started : Get assembly info')
       start_time = MPI_Wtime()
@@ -194,7 +193,7 @@ subroutine petsc_solve(mtype)
 !..compute global node ownership
    count = NRNODS
    call MPI_ALLREDUCE(MPI_IN_PLACE,NOD_OWN,count,MPI_INTEGER,MPI_MIN,MPI_COMM_WORLD, ierr)
-   !call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
+   call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
 !
    allocate(MAXDOFS(NR_PHYSA))
    MAXDOFS = 0; MAXDOFM = 0
@@ -326,7 +325,7 @@ subroutine petsc_solve(mtype)
    nrdof_subd(RANK+1) = nrdof
    count = NUM_PROCS
    call MPI_ALLREDUCE(MPI_IN_PLACE,nrdof_subd,count,MPI_INTEGER,MPI_MAX,MPI_COMM_WORLD, ierr)
-   !call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
+   call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
 !
 !..compute non-zero entries per row, accounting for own subdomain interaction
    allocate(petsc_dnz(nrdof_subd(RANK+1))); petsc_dnz = 0
@@ -408,7 +407,7 @@ subroutine petsc_solve(mtype)
 !  in order to account for non-zero entries from interaction between subdomains
    count = NUM_PROCS * NUM_PROCS
    call MPI_ALLREDUCE(MPI_IN_PLACE,NOD_COMM(1:NUM_PROCS,1:NUM_PROCS),count,MPI_INTEGER,MPI_SUM,MPI_COMM_WORLD, ierr)
-   !call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
+   call mpi_w_handle_err(ierr,'MPI_ALLREDUCE')
 !
 !..exchange off-diagonal contributions
    nr_send = 0; nr_recv = 0
@@ -420,7 +419,7 @@ subroutine petsc_solve(mtype)
          tag   = count
          nr_send  = nr_send + 1
          call MPI_ISEND(ONZ(i)%SEND_BUF(1:count),count,MPI_INTEGER,rcv,tag,MPI_COMM_WORLD, reqs(nr_send),ierr)
-         !call mpi_w_handle_err(ierr,'MPI_ISEND')
+         call mpi_w_handle_err(ierr,'MPI_ISEND')
       endif
    enddo
    do i=1,NUM_PROCS
@@ -432,7 +431,7 @@ subroutine petsc_solve(mtype)
[…]
