### 2024-02-21 6aff36e
makefile change and minor fixes in problems
