+++ trunk/problems/MAXWELL/ULTRAWEAK_DPG/main.F90
@@ -23,8 +23,7 @@ program main
    use assembly_sc, only: IPRINT_TIME
    use stc        , only: STORE_STC,HERM_STC
 !
-   use MPI        , only: MPI_COMM_WORLD,MPI_BARRIER,MPI_ABORT, &
-                          MPI_GET_PROCESSOR_NAME,MPI_MAX_PROCESSOR_NAME
+   use MPI
    use mpi_param  , only: ROOT,RANK,NUM_PROCS
    use mpi_wrapper, only: mpi_w_init,mpi_w_finalize
 !
@@ -42,7 +41,7 @@ program main
    character(MPI_MAX_PROCESSOR_NAME) :: pname
 !
 !..timer
-   real(8) :: MPI_Wtime, start_time, end_time
+   real(8) :: start_time, end_time
 !
    character       :: arg
    character(1024) :: cmd
