### 2024-02-21 9387725
resolve type mismatch in frontal solver
