### 2024-02-21 0f4434f
Enable using mpi_wrapper in src, test, and problems (#167)
