+++ trunk/problems/LASER/COMMON_FILES/exact_error.F90
@@ -17,8 +17,7 @@ subroutine exact_error(Nflag,PhysNick)
    use environment
    use assembly_sc, only: NRDOF_TOT,NRDOF_CON
    use par_mesh   , only: DISTRIBUTED,HOST_MESH
-   use mpi_param  , only: ROOT,RANK
-   use MPI        , only: MPI_SUM,MPI_COMM_WORLD,MPI_REAL8
+   use mpi_wrapper
 !
    implicit none
 !
