+++ trunk/problems/LASER/UW_COUPLED/elem/elem_maxwell_fi_pris.F90
@@ -55,7 +55,7 @@ subroutine elem_maxwell_fi_pris(Mdle,Fld_flag,                &
    use data_structure3D
    use laserParam
    use commonParam
-   use MPI, only: MPI_Wtime
+   use mpi_wrapper
 !..no implicit statements
    implicit none
 !..declare input/output variables
