+++ trunk/src/solver/omp_mumps/mumps_sc.F90
@@ -26,10 +26,6 @@ subroutine mumps_sc(mtype)
 !
    use data_structure3D, only: NRNODS, NRELES, ELEM_ORDER
    use assembly,         only: NR_RHS, MAXDOFM, MAXDOFS,       &
-                               MAXbrickH, MAXmdlbH,            &
-                               MAXbrickE, MAXmdlbE,            &
-                               MAXbrickV, MAXmdlbV,            &
-                               MAXbrickQ,                      &
                                NEXTRACT, IDBC, ZDOFD, ZERO,    &
                                ALOC, BLOC, AAUX, ZAMOD, ZBMOD, &
                                NR_PHYSA, MAXNODM
