+++ trunk/src/solver/frontal/base/complex/frwrs.F
@@ -120,7 +120,7 @@ c flag for solin2 (ifg=1: get rhs ; =2: get lhs as well)
 c
          ifg = 1
          md = int(Amdest(1))
-         call solin2 (iel, ifg, NDOFM, NRHS, Amdest, Elem, Elem)
+         call solin2 (iel, ifg, NDOFM, NRHS, md, Elem, Elem)
 c        -------------------------------------------------------
 c debug print
 c
