+++ trunk/problems/MPI_TEST/TEST/elem.F90
@@ -22,4 +22,13 @@
 !
 subroutine elem(Mdle, Itest,Itrial)
 !
+   use physics, only: NR_PHYSA
+!
+   implicit none
+!
+   integer, intent(in) :: Mdle
+   integer, dimension(NR_PHYSA), intent(out) :: Itest,Itrial
+!
+   Itest = 1; Itrial = 1
+!
 end subroutine elem
