+++ trunk/src/graphics/x11/interf/grint.F
@@ -105,7 +105,8 @@ c.........switch to any active window
       endif
 c
       return
-      end
+      end subroutine closwind
+c
 c----------------------------------------------------------------------
 c
 c   routine name       - openwind
@@ -254,7 +255,9 @@ c
   123 continue
 c
       return
-      end
+      end subroutine openwind
+c
+c
       subroutine selwind(Iwin)
 c
 c-----------------------------------------------------------------------
@@ -318,8 +321,10 @@ c        if (LWTYPE(ICURRWIN).eq.1) call wvdfcl(iclip)
       endif
 c
       return
-      end
-      subroutine  drawline(X1,Y1,X2,Y2,Ilcol)
+      end subroutine selwind
+c
+c
+      subroutine drawline(X1,Y1,X2,Y2,Ilcol)
 c
 c---------------------------------------------------------------------
 c
@@ -402,8 +407,9 @@ c          if (LWTYPE(ICURRWIN).eq.1) call wvdfdl(sx1,sy1,sx2,sy2,il)
       endif
 c
       return
-      end
-
+      end subroutine drawline
+c
+c
       subroutine drawpoly(N,Coor,Ilcol)
 c
 c-----------------------------------------------------------------------c
@@ -434,8 +440,9 @@ c
       call fillpoly(N,Coor,-1,Ilcol)
 c
       return
-      end
-
+      end subroutine drawpoly
+c
+c
       subroutine fillpoly(Nin,Coor,Ifcol,Ilcol)
 c
 c-----------------------------------------------------------------------
@@ -519,8 +526,9 @@ c..ldem        call wvpsfp(n,xcoor,il,if)
       endif
 c
       return
-      end
-
+      end subroutine fillpoly
+c
+c
       subroutine symbol(X1,Y1,Height,Mess,Angle,N,Icol)
 c
 c-----------------------------------------------------------------------c
@@ -590,7 +598,7 @@ c
       endif
 c      write(*,*) 'symbol2'
 c
-      if (IGPSP(ICURRWIN).eq.1) THEN
+      if (IGPSP(ICURRWIN).eq.1) then
         if (LSCALE(ICURRWIN).eq.0) then
           xx1 = X1*10.0/MAXSCX
           xy1 = Y1*10.0/MAXSCX
@@ -605,8 +613,9 @@ c        write(*,*) 'before wvpste'
 c      write(*,*) 'symbol3'
 c
       return
-      end
-
+      end subroutine symbol
+c
+c
       subroutine setpost(ionoff)
 c
 c-----------------------------------------------------------------------c
@@ -650,7 +659,9 @@ c
       endif
 c
       return
-      end
+      end subroutine setpost
+c
+c
       subroutine clrwind(Iback)
 c
 c-----------------------------------------------------------------------c
@@ -690,9 +701,9 @@ c
       LSCALE(ICURRWIN)=is
 c
       return
-      end
-
-
+      end subroutine clrwind
+c
+c
       subroutine scaleon(Ionoff)
 c
 c-----------------------------------------------------------------------c
@@ -722,7 +733,8 @@ c
       LSCALE(ICURRWIN) = Ionoff
 c
       return
-      end
+      end subroutine scaleon
+c
       subroutine newscale(Range)
 c
 c-----------------------------------------------------------------------c
@@ -759,7 +771,7 @@ c
       call scaleon(1)
 c
       return
-      end
+      end subroutine newscale
 
       subroutine initcm
 c
@@ -811,7 +823,7 @@ c##   30     continue
 c##          return
 c##      endif
 c##c
-c##      if (ISET.EQ.1) THEN
[…]
