### 2024-02-24 09acebb
disable verbose interactive mode output in model problems
