### 2024-02-24 51f8ce7
Fix two minor issues (#171)
* define module graphmod when X11 is disabled.

* disable verbose interactive-mode output in model problems.
