### 2024-02-24 a1149d5
more cleanup in model problems
