### 2024-02-24 a821a9d
unused variable cleanup
