+++ trunk/problems/SHEATHED_HOSE/elem_residual.F90
@@ -110,7 +110,6 @@
       character uplo
       integer :: nk
       nk(k1,k2) = (k2-1)*k2/2+k1
-
 !
 !-----------------------------------------------------------------------------------
 !      I N I T I A L I Z A T I O N                                                 |
