### 2024-02-24 e7f53a6
adding new utility for computing packed indices
