### 2024-02-24 dcdc70b
replace obsolete statement functions
