+++ trunk/problems/MAXWELL/ULTRAWEAK_DPG/elem/elem_maxwell.F90
@@ -139,19 +139,17 @@
       complex(8) :: zc1
       complex(8) :: za(3,3), zc(3,3)
 !
-#if DEBUG_MODE
-      integer :: iprint
-#endif
+      integer, external :: ij_upper_to_packed
 !
-!  ...for Gram matrix compressed storage format
-      integer :: nk
-      nk(k1,k2) = (k2-1)*k2/2+k1
+#if DEBUG_MODE
+!  ...Set iprint = 0/1 (Non-/VERBOSE)
+      integer :: iprint
+      iprint = 0
+#endif
 !
 !-------------------------------------------------------------------------------
 !
 #if DEBUG_MODE
-!  ...Set iprint = 0/1 (Non-/VERBOSE)
-      iprint = 0
       if (iprint.eq.1) then
          write(*,*) 'elem_maxwell: Mdle = ', Mdle
       endif
@@ -341,7 +339,7 @@
 !             (F_j,F_i) terms = Int[F_^*i F_j] terms (G_11)
                n = 2*k1-1; m = 2*k2-1
 !
-               k = nk(n,m)
+               k = ij_upper_to_packed(n,m)
                zaux = conjg(epsTfldF(1))*epsTfldE(1) + &
                       conjg(epsTfldF(2))*epsTfldE(2) + &
                       conjg(epsTfldF(3))*epsTfldE(3)
@@ -350,7 +348,7 @@
 !
 !              (G_j,F_i) terms = Int[F_^*i G_j] terms (G_12)
                n = 2*k1-1; m = 2*k2
-               k = nk(n,m)
+               k = ij_upper_to_packed(n,m)
                zaux = - (fldF(1)*epscrlE(1) + &
                          fldF(2)*epscrlE(2) + &
                          fldF(3)*epscrlE(3) )
@@ -364,7 +362,7 @@
                if (k1 .ne. k2) then
 !                 (F_j,G_i) terms = Int[G_^*i F_j] terms (G_21)
                   n = 2*k1; m = 2*k2-1
-                  k = nk(n,m)
+                  k = ij_upper_to_packed(n,m)
                   zaux = - (crlF(1)*epsTfldE(1) + &
                             crlF(2)*epsTfldE(2) + &
                             crlF(3)*epsTfldE(3) )
@@ -376,7 +374,7 @@
 !
 !              (G_j,G_i) terms = Int[G_^*i G_j] terms (G_22)
                n = 2*k1; m = 2*k2
-               k = nk(n,m)
+               k = ij_upper_to_packed(n,m)
                zcux = abs(zc1)**2*(fldF(1)*fldE(1) + &
                                    fldF(2)*fldE(2) + &
                                    fldF(3)*fldE(3) )
