### 2024-02-24 d0d7956
cleanup unused vars
