+++ trunk/problems/LASER/COMMON_FILES/norms.F90
@@ -13,9 +13,8 @@
 !-------------------------------------------------------------------
 subroutine get_L2NormDiff(Attr1,Attr2, L2NormDiff)
 !
-   use control         , only: NEXACT
    use data_structure3D
-   use environment     , only: QUIET_MODE,L2PROJ,FILE_ERR
+   use environment     , only: QUIET_MODE
    use physics
    use par_mesh        , only: DISTRIBUTED,HOST_MESH
    use mpi_wrapper
@@ -28,7 +27,7 @@ subroutine get_L2NormDiff(Attr1,Attr2, L2NormDiff)
    real(8) :: currL2NormDiff
 !
 !..miscellanea
-   integer :: mdle,nint,nrdof_tot,ic,iel
+   integer :: mdle,iel
 !
 !..auxiliary
    integer :: count,ierr
@@ -133,7 +132,6 @@ subroutine get_elem_L2NormDiff(Mdle,Attr1,Attr2, L2NormDiff)
 !
    use control          , only : INTEGRATION
    use data_structure3D
-   use environment      , only : L2PROJ
    use physics
 !
    implicit none
@@ -173,7 +171,7 @@ subroutine get_elem_L2NormDiff(Mdle,Attr1,Attr2, L2NormDiff)
    VTYPE, dimension(  MAXEQNQ  ) ::  zsolQ
 !
 !..miscellanea
-   integer :: nint,icase,iflag,l,i,j,ibeg1,ibeg2,icomp,ndom,ivar,nflag
+   integer :: nint,iflag,l,i,j,ibeg1,ibeg2,icomp,ndom,nflag
    real(8) :: weight,wa
 !
 !---------------------------------------------------------------------------------------
@@ -261,9 +259,8 @@ end subroutine get_elem_L2NormDiff
 !-------------------------------------------------------------------
 subroutine get_L2NormAttr(Attr, L2Norm)
 !
-   use control         , only: NEXACT
    use data_structure3D
-   use environment     , only: QUIET_MODE,L2PROJ,FILE_ERR
+   use environment     , only: QUIET_MODE
    use physics
    use par_mesh        , only: DISTRIBUTED,HOST_MESH
    use mpi_wrapper
@@ -276,7 +273,7 @@ subroutine get_L2NormAttr(Attr, L2Norm)
    real(8) :: currL2Norm
 !
 !..miscellanea
-   integer :: mdle,nint,nrdof_tot,ic,iel
+   integer :: mdle,iel
 !
 !..auxiliary
    integer :: count,ierr
@@ -364,7 +361,6 @@ subroutine get_elem_L2NormAttr(Mdle,Attr, L2Norm)
 !
    use control          , only : INTEGRATION
    use data_structure3D
-   use environment      , only : L2PROJ
    use physics
 !
    implicit none
@@ -404,7 +400,7 @@ subroutine get_elem_L2NormAttr(Mdle,Attr, L2Norm)
    VTYPE, dimension(  MAXEQNQ  ) ::  zsolQ
 !
 !..miscellanea
-   integer :: nint,icase,iflag,l,i,ibeg,icomp,ndom,ivar,nflag
+   integer :: nint,iflag,l,i,ibeg,icomp,ndom,nflag
    real(8) :: weight,wa
 !
 !---------------------------------------------------------------------------------------
@@ -496,9 +492,8 @@ end subroutine get_elem_L2NormAttr
 !-------------------------------------------------------------------
 subroutine get_Norm(Flag, FieldNormH,FieldNormE,FieldNormV,FieldNormQ)
 !
-   use control          , only: NEXACT
[…]
