+++ trunk/problems/POISSON/ULTRAWEAK_DPG/soldis.F90
@@ -5,124 +5,130 @@
 !! @param[in] Xi     - master element coordinates
 !! @param[in] X      - physical coordinates
 !! @param[in] Rn     - outward normal unit vector
-!! @param[in] ZsolH  - H1    sol
-!! @param[in] ZgradH - H1    grad
-!! @param[in] ZsolE  - Hcurl sol
-!! @param[in] ZcurlE - Hcurl curl
-!! @param[in] ZsolV  - Hdiv  sol
-!! @param[in] ZdivV  - Hdiv  div
-!! @param[in] ZsolQ  - L2    sol
+!! @param[in] RsolH  - H1    sol
+!! @param[in] RgradH - H1    grad
+!! @param[in] RsolE  - Hcurl sol
+!! @param[in] RcurlE - Hcurl curl
+!! @param[in] RsolV  - Hdiv  sol
+!! @param[in] RdivV  - Hdiv  div
+!! @param[in] RsolQ  - L2    sol
 !
-!! @param[out] val   - quantity to display
+!! @param[out] Val   - quantity to display
 !---------------------------------------------------------------------------------------
 !
 #include "typedefs.h"
 !
 subroutine soldis(Mdle,Xi,X,Rn,RsolH,RgradH,RsolE,RcurlE,RsolV,RdivV,RsolQ, Val)
 !
-    use data_structure3D
-    use common_prob_data
+   use data_structure3D
+   use common_prob_data
 !
-    implicit none
+   implicit none
 !
-    integer,                       intent(in)  :: Mdle
-    real(8),dimension(3),          intent(in)  :: Xi,X,Rn
-    VTYPE  ,dimension(  MAXEQNH  ),intent(in)  :: RsolH
-    VTYPE  ,dimension(  MAXEQNH,3),intent(in)  :: RgradH
-    VTYPE  ,dimension(3,MAXEQNE  ),intent(in)  :: RsolE
-    VTYPE  ,dimension(3,MAXEQNE  ),intent(in)  :: RcurlE
-    VTYPE  ,dimension(3,MAXEQNV  ),intent(in)  :: RsolV
-    VTYPE  ,dimension(  MAXEQNV  ),intent(in)  :: RdivV
-    VTYPE  ,dimension(  MAXEQNQ  ),intent(in)  :: RsolQ
-    real(8),                       intent(out) :: Val
+   integer,                       intent(in)  :: Mdle
+   real(8),dimension(3),          intent(in)  :: Xi,X,Rn
+   VTYPE  ,dimension(  MAXEQNH  ),intent(in)  :: RsolH
+   VTYPE  ,dimension(  MAXEQNH,3),intent(in)  :: RgradH
+   VTYPE  ,dimension(3,MAXEQNE  ),intent(in)  :: RsolE
+   VTYPE  ,dimension(3,MAXEQNE  ),intent(in)  :: RcurlE
+   VTYPE  ,dimension(3,MAXEQNV  ),intent(in)  :: RsolV
+   VTYPE  ,dimension(  MAXEQNV  ),intent(in)  :: RdivV
+   VTYPE  ,dimension(  MAXEQNQ  ),intent(in)  :: RsolQ
+   real(8),                       intent(out) :: Val
+!
+#if DEBUG_MODE
+   integer :: iprint
+   iprint = 0
+#endif
 !
 !---------------------------------------------------------------------------------------
 !
-    integer :: iprint
-
-    select case (IEXACT_DISP)
-        case(0)
-            select case (ICHOOSE_COMP)
-                case(1)
-                    Val = (RsolQ(1))
-                case(2)
-                    Val = (RsolQ(2))
[…]
