### 2024-02-24 6e57d01
minor update to CI workflow
