+++ trunk/src/solver/par_mumps/par_mumps.F90
@@ -48,7 +48,11 @@ subroutine mumps_start(mumps,mumps_comm)
    type (DMUMPS_STRUC), intent(inout) :: mumps
 #endif
 !
+#if HP3D_USE_MPI_F08
    type(MPI_Comm), intent(in) :: mumps_comm
+#else
+   integer       , intent(in) :: mumps_comm
+#endif
 !
 !..NOTE:
 !..If you plan to run MUMPS sequentially from a parallel MPI application,
@@ -56,7 +60,11 @@ subroutine mumps_start(mumps,mumps_comm)
 !  communicator containing a single processor to the MUMPS library
 !
 !..Define a communicator for the package.
+#if HP3D_USE_MPI_F08
    mumps%COMM = mumps_comm%MPI_VAL
+#else
+   mumps%COMM = mumps_comm
+#endif
 !
 !..PAR
 !     0 : host is not involved in factorization/solve phases
@@ -195,7 +203,11 @@ subroutine mumps_start_subd
 !..this routine initiates sparse solver for interior of subdomain
 !
 !..Define a communicator for the package.
+#if HP3D_USE_MPI_F08
    mumps_bub%COMM = MPI_COMM_SELF%MPI_VAL
+#else
+   mumps_bub%COMM = MPI_COMM_SELF
+#endif
 !
 !..PAR
 !     0 : host is not involved in factorization/solve phases
