### 2024-02-25 7f214f2
updating CI workflow configs
