### 2024-02-25 fe7a47a
Add MPI F08 binding builds and tests to CI workflow
