+++ .github/workflows/builds-and-tests.yml
@@ -127,10 +127,11 @@ jobs:
       # Build hp3D (MPI F90)
       - name: Build hp3D (MPI F90)
         run: |
-          echo "Configuring hp3D (MPI F90)"
+          export HP3D_USE_MPI_F08=0
+
+          echo "Configuring hp3D"
           cd trunk
           cp m_options_files/workflows/m_options_$CONFIG ./m_options
-          export HP3D_USE_MPI_F08=0
 
           echo "Building hp3D"
           make
@@ -141,17 +142,19 @@ jobs:
       # Run hp3D tests (MPI F90)
       - name: Run hp3D tests (MPI F90)
         run: |
-          echo "Running hp3D tests (MPI F90)"
-          cd trunk
           export HP3D_USE_MPI_F08=0
+
+          echo "Running hp3D tests"
+          cd trunk
           make test
 
       # Build hp3D problems (MPI F90)
       - name: Build hp3D problems (MPI F90)
         run: |
+          export HP3D_USE_MPI_F08=0
+
           echo "Building hp3D problems"
           cd trunk
-          export HP3D_USE_MPI_F08=0
           make problems
 
       # Clean hp3D directory
@@ -164,10 +167,11 @@ jobs:
       # Build hp3D (MPI F08)
       - name: Build hp3D (MPI F08)
         run: |
-          echo "Configuring hp3D (MPI F08)"
+          export HP3D_USE_MPI_F08=1
+
+          echo "Configuring hp3D"
           cd trunk
           cp m_options_files/workflows/m_options_$CONFIG ./m_options
-          export HP3D_USE_MPI_F08=1
 
           echo "Building hp3D"
           make
@@ -178,17 +182,19 @@ jobs:
       # Run hp3D tests (MPI F08)
       - name: Run hp3D tests (MPI F08)
         run: |
-          echo "Running hp3D tests (MPI F90)"
-          cd trunk
           export HP3D_USE_MPI_F08=1
+
+          echo "Running hp3D tests"
+          cd trunk
           make test
 
       # Build hp3D problems (MPI F08)
       - name: Build hp3D problems (MPI F08)
         run: |
+          export HP3D_USE_MPI_F08=1
+
           echo "Building hp3D problems"
           cd trunk
-          export HP3D_USE_MPI_F08=1
           make problems
 
       # Report job status
