+++ trunk/src/solver/par_mumps/par_fiber.F90
@@ -73,9 +73,14 @@ recursive subroutine par_fiber(mumps,nrdof,nproc,level)
 !
 !..aux variables for subcommunicators
    integer :: mRANK,mPROCS,mSUB_RANK,mINT_PROCS,mINT_RANK
-   integer :: mpi_comm_sub,mpi_comm_int
    integer :: group,key,color,count,src,rcv,tag,ierr
    integer :: nrdof_int(nproc/mSUB_PROCS)
+!
+#if HP3D_USE_MPI_F08
+   type(MPI_Comm) :: mumps_comm, mpi_comm_sub, mpi_comm_int
+#else
+   integer        :: mumps_comm, mpi_comm_sub, mpi_comm_int
+#endif
 !
 !..aux variables for subproblem assembly
    integer :: dof_off,dof_sub,dof_int,dof_int_L,dof_int_R
@@ -106,8 +111,14 @@ recursive subroutine par_fiber(mumps,nrdof,nproc,level)
 !
  123 format('[',I3,'] ',A10,': ',I8)
 !
-   call MPI_COMM_RANK(mumps%COMM, mRANK ,ierr)
-   call MPI_COMM_SIZE(mumps%COMM, mPROCS,ierr)
+#if HP3D_USE_MPI_F08
+   mumps_comm%MPI_VAL = mumps%COMM
+#else
+   mumps_comm         = mumps%COMM
+#endif
+!
+   call MPI_COMM_RANK(mumps_comm, mRANK ,ierr)
+   call MPI_COMM_SIZE(mumps_comm, mPROCS,ierr)
    if (mPROCS .ne. nproc) then
       write(*,*) 'par_fiber: mPROCS != nproc : ',mPROCS,nproc
    endif
@@ -135,14 +146,14 @@ recursive subroutine par_fiber(mumps,nrdof,nproc,level)
       goto 90
    endif
 !
-   call MPI_BARRIER(mumps%COMM, ierr); start_time = MPI_Wtime()
+   call MPI_BARRIER(mumps_comm, ierr); start_time = MPI_Wtime()
 !
 !..divide into "small enough" subproblems (groups) and solve separately
 !  [Step 1: define communicators for the subproblems]
    group = mRANK / mSUB_PROCS
    color = group
    key   = MOD(mRANK,mSUB_PROCS)
-   call MPI_COMM_SPLIT(mumps%COMM,color,key, mpi_comm_sub,ierr)
+   call MPI_COMM_SPLIT(mumps_comm,color,key, mpi_comm_sub,ierr)
    call MPI_COMM_RANK(mpi_comm_sub, mSUB_RANK,ierr)
    if (mSUB_RANK .ne. key) then
       write(*,5050) 'RANK, mSUB_RANK, key = ',RANK,mSUB_RANK,key
@@ -155,7 +166,7 @@ recursive subroutine par_fiber(mumps,nrdof,nproc,level)
    color = MPI_UNDEFINED; mINT_RANK = -1
    key   = group
    if (mSUB_RANK .eq. ROOT) color = 0
-   call MPI_COMM_SPLIT(mumps%COMM,color,key, mpi_comm_int,ierr)
+   call MPI_COMM_SPLIT(mumps_comm,color,key, mpi_comm_int,ierr)
    if (mSUB_RANK .eq. ROOT) then
       call MPI_COMM_RANK(mpi_comm_int, mINT_RANK ,ierr)
       call MPI_COMM_SIZE(mpi_comm_int, mINT_PROCS,ierr)
[…]
