### 2024-02-25 4db4a2f
updating DEBUG and COMPLEX flags everywhere
