+++ trunk/src/solver/frontal/base/real/symelmR.F
@@ -49,7 +49,7 @@ c
       real(8), parameter :: dzero = 0.0d0
       real(8), parameter :: sml2  = 1.d-15
 c
-#if DEBUG_MODE
+#if HP3D_DEBUG
       integer :: iprint
       iprint=0
 c
@@ -75,7 +75,7 @@ c pull the diagonal entry
 c
       pivot = Flhs(mp)
 c
-#if DEBUG_MODE
+#if HP3D_DEBUG
       if(IPRPIV .eq. 1 .and. iprint .eq. 1) then
          write(NFSOUT,1000) Iel,nfw,id,pivot
  1000    format(5x,'IEL,NFW,ID,PIVOT',3i5,1pe11.3)
