+++ trunk/src/hpinterp/face/dhpfaceV.F90
@@ -286,13 +286,13 @@ subroutine dhpfaceV(Mdle,Iflag,No,Etav,Ntype,Icase,Bcond,   &
       write(*,7015) j, zbV(j,1:MAXEQNV)
       write(*,7016) aaV(j,1:ndofV_face)
     enddo
-    #if HP3D_COMPLEX
-      7015 format(i5, 6(2e10.3,2x))
-    #else
-      7015 format(i5, 10e12.5)
-    #endif
-    7016 format(10e12.5)
   endif
+#if HP3D_COMPLEX
+  7015 format(i5,2x,6(2e10.3,2x))
+#else
+  7015 format(i5,2x,10e12.5)
+#endif
+  7016 format(10e12.5)
 #endif
 !
 !-----------------------------------------------------------------------
