+++ trunk/src/hpinterp/edge/dhpedgeH.F90
@@ -271,13 +271,13 @@ subroutine dhpedgeH(Mdle,Iflag,No,Etav,Ntype,Icase,Bcond,   &
       write(*,7015) j, zbH(j,1:MAXEQNH)
       write(*,7016) aaH(j,1:ndofH_edge)
     enddo
-    #if HP3D_COMPLEX
-      7015 format(i5, 6(2e10.3,2x))
-    #else
-      7015 format(i5, 10e12.5)
-    #endif
-    7016 format(10e12.5)
   endif
+#if HP3D_COMPLEX
+  7015 format(i5,2x,6(2e10.3,2x))
+#else
+  7015 format(i5,2x,10e12.5)
+#endif
+  7016 format(10e12.5)
 #endif
 !
 ! projection matrix leading dimension (maximum number of 1D bubbles)
