+++ trunk/src/utility/invert.F90
@@ -5,73 +5,73 @@
 !> @brief Computes determinant of real-valued 3x3 matrix
 !> @date Mar 2023
 subroutine get_ddet(A, Adet)
-  implicit none
-  real(8), dimension(3,3), intent(in)  :: A
-  real(8),                 intent(out) :: Adet
-  Adet = A(1,1)*A(2,2)*A(3,3) &
-       + A(2,1)*A(3,2)*A(1,3) &
-       + A(3,1)*A(1,2)*A(2,3) &
-       - A(3,1)*A(2,2)*A(1,3) &
-       - A(1,1)*A(3,2)*A(2,3) &
-       - A(2,1)*A(1,2)*A(3,3)
+   implicit none
+   real(8), intent(in)  :: A(3,3)
+   real(8), intent(out) :: Adet
+   Adet = A(1,1)*A(2,2)*A(3,3) &
+        + A(2,1)*A(3,2)*A(1,3) &
+        + A(3,1)*A(1,2)*A(2,3) &
+        - A(3,1)*A(2,2)*A(1,3) &
+        - A(1,1)*A(3,2)*A(2,3) &
+        - A(2,1)*A(1,2)*A(3,3)
 end subroutine get_ddet
 !
 !--------------------------------------------------------------------
 !> @brief Computes determinant of complex-valued 3x3 matrix
 !> @date Mar 2023
 subroutine get_zdet(A, Adet)
-  implicit none
-  VTYPE, dimension(3,3), intent(in)  :: A
-  VTYPE,                 intent(out) :: Adet
-  Adet = A(1,1)*A(2,2)*A(3,3) &
-       + A(2,1)*A(3,2)*A(1,3) &
-       + A(3,1)*A(1,2)*A(2,3) &
-       - A(3,1)*A(2,2)*A(1,3) &
-       - A(1,1)*A(3,2)*A(2,3) &
-       - A(2,1)*A(1,2)*A(3,3)
+   implicit none
+   VTYPE, intent(in)  :: A(3,3)
+   VTYPE, intent(out) :: Adet
+   Adet = A(1,1)*A(2,2)*A(3,3) &
+        + A(2,1)*A(3,2)*A(1,3) &
+        + A(3,1)*A(1,2)*A(2,3) &
+        - A(3,1)*A(2,2)*A(1,3) &
+        - A(1,1)*A(3,2)*A(2,3) &
+        - A(2,1)*A(1,2)*A(3,3)
 end subroutine get_zdet
 !
 !--------------------------------------------------------------------
 !> @brief Computes inverse of real-valued 3x3 matrix
 !> @date Mar 2023
 subroutine dinvert(A, Ainv,Adet)
-  implicit none
-  real(8), dimension(3,3), intent(in)  :: A
-  real(8), dimension(3,3), intent(out) :: Ainv
-  real(8),                 intent(out) :: Adet
-  real(8) :: det(3)
-  logical, external :: dnear
+   implicit none
+   real(8), dimension(3,3), intent(in)  :: A
+   real(8), dimension(3,3), intent(out) :: Ainv
+   real(8),                 intent(out) :: Adet
+   real(8) :: det(3)
+   logical, external :: dnear
 
[…]
