+++ trunk/src/utility/fact.F90
@@ -10,13 +10,13 @@ integer function fact(N)
 !
    integer :: i
 !
-   if (n.lt.0) then
+   if     (N.lt.0) then
       fact = 0
-   elseif (n.eq.0) then
+   elseif (N.eq.0) then
       fact = 1
    else
       fact = 1
-      do i = 1,n
+      do i = 1,N
          fact = fact*i
       enddo
    endif
