### 2024-02-25 62af052
Fix various compiler warnings (#169)
