+++ trunk/src/solver/frontal/base/real/elmsolR.F
@@ -67,7 +67,7 @@ c
       iprint=0
 #endif
 c
-      id = Aidx
+      id = int(Aidx)
       idm = id - 1
       idp = id + 1
 #if DEBUG_MODE
