+++ trunk/src/numeric/blas.F
@@ -9,12 +9,15 @@ c
       integer :: ip(n)
       integer :: Iflag
 c
-      integer :: i,j,k,imax,inc
+      integer :: j,k,imax,inc
       real(8) :: c,pivmin
+c
+      real(8), parameter :: eps = 1.0d-15
 c
       integer, external :: idamax
 c
 #if DEBUG_MODE
+      integer :: i
       integer :: iprint
       iprint=0
       if (iprint.eq.1) then
@@ -42,7 +45,7 @@ c
           call pause
         endif
         pivmin = dmin1(pivmin,abs(a(ip(k),k)) )
-        if (abs(a(ip(k),k)).ne.0.d0) go to 20
+        if (dabs(a(ip(k),k)) > eps) go to 20
         Iflag=1
         return
 c
@@ -86,10 +89,11 @@ c
       real(8) :: a(na,n),b(n),x(n)
       integer :: ip(n)
 c
-      integer :: inc,i,j,k
+      integer :: inc,j,k
       real(8) :: c
 c
 #if DEBUG_MODE
+      integer :: i
       integer :: iprint
       iprint = 0
 c
