+++ trunk/problems/LASER/UW_COUPLED/elem/elem_maxwell.F90
@@ -199,7 +199,7 @@ subroutine elem_maxwell(Mdle,Fld_flag,                &
 !..timer
 !   real(8) :: start_time,end_time
 !
-#if DEBUG_MODE
+#if HP3D_DEBUG
    integer :: iprint
    iprint = 0
 #endif
