### 2024-02-25 3e7b9a0
define variables for MPI launch
