### 2024-02-25 b6737ca
minor makefile fix
