### 2024-02-26 9e10e82
update x11 legacy code and other minor legacy updates
