+++ trunk/src/graphics/x11/base/int_coord.F90
@@ -1,61 +1,61 @@
 #if HP3D_USE_X11
 
-c----------------------------------------------------------------------
-c
-c   routine name       - int_coord
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine determines integer coordinates of
-c                        vertices for a subtriangle of a 2D element,
-c                        the corresponding color flag, and flags
-c                        indicating adjacency to element sides
-c
-c   arguments :
-c     in:
-c             Numlev   - a flag = 0 for mesh display,
-c                               > 0 for contour maps
-c             Nsub     - (even,.ge.4) number of subdivisions
-c             Ntype    - type of element
-c             Norder   - order of approximation (4 edges + 1 middle)
-c             I,J      - integer indices of the field
-c     out:
-c             Ixi1     - integer coordinates of the lower triangle
-c             Ixi2     - integer coordinates of the upper triangle
-c             Np1      - color code for the lower triangle
-c             Np2      - color code for the upper triangle
-c             Nsid1    - adjacency info for the lower triangle
-c             Nsid2    - adjacency info for the upper triangle
-c
-c----------------------------------------------------------------------
-c
-      subroutine int_coord(Numlev,Nsub,Ntype,Norder,I,J,
-     .                     Ixi1,Ixi2,Np1,Np2,Nsid1,Nsid2)
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - int_coord
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - routine determines integer coordinates of
+!                        vertices for a subtriangle of a 2D element,
+!                        the corresponding color flag, and flags
+!                        indicating adjacency to element sides
+!
+!   arguments :
+!     in:
+!             Numlev   - a flag = 0 for mesh display,
+!                               > 0 for contour maps
+!             Nsub     - (even,.ge.4) number of subdivisions
+!             Ntype    - type of element
+!             Norder   - order of approximation (4 edges + 1 middle)
+!             I,J      - integer indices of the field
+!     out:
+!             Ixi1     - integer coordinates of the lower triangle
+!             Ixi2     - integer coordinates of the upper triangle
+!             Np1      - color code for the lower triangle
+!             Np2      - color code for the upper triangle
+!             Nsid1    - adjacency info for the lower triangle
+!             Nsid2    - adjacency info for the upper triangle
+!
+!----------------------------------------------------------------------
+!
+      subroutine int_coord(Numlev,Nsub,Ntype,Norder,I,J, &
+                           Ixi1,Ixi2,Np1,Np2,Nsid1,Nsid2)
+!
       use node_types
-c
+!
       implicit none
-c
+!
       integer, intent(in)  :: Numlev,Nsub,Ntype
       integer, intent(in)  :: Norder(5)
[…]
