+++ trunk/src/graphics/x11/base/lsvisidb.F90
@@ -1,82 +1,82 @@
 #if HP3D_USE_X11
 
-c----------------------------------------------------------------------
-c
-c   routine name       - lsvisidb
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine makes a list of all visible
-c                        triangles and stores rescaled coordinates
-c                        of their end-points (vertices) in an array
-c                        RGTR
-c
-c   arguments :
-c     in:
-c               Numlev = 0 display graphical image of the mesh
-c                      > 0 number of levels to plot solution values
-c               Iflagn = 1 encode element (middle node) numbers
-c                      = 2 encode face nodes numbers
-c
-c----------------------------------------------------------------------
+!----------------------------------------------------------------------
+!
+!   routine name       - lsvisidb
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - routine makes a list of all visible
+!                        triangles and stores rescaled coordinates
+!                        of their end-points (vertices) in an array
+!                        RGTR
+!
+!   arguments :
+!     in:
+!               Numlev = 0 display graphical image of the mesh
+!                      > 0 number of levels to plot solution values
+!               Iflagn = 1 encode element (middle node) numbers
+!                      = 2 encode face nodes numbers
+!
+!----------------------------------------------------------------------
 #include "typedefs.h"
       subroutine lsvisidb(Numlev,Iflagn)
-c
+!
       use element_data
       use data_structure3D
       use graphmod
-c
+!
       implicit none
-c
+!
       integer, intent(in) :: Numlev,Iflagn
-c
-c  ...geometry and solution dof
+!
+!  ...geometry and solution dof
       real(8) :: xnod(3,MAXbrickH)
-      VTYPE   :: zdofH(MAXEQNH,MAXbrickH),zdofE(MAXEQNE,MAXbrickE),
-     .           zdofV(MAXEQNV,MAXbrickV),zdofQ(MAXEQNQ,MAXbrickQ)
-c
+      VTYPE   :: zdofH(MAXEQNH,MAXbrickH),zdofE(MAXEQNE,MAXbrickE),  &
+                 zdofV(MAXEQNV,MAXbrickV),zdofQ(MAXEQNQ,MAXbrickQ)
+!
       real(8) :: xmax(3),xmin(3)
       integer :: nedge_orient(12),nface_orient(6),norder(19)
-c
-c  ...element neighbors through faces
+!
+!  ...element neighbors through faces
       integer :: neig(4,6)
-c
-c  ...element nodes and their orientation
+!
+!  ...element nodes and their orientation
       integer :: nodesl(27),norientl(27)
       real(8) :: solev(NR_COLORS-10)
 
       integer :: nrinter(8)
       real(8) :: xlocinter(3,500)
       real(8) :: xcenter(3)
-c
+!
       real(8) :: cl,daux,daux1
       integer :: icut,iel,ifc,iloc,ineig,iv,ivar,ivis
       integer :: loc,mdle,ndom,nrintertot
-c
+!
       real(8), parameter :: bigp  =  1.d30
       real(8), parameter :: bign  = -1.d30
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint=0
 #endif
-c
-c---------------------------------------------------------------
-c
-c  ...find limiting values for contour plot
+!
[…]
