### 2024-02-26 f76a7c7
update x11 legacy code and other minor legacy updates
