+++ trunk/src/graphics/x11/interf/psgr.F
@@ -1,532 +1,0 @@
-#if HP3D_USE_X11
-
-C******************************************************
-C
-C THE IMPORTANT ROUTINES FOR THE POSTSCRIPT FILES
-C
-C wvopps - to open a ps file
-C wvpscl(clip(4)) - to set clipping region
-C wvpsdl(X1,Y1,X2,Y2,IC) - to draw a line
-C wvpsfp(N,Xc,Icf,Icl) - to fill a polygon
-C wvpspx(X,Y,IC) - to draw a pixel
-C wvpste(X1,Y1,H,MESS,A,N,ICOL) - to write text
-C wvpsout(x,y,icode) - to write ps parameters into buffer
-C flushps - to rewrite contents os the buffer into output ps file
-C wvclps - to close the ps file
-C
-C OTHER PARAMETERS ARE DISCUSSED IN THE ROUTINES
-C
-C********************************************************
-
-      subroutine wvopps
-C---------------------------------------------------------
-C      AUTHOR(S)        - Satish Chavva,Tadeusz Liszka and others
-C
-C      version  -   1998
-C
-C    purpose - to write header of ps file
-C
-C========================================================
-cwr07.08.99
-cwr03.14.02
-c
-      implicit real(8) (a-h,o-y)
-#include "wvglob.blk"
-#include "wvpost.blk"
-#include "wvscrn.blk"
-c
-      character(10) name
-c
-  110 write(*,*)  'Give name for Postscript file (0 - no postscript)'
-   10 read(*,1011) name
-
- 1011 format(a10)
-      if (name(1:1).le.' ') goto 10
-      if (name(1:1).eq.'0') then
-          iflag=0
-      else
-          ifile=79
-          open(ifile,file=name,status='new',err=100)
-          write(*,*)' PostScript send to ',name
-          write(ifile,1000) name
-c        Check if color Postscript required
-
-          write(*,*) 'B/W or  Color Postscript (0 - B/W  1-Color)'
-          read(*,*) ICOLVAL
-
- 1000     format('%!  TL PostScript file, name ',a10/
-     &        '%%BoundingBox: 36 36 518 756',/
-     &        '0 setlinewidth',/
-     &        '/n {newpath moveto} def',/
-     &        '/c {lineto closepath} def',/
-     &        '/m {moveto} def',/
-     &        '/d {lineto} def',/
-     &        '/f {setgray fill} def',/
-     &        '/s {0 setgray stroke} def',/
-     &        '/rf {setrgbcolor fill} def',/
-     &        '/rs {setrgbcolor stroke} def',/
-     &        '/gs {gsave} def',/
-     &        '/gr {grestore} def',/
-     &        '/fo { /Times-Roman findfont } def ',
-     &        '/sc { scalefont setfont } def'/
-     &        '50 750 translate -90 rotate fo 12 sc newpath')
-          ilbuf=0
-      endif
-      ilepsp=1
-c      write(*,*) 'NO POSTSCRIPT GENERATED, iflag=',iflag
-      return
-  100 continue
-      write(*,*) ' File ',name,' exists !!'
-      write(*,*) ' I will not overwrite it !!'
-      goto 110
-      end subroutine wvopps
-
-
-c-----------------------------------------------------------------
-      subroutine wvpscl(Clip)
-c
-c----------------------------------------------------------------------
-c
-c   computer           - machine independent
-c
-c   latest revision    -
-c
-c   purpose            - to set clipping region for the ps page
-c
-c
-c----------------------------------------------------------------------
-c
-      dimension clip(4)
-c
-#include "wvglob.blk"
-#include "wvpost.blk"
-#include "wvscrn.blk"
-c
-      iprint=0
-      if (iprint.eq.1) then
[…]
