### 2024-02-26 c960ca0
update legacy code in various problems
