+++ trunk/problems/SHEATHED_HOSE/module/sheathed_isotropic_materials.F90
@@ -7,31 +7,31 @@ module sheathed_isotropic_materials
   implicit none
 !
 !  G E O M E T R Y
-  real*8, parameter :: R_inside  = 0.5d0
-  real*8, parameter :: R_middle  = 0.99d0
-  real*8, parameter :: R_outside = 1.0d0
-  real*8, parameter :: X_1       = 0.0d0
-  real*8, parameter :: X_2       = 1.0d0
+  real(8), parameter :: R_inside  = 0.5d0
+  real(8), parameter :: R_middle  = 0.99d0
+  real(8), parameter :: R_outside = 1.0d0
+  real(8), parameter :: X_1       = 0.0d0
+  real(8), parameter :: X_2       = 1.0d0
 !
 !  M A T E R I A L   D A T A
-  real*8, parameter :: E_steel   = 2.0d4
-  real*8, parameter :: E_rubber  = 1.0d0
-  real*8, parameter :: NU_steel  = 0.285d0
-  real*8, parameter :: NU_rubber = 0.5d0
+  real(8), parameter :: E_steel   = 2.0d4
+  real(8), parameter :: E_rubber  = 1.0d0
+  real(8), parameter :: NU_steel  = 0.285d0
+  real(8), parameter :: NU_rubber = 0.5d0
 !
-  real*8, parameter :: LAMBDA_steel  = E_steel*NU_steel/((1+NU_steel)*(1-2*NU_steel))
-  real*8, parameter :: MU_steel      = 0.5d0*E_steel/(1+NU_steel)
-  real*8, parameter :: MU_rubber     = 0.5d0*E_rubber/(1+NU_rubber)
+  real(8), parameter :: LAMBDA_steel  = E_steel*NU_steel/((1+NU_steel)*(1-2*NU_steel))
+  real(8), parameter :: MU_steel      = 0.5d0*E_steel/(1+NU_steel)
+  real(8), parameter :: MU_rubber     = 0.5d0*E_rubber/(1+NU_rubber)
 !
-  real*8, parameter :: RHO   = 0.d0
-  real*8, parameter :: OMEG  = 0.d0
+  real(8), parameter :: RHO   = 0.d0
+  real(8), parameter :: OMEG  = 0.d0
 !
 !  Pressure
-  real*8, parameter :: P_inner =-1.d-1
-  real*8, parameter :: P_outer = 0.d0
+  real(8), parameter :: P_inner =-1.d-1
+  real(8), parameter :: P_outer = 0.d0
 
 !  Kronecker's delta
-  real*8, parameter, dimension(3,3) :: del =  &
+  real(8), parameter, dimension(3,3) :: del =  &
       reshape((/ 1.d0,0.d0,0.d0,0.d0,1.d0,0.d0,0.d0,0.d0,1.d0 /),(/3,3/))
 
 contains
@@ -47,12 +47,12 @@ contains
   subroutine getC(X,Dom, C)
     implicit none
   !------------------------------------------------------------------------------
-    real*8,dimension(3),       intent(in)  :: X
+    real(8),dimension(3),       intent(in)  :: X
     integer,                   intent(in)  :: Dom
-    real*8,dimension(3,3,3,3), intent(out) :: C
+    real(8),dimension(3,3,3,3), intent(out) :: C
   !------------------------------------------------------------------------------
     integer :: i,j,k,l
-    real*8  :: mu,lambda
+    real(8)  :: mu,lambda
   !------------------------------------------------------------------------------
   !
   if (Dom.eq.1) then
@@ -87,12 +87,12 @@ contains
   subroutine getA(X,Dom, A)
     implicit none
   !------------------------------------------------------------------------------
-    real*8,dimension(3),       intent(in)  :: X
+    real(8),dimension(3),       intent(in)  :: X
[…]
