+++ trunk/problems/SHEATHED_HOSE/common/getf.F90
@@ -12,15 +12,15 @@ subroutine getf(Mdle,X, Fval)
   implicit none
 !------------------------------------------------------------------------------
   integer,                       intent(in)  :: Mdle
-  real*8, dimension(3),          intent(in)  :: X
-  real*8, dimension(1:3,NR_RHS), intent(out) :: Fval
+  real(8), dimension(3),          intent(in)  :: X
+  real(8), dimension(1:3,NR_RHS), intent(out) :: Fval
 !------------------------------------------------------------------------------
 ! solution variables
-  real*8, dimension(3)   :: u
-  real*8, dimension(3,3) :: gradu
-  real*8, dimension(3,3) :: epsilon
-  real*8, dimension(3,3) :: sigma
-  real*8, dimension(3)   :: divsigma
+  real(8), dimension(3)   :: u
+  real(8), dimension(3,3) :: gradu
+  real(8), dimension(3,3) :: epsilon
+  real(8), dimension(3,3) :: sigma
+  real(8), dimension(3)   :: divsigma
 ! counters
   integer :: iload
 !------------------------------------------------------------------------------
@@ -72,16 +72,16 @@ subroutine getg(Mdle,Ibc,X,Rn, Gval)
   implicit none
 !------------------------------------------------------------------------------
   integer,                     intent(in)  :: Mdle
-  real*8, dimension(3),        intent(in)  :: X,Rn
+  real(8), dimension(3),        intent(in)  :: X,Rn
   integer,                     intent(in)  :: Ibc
-  real*8, dimension(3,NR_RHS), intent(out) :: Gval
+  real(8), dimension(3,NR_RHS), intent(out) :: Gval
 !------------------------------------------------------------------------------
 ! solution variables
-  real*8, dimension(3)   :: u
-  real*8, dimension(3,3) :: gradu
-  real*8, dimension(3,3) :: epsilon
-  real*8, dimension(3,3) :: sigma
-  real*8, dimension(3)   :: divsigma
+  real(8), dimension(3)   :: u
+  real(8), dimension(3,3) :: gradu
+  real(8), dimension(3,3) :: epsilon
+  real(8), dimension(3,3) :: sigma
+  real(8), dimension(3)   :: divsigma
 ! counters
   integer :: iload,j
 !------------------------------------------------------------------------------
