+++ trunk/problems/SHEATHED_HOSE/paraview/paraview_attr_custom.F90
@@ -20,7 +20,7 @@ subroutine paraview_attr_custom(Id,Scomp)
       integer, intent(in) :: Id
       integer, intent(in) :: Scomp
 !
-      character*8, parameter, dimension(3) :: sAttr = (/'sigma_rr','sigma_rt','sigma_tt'/)
+      character(len=8), parameter, dimension(3) :: sAttr = (/'sigma_rr','sigma_rt','sigma_tt'/)
       character(len=60) :: fname,nick
       integer :: ic
       character(len=5) :: postfix
