### 2024-02-26 625c6d3
update legacy code in various problems
