+++ trunk/problems/SHEATHED_HOSE/set_initial_mesh.F90
@@ -19,7 +19,6 @@ subroutine set_initial_mesh(Nelem_order)
    use GMP
    use data_structure3D
    use common_prob_data, only : IBC_PROB,IP
-   use mpi_param, only: RANK,ROOT
 !------------------------------------------------------------------------------------
    implicit none
 !
