+++ trunk/problems/SHEATHED_HOSE/soldis.F90
@@ -21,32 +21,32 @@ subroutine soldis(Mdle,Xi,X,Rn,ZsolH,ZgradH,ZsolE,ZcurlE,ZsolV,ZdivV,ZsolQ, Val)
 !---------------------------------------------------------------------------------------
   implicit none
   integer,                      intent(in)  :: Mdle
-  real*8,dimension(3),          intent(in)  :: Xi,X,Rn
-  real*8,dimension(  MAXEQNH  ),intent(in)  :: ZsolH
-  real*8,dimension(  MAXEQNH,3),intent(in)  :: ZgradH
-  real*8,dimension(3,MAXEQNE  ),intent(in)  :: ZsolE
-  real*8,dimension(3,MAXEQNE  ),intent(in)  :: ZcurlE
-  real*8,dimension(3,MAXEQNV  ),intent(in)  :: ZsolV
-  real*8,dimension(  MAXEQNV  ),intent(in)  :: ZdivV
-  real*8,dimension(  MAXEQNQ  ),intent(in)  :: ZsolQ
-  real*8,                       intent(out) :: Val
+  real(8),dimension(3),          intent(in)  :: Xi,X,Rn
+  real(8),dimension(  MAXEQNH  ),intent(in)  :: ZsolH
+  real(8),dimension(  MAXEQNH,3),intent(in)  :: ZgradH
+  real(8),dimension(3,MAXEQNE  ),intent(in)  :: ZsolE
+  real(8),dimension(3,MAXEQNE  ),intent(in)  :: ZcurlE
+  real(8),dimension(3,MAXEQNV  ),intent(in)  :: ZsolV
+  real(8),dimension(  MAXEQNV  ),intent(in)  :: ZdivV
+  real(8),dimension(  MAXEQNQ  ),intent(in)  :: ZsolQ
+  real(8),                       intent(out) :: Val
 !---------------------------------------------------------------------------------------
 ! exact solution workspace
   integer                         :: icase
-  real*8,dimension(  MAXEQNH    ) ::   zvalH
-  real*8,dimension(  MAXEQNH,3  ) ::  zdvalH
-  real*8,dimension(  MAXEQNH,3,3) :: zd2valH
-  real*8,dimension(3,MAXEQNE    ) ::   zvalE
-  real*8,dimension(3,MAXEQNE,3  ) ::  zdvalE
-  real*8,dimension(3,MAXEQNE,3,3) :: zd2valE
-  real*8,dimension(3,MAXEQNV    ) ::   zvalV
-  real*8,dimension(3,MAXEQNV,3  ) ::  zdvalV
-  real*8,dimension(3,MAXEQNV,3,3) :: zd2valV
-  real*8,dimension(  MAXEQNQ    ) ::   zvalQ
-  real*8,dimension(  MAXEQNQ,3  ) ::  zdvalQ
-  real*8,dimension(  MAXEQNQ,3,3) :: zd2valQ
+  real(8),dimension(  MAXEQNH    ) ::   zvalH
+  real(8),dimension(  MAXEQNH,3  ) ::  zdvalH
+  real(8),dimension(  MAXEQNH,3,3) :: zd2valH
+  real(8),dimension(3,MAXEQNE    ) ::   zvalE
+  real(8),dimension(3,MAXEQNE,3  ) ::  zdvalE
+  real(8),dimension(3,MAXEQNE,3,3) :: zd2valE
+  real(8),dimension(3,MAXEQNV    ) ::   zvalV
+  real(8),dimension(3,MAXEQNV,3  ) ::  zdvalV
+  real(8),dimension(3,MAXEQNV,3,3) :: zd2valV
+  real(8),dimension(  MAXEQNQ    ) ::   zvalQ
+  real(8),dimension(  MAXEQNQ,3  ) ::  zdvalQ
[…]
