+++ trunk/src/meshmod/execute_pref.F90
@@ -1,45 +1,45 @@
-c---------------------------------------------------------------------
-c
-c   routine name       - execute_pref
-c
-c---------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine executes p-refinement for a group
-c                        of elements
-c                        REMARK: this routine executes ONLY isotropic
-c                                refinements
-c
-c   arguments :
-c
-c     in:
-c             List     - a list of elements (active or not)
-c             Nlist    - length of the list
-c
-c-----------------------------------------------------------------------
-c
+!---------------------------------------------------------------------
+!
+!   routine name       - execute_pref
+!
+!---------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - routine executes p-refinement for a group
+!                        of elements
+!                        REMARK: this routine executes ONLY isotropic
+!                                refinements
+!
+!   arguments :
+!
+!     in:
+!             List     - a list of elements (active or not)
+!             Nlist    - length of the list
+!
+!-----------------------------------------------------------------------
+!
       subroutine execute_pref(List,Nlist)
-c
+!
       use data_structure3D
       use element_data
       use constrained_nodes
-c
+!
       implicit none
-c
+!
       integer, intent(in) :: Nlist
       integer, intent(in) :: List(Nlist)
-c
-c  ...element nodes
+!
+!  ...element nodes
       integer :: nodesl(27), norientl(27)
-c
-c  ...order for element nodes implied by the order of the middle node
+!
+!  ...order for element nodes implied by the order of the middle node
       integer :: norder(19)
-c
-c  ...miscellanea
+!
+!  ...miscellanea
       integer :: icase,iel,ie,ip,is,j,mdle,nc,nrs
       integer :: nod,newp,nodp,nods,nrv,nre,nrf,ntype
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint=0
@@ -49,104 +49,104 @@ c
         call pause
       endif
 #endif
-c
+!
       call reset_visit
-c
-c  ...Step 1: loop over elements from the list and record
-c     the new order of approximation for their nodes
+!
+!  ...Step 1: loop over elements from the list and record
+!     the new order of approximation for their nodes
       do iel=1,Nlist
         mdle = List(iel)
-c
-c   ....skip if the element has been h-refined
+!
+!   ....skip if the element has been h-refined
         if (NODES(mdle)%ref_kind.ne.0) cycle
-c
-c  .....determine nodes for the element (active and constrained)
-c       and build the data base for the constrained nodes
-c       (module constrained_nodes)
+!
+!  .....determine nodes for the element (active and constrained)
+!       and build the data base for the constrained nodes
+!       (module constrained_nodes)
         call get_connect_info(mdle, nodesl,norientl)
         ntype = NODES(mdle)%ntype
         nrv = nvert(ntype); nre = nedge(ntype); nrf = nface(ntype)
-c
-c  .....loop through element higher order nodes
-C         do j=nrv+1,nrv+nre+nrf+1
+!
+!  .....loop through element higher order nodes
+!         do j=nrv+1,nrv+nre+nrf+1
 !     ...ONLY THROUGH FACES AND THE MIDDLE NODE
         do j=nrv+nre+1,nrv+nre+nrf+1
           nod = Nodesl(j)
-c
-c  .......if nod is active
+!
+!  .......if nod is active
           if (Is_active(nod)) then
-c
+!
             NODES(nod)%visit = 1
             nrs = NODES(nod)%nr_sons
             do is=1,nrs
               nods = Son(nod,is)
               NODES(nods)%visit = 1
[…]
