+++ trunk/src/meshmod/util/verify_orient.F90
@@ -1,60 +1,60 @@
 #if HP3D_DEBUG
 
-c----------------------------------------------------------------------
-c
-c   routine name       - verify_orient
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine verifies consistency of node
-c                        orientations as seen by elements
-c
-c   arguments          - none
-c
-c----------------------------------------------------------------------
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - verify_orient
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - routine verifies consistency of node
+!                        orientations as seen by elements
+!
+!   arguments          - none
+!
+!----------------------------------------------------------------------
+!
       subroutine verify_orient
-c
+!
       use data_structure3D
       use element_data
-c
+!
       implicit none
-c
-c  ...work space for elem_nodes
+!
+!  ...work space for elem_nodes
       integer :: nodesl(27),norientl(27)
-c
-c  ...vertex nodes on an edge or face
+!
+!  ...vertex nodes on an edge or face
       integer :: nv(4)
-c
-c  ...miscellanea
+!
+!  ...miscellanea
       integer :: ic,ie,iel,ifc,inod,iv,iaux,mdle
       integer :: nrv,nre,nrf,nod,nflag,nsum
-c
+!
       integer, allocatable ::  node_orient(:,:,:)
       integer ::  node_orient_aux(4)
-c     for a mid-edge or mid0face node 'nod' and i-th element
-c     connected to the node,
-c        node_orient(1,i,nod) = element (middle node) number
-c        node_orient(2:3(4,5),i,nod) = vertices on the edge
-c                               (face) listed in the order
-c                               dictated by the global orientation
-c                               (as seen by the element)
-c
+!     for a mid-edge or mid0face node 'nod' and i-th element
+!     connected to the node,
+!        node_orient(1,i,nod) = element (middle node) number
+!        node_orient(2:3(4,5),i,nod) = vertices on the edge
+!                               (face) listed in the order
+!                               dictated by the global orientation
+!                               (as seen by the element)
+!
       allocate(node_orient(5,10,NRNODS))
       node_orient = 0
-c
-c  ...loop through elements and record mid-edge node and mid-face
-c     node orientations as seen by the elements
+!
+!  ...loop through elements and record mid-edge node and mid-face
+!     node orientations as seen by the elements
       do iel=1,NRELES
         mdle = ELEM_ORDER(iel)
         nrv = nvert(NODES(mdle)%ntype)
         nre = nedge(NODES(mdle)%ntype)
         nrf = nface(NODES(mdle)%ntype)
         call elem_nodes(mdle, nodesl,norientl)
-c
-c  .....loop through the element edges
+!
+!  .....loop through the element edges
         do ie=1,nre
           inod = nrv+ie
           nod = nodesl(inod)
@@ -69,12 +69,12 @@ c  .....loop through the element edges
           node_orient(1,ic,nod) = mdle
           call edge_to_vert(Nodes(mdle)%ntype,ie, nv(1),nv(2))
           nv(1:2) = nodesl(nv(1:2))
-          call apply_orient(norientl(inod),NODES(nod)%ntype,
-     .                      nv, node_orient_aux)
+          call apply_orient(norientl(inod),NODES(nod)%ntype, &
[…]
