+++ trunk/src/meshmod/perform_pref.F90
@@ -1,53 +1,53 @@
-c---------------------------------------------------------------------
-c
-c   routine name       - perform_pref
-c
-c---------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine executes p-refinements for a group
-c                        of element middle nodes, along with the
-c                        corresponding p-refinements of edge and face
-c                        nodes
-c
-c   arguments :
-c
-c     in:
-c            List_elem - a list of active elements
-c            List_ord  - the corresponding orders for the middle
-c                        nodes
-c            Nlist     - length of both lists
-c
-c    REMARK: routine appears to call nodmod more often than necessary
-c            need to check correctness of the routine
-c
-c-----------------------------------------------------------------------
-c
+!---------------------------------------------------------------------
+!
+!   routine name       - perform_pref
+!
+!---------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - routine executes p-refinements for a group
+!                        of element middle nodes, along with the
+!                        corresponding p-refinements of edge and face
+!                        nodes
+!
+!   arguments :
+!
+!     in:
+!            List_elem - a list of active elements
+!            List_ord  - the corresponding orders for the middle
+!                        nodes
+!            Nlist     - length of both lists
+!
+!    REMARK: routine appears to call nodmod more often than necessary
+!            need to check correctness of the routine
+!
+!-----------------------------------------------------------------------
+!
       subroutine perform_pref(List_elem,List_nord,Nlist)
-c
+!
       use data_structure3D
       use element_data
       use constrained_nodes
-c
+!
       implicit none
-c
+!
       integer, intent(in) :: Nlist
       integer, intent(in) :: List_elem(Nlist), List_nord(Nlist)
-c
-c  ...element nodes
+!
+!  ...element nodes
       integer :: nodesl(27), norientl(27)
-c
-c  ...order for element nodes
+!
+!  ...order for element nodes
       integer :: norder(19)
-c
+!
       integer :: ie,iel,ip,j,icase,nc,mdle,ntype,nrv,nre,nrf
       integer :: nod,nodp,newp,newph,newpv,nord,nordh,nordv,nordm
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint=0
-c
+!
       if (iprint.eq.1) then
         write(*,7010) List_elem(1:Nlist)
  7010   format('perform_pref: List_elem = ',20(/,10i6))
@@ -56,115 +56,115 @@ c
         call pause
       endif
 #endif
-c
-c  ...Step 1: loop over elements from the list and perform the requested
-c     p-refinements
+!
+!  ...Step 1: loop over elements from the list and perform the requested
+!     p-refinements
       do iel=1,Nlist
         mdle = List_elem(iel)
         nord = List_nord(iel)
-c
-c   ....skip if not an active element
+!
+!   ....skip if not an active element
         if (NODES(mdle)%ref_kind.ne.0) cycle
-c
-c  .....enforce the new order for the element middle node
+!
+!  .....enforce the new order for the element middle node
         call nodmod(mdle,nord)
 
-c  .....determine nodes for the element (active and constrained)
-c       and build the data base for the constrained nodes
-c       (module constrained_nodes)
+!  .....determine nodes for the element (active and constrained)
+!       and build the data base for the constrained nodes
+!       (module constrained_nodes)
         call get_connect_info(mdle, nodesl,norientl)
-c
-c  .....determine the order for element nodes implied by the middle node
+!
[…]
