+++ trunk/src/datstrs/neig_edge.F90
@@ -1,82 +1,82 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - neig_edge
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - find neigbors for an element edge
-c
-c   arguments :
-c     in:
-c         Medge        - a mid-edge node
-c         Maxn         - dimension of the arrays below
-c                        (max number of neighbors)
-c     out:
-c         Neig         - list of neighbors of the edge
-c         Nedg_list    - local edge numbers
-c         Norient_list - orientations of the mid-edge node
-c         Nrneig       - number of adjacent elements
-c
-c----------------------------------------------------------------------
-c
-      subroutine neig_edge(Medge,Maxn,
-     .                     Neig,Nedg_list,Norient_list,Nrneig)
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - neig_edge
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - find neigbors for an element edge
+!
+!   arguments :
+!     in:
+!         Medge        - a mid-edge node
+!         Maxn         - dimension of the arrays below
+!                        (max number of neighbors)
+!     out:
+!         Neig         - list of neighbors of the edge
+!         Nedg_list    - local edge numbers
+!         Norient_list - orientations of the mid-edge node
+!         Nrneig       - number of adjacent elements
+!
+!----------------------------------------------------------------------
+!
+      subroutine neig_edge(Medge,Maxn, &
+                           Neig,Nedg_list,Norient_list,Nrneig)
+!
       use GMP
       use data_structure3D
       use element_data
-c
+!
       implicit none
-c
+!
       integer, intent(in)  :: Medge
       integer, intent(in)  :: Maxn
       integer, intent(out) :: Neig(2,Maxn)
       integer, intent(out) :: Nedg_list(Maxn)
       integer, intent(out) :: Norient_list(Maxn)
       integer, intent(out) :: Nrneig
-c
-c  ...neighbors of a curve
+!
+!  ...neighbors of a curve
       integer :: neigbl(100)
-c
-c  ...decoded orientations for edges
+!
+!  ...decoded orientations for edges
       integer :: nedge_orient(12)
-c
-c  ...element type
+!
+!  ...element type
       integer :: etype
-c
-c  ...miscellanea
+!
+!  ...miscellanea
       integer :: nod,nc,nrbl,nb,lab,ib,ie,nel,ibeg,iend
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint = 0
 #endif
-c
-c----------------------------------------------------------------------
-c
-c  ...initialize
+!
+!----------------------------------------------------------------------
+!
+!  ...initialize
       Neig = 0; Nedg_list = 0; Norient_list = 0; Nrneig = 0
-c
+!
       if (NODES(Medge)%ntype.ne.MEDG) then
         write(*,7001) Medge,S_Type(NODES(Medge)%ntype)
  7001   format('neig_edge: Medge,NODES(Medge)%type = ',i6,2x,a5)
         stop 1
       endif
-c
-c  ...a temporary version for the initial mesh only...
+!
+!  ...a temporary version for the initial mesh only...
       if (NODES(Medge)%father.gt.0) then
         write(*,7002) Medge,NODES(Medge)%father
[…]
