+++ trunk/src/constrs/util/hdiv/setcnstr_trian_iso_hdiv.F90
@@ -1,89 +1,89 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - setcnstr_trian_iso_hdiv
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine evaluates constraint coefficients for
-c                        the triangular master element and H(div) (L2)
-c                        constrained approximation
-c
-c   arguments          - none
-c
-c----------------------------------------------------------------------
-c
-c    'big' parent node (just one)
-c
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *       *
-c      *         *
-c      *           *
-c      *             *
-c      *       1       *
-c      *                 *
-c      *                   *
-c      *                     *
-c      *************************
-c
-c
-c      'small' constrained nodes
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *   3   *
-c      *         *
-c      *************
-c      * *         * *
-c      *   *   4   *   *
-c      *     *     *     *
-c      *   1   *   *   2   *
-c      *         * *         *
-c      *************************
-c
-c--------------------------------------------------------------------
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - setcnstr_trian_iso_hdiv
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - routine evaluates constraint coefficients for
+!                        the triangular master element and H(div) (L2)
+!                        constrained approximation
+!
+!   arguments          - none
+!
+!----------------------------------------------------------------------
+!
+!    'big' parent node (just one)
+!
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *       *
+!      *         *
+!      *           *
+!      *             *
+!      *       1       *
+!      *                 *
+!      *                   *
+!      *                     *
+!      *************************
+!
+!
+!      'small' constrained nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *   3   *
+!      *         *
+!      *************
+!      * *         * *
+!      *   *   4   *   *
+!      *     *     *     *
+!      *   1   *   *   2   *
+!      *         * *         *
+!      *************************
+!
+!--------------------------------------------------------------------
+!
       subroutine setcnstr_trian_iso_hdiv
-c
+!
       use parameters
       use constraints
       use element_data
-c
+!
       implicit none
-c
-c  ...big(parent) and small element shape functions, collocation points
-      real(8) :: shapsma(MAXtriaQ,MAXtriaQ),
-     .           shapbig(MAXtriaQ,MAXtriaQ),
-     .           xibig(2),xisma(2)
+!
+!  ...big(parent) and small element shape functions, collocation points
+      real(8) :: shapsma(MAXtriaQ,MAXtriaQ), &
+                 shapbig(MAXtriaQ,MAXtriaQ), &
+                 xibig(2),xisma(2)
       integer :: ip(MAXtriaQ)
-c
+!
       integer :: norder(4),nsize(2)
-c
+!
[…]
