+++ trunk/src/constrs/util/hcurl/setcnstr_trian_iso_hcurl.F90
@@ -1,158 +1,158 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - setcnstr_trian_iso_hcurl
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine evaluates constraint coefficients for
-c                        the triangular master element
-c
-c   arguments          - none
-c
-c----------------------------------------------------------------------
-c
-c    'big' parent nodes
-c
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *       *
-c      *         *
-c    4 *           *  3
-c      *             *
-c      *       1       *
-c      *                 *
-c      *                   *
-c      *                     *
-c      *************************
-c                  2
-c
-c      'small' constrained nodes
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *   3   *
-c      *         *
-c      ******7******
-c      * *         * *
-c      *   *   4   *   *
-c      *     5     6     *
-c      *   1   *   *   2   *
-c      *         * *         *
-c      *************************
-c
-c--------------------------------------------------------------------
-c
-      subroutine setcnstr_trian_iso_hcurl
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - setcnstr_trian_iso_hcurl
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2024
+!
+!   purpose            - routine evaluates constraint coefficients for
+!                        the triangular master element
+!
+!   arguments          - none
+!
+!----------------------------------------------------------------------
+!
+!    'big' parent nodes
+!
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *       *
+!      *         *
+!    4 *           *  3
+!      *             *
+!      *       1       *
+!      *                 *
+!      *                   *
+!      *                     *
+!      *************************
+!                  2
+!
+!      'small' constrained nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *   3   *
+!      *         *
+!      ******7******
+!      * *         * *
+!      *   *   4   *   *
+!      *     5     6     *
+!      *   1   *   *   2   *
+!      *         * *         *
+!      *************************
+!
+!--------------------------------------------------------------------
+!
+   subroutine setcnstr_trian_iso_hcurl
+!
       use parameters
       use constraints
       use element_data
-c
+!
       implicit none
-c
-c  ...use collocation points for lagrange shape functions h1
+!
+!  ...use collocation points for lagrange shape functions h1
       integer, parameter :: npts  = 4*MAXtriaE
       integer, parameter :: lwork = MAXtriaE**2
-c
-      real(8) :: shapsma(2,MAXtriaE),curl(MAXtriaE),
-     .           shapbig(2,MAXtriaE),
-     .           evalsma(npts,MAXtriaE),evalbig(npts,MAXtriaE),
[…]
