+++ trunk/src/constrs/util/hcurl/setcnstr_edge_hcurl.F
@@ -1,270 +1,0 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - setcnstr_edge_hcurl
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine sets coefficients for the constrained
-c                        approximation on the master edge
-c
-c   arguments          - none
-c
-c----------------------------------------------------------------------
-c
-c     parent nodes:
-c
-c     |----------------->------------------|
-c    big 2           big 1                big 3
-c
-c
-c     constrained nodes:
-c
-c     |--------->-------|--------->--------|
-c             sm.1     sm.3     sm.2
-c
-c
-c---------------------------------------------------------------------
-c
-      subroutine setcnstr_edge_hcurl
-c
-      use parameters
-      use constraints
-c
-      implicit none
-c
-c  ...values of small elements shape functions, big element
-c     shape functions, derivatives (not used)
-      real(8) :: shapsma(MAXP,MAXP),shapbig(MAXP,MAXP)
-c
-c  ...pivoting array, work array
-      integer :: ip(MAXP)
-      real(8) :: r(MAXP,MAXP)
-c
-      real(8) :: aux,xibig,xisma
-      integer :: i,j,icl,is,info
-      integer :: nord,nvoid
-c
-#if HP3D_DEBUG
-      real(8) :: val
-      integer :: ians,nel
-      integer :: iprint
-      iprint = 0
-#endif
-c
-c  ...clear:
-      RRRE = 0.d0
-c
-      nord = MAXP
-c
-#if HP3D_DEBUG
-      if (iprint.eq.1) then
-        write(*,*) 'Compute constraints wrt p = ', nord
-      end if
-#endif
-c
-c---------------------------------------------------------------------
-c
-c  ...small nodes 1:
-c
-c---------------------------------------------------------------------
-c
-      icl = 0
-c
-c  ...loop through collocation points:
-      do i=1,nord
-        icl = icl+1
-c
-c  .....local coordinates for big and small elements
-        xibig = (i-1)*0.5d0/(nord-1)
-        xisma = (i-1)*1.0d0/(nord-1)
-c
-c  .....find small shape functions at this point:
-        call TraceEshapeE(xisma,nord,0, nvoid,shapsma(1:MAXP,icl))
-c
-c  .....find big shape functions at this point:
-        call TraceEshapeE(xibig,nord,0, nvoid,shapbig(1:MAXP,icl))
-      enddo
-c
-c  ...transpose shapsma and shapbig
-      do i=1,nord
-        do j=1,i
-c
-          aux = shapsma(i,j)
-          shapsma(i,j) = shapsma(j,i)
-          shapsma(j,i) = aux
-c
-          aux = shapbig(i,j)
-          shapbig(i,j) = shapbig(j,i)
-          shapbig(j,i) = aux
-        enddo
-      enddo
-c
-c  ...solve the collocation system to determine the constrained
-c     approximation coefficients
-c
-c  ...invert the collocation matrix
-#if HP3D_DEBUG
-      if (iprint.eq.1) then
-        write(*,7010)
- 7010   format('setcnstr_edge_hcurl: COLLOCATION MATRIX = ')
-        do i=1,nord
-          write(*,7011) i,shapsma(i,1:nord)
- 7011     format('i = ',i3,10e12.5)
-        enddo
-        call pause
-      endif
-#endif
-      call dgesv(nord,nord,shapsma,MAXP,ip,shapbig,MAXP,info)
[…]
