+++ trunk/src/constrs/util/h1/setcnstr_trian_aniso_h1.F
@@ -1,277 +1,0 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - setcnstr_trian_aniso_h1
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine evaluates constraint coefficients for
-c                        the triangular master element
-c
-c----------------------------------------------------------------------
-c
-c     'big' parent nodes
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *       *
-c      *         *
-c    4 *           *  3
-c      *             *
-c      *       1       *
-c      *                 *
-c      *                   *
-c      *                     *
-c      *************************
-c                  2
-c
-c      Iref = 2; 'small' constrained nodes
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *       *
-c      *         *
-c      *           *
-c      * *      1    *
-c      *   *           *
-c      *     ?           *
-c      *   ?   *           *
-c      *         *           *
-c      *************************
-c
-c      Iref = 3; 'small' constrained nodes
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *       *
-c      *         *
-c      *         * *
-c      *    1    *   *
-c      *         *     *
-c      *         ?   ?   *
-c      *         *         *
-c      *         *           *
-c      *************************
-c
-c      Iref = 4; 'small' constrained nodes
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *   ?   *
-c      *         *
-c      *           *
-c      * * * ? * * * *
-c      *               *
-c      *        1        *
-c      *                   *
-c      *                     *
-c      *************************
-c
-c--------------------------------------------------------------------
-c
-      subroutine setcnstr_trian_aniso_h1(Iref)
-c
-      use parameters
-      use constraints
-c
-      implicit none
-c
-      integer, intent(in) :: Iref
-c
-      real(8) :: shapsma(MAXquadH,MAXquadH),
-     .           shapbig(MAXquadH,MAXquadH),
-     .           void(2,MAXquadH),
-     .           r(MAXquadH,MAXquadH),xibig(2),xisma(2)
-      integer :: ip(MAXquadH)
-c
-c  ...order and edge orientations
-      integer :: norder(5),norient(4)
-c
-      real(8) :: aux
-      integer :: i,ie,i1,i2,i1beg,ipp,jp,jpbeg,jpend,j,icl
-      integer :: n,na,nord,nrb,nrdof,nrs,info
-c
-#if HP3D_DEBUG
-      integer :: iprint
-      iprint = 0
-#endif
-c
-      nord = MAXP
-c
-c  ...set uniform order MAXP:
-      norder(1:4) = nord
-      norder(5) = nord*10 + nord
-c
-c  ...set orientations
-      norient = 0
-c
-c*********************************************************************
-c
-      icl = 0
-c
-c  ...loop through collocation points:
-      do j = 1,nord+1
-        do i = 1,nord+1
-          xisma(1) = (i-1)*1.0d0/nord
-          xisma(2) = (j-1)*1.0d0/nord
-          call map_quad(Iref,xisma(1:2), xibig(1:2))
-c
-          icl = icl + 1
-c
[…]
