+++ trunk/src/constrs/util/h1/setcnstr_trian_iso_h1.F90
@@ -1,136 +1,136 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - setcnstr_trian_iso_h1
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine evaluates constraint coefficients for
-c                        the triangular master element
-c
-c   arguments          - none
-c
-c----------------------------------------------------------------------
-c
-c    'big' parent nodes
-c
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *       *
-c      *         *
-c    4 *           *  3
-c      *             *
-c      *       1       *
-c      *                 *
-c      *                   *
-c      *                     *
-c      *************************
-c                  2
-c
-c      'small' constrained nodes
-c      *
-c      * *
-c      *   *
-c      *     *
-c      *   3   *
-c      *         *
-c      ******7******
-c      * *         * *
-c      *   *   4   *   *
-c      *     5     6     *
-c      *   1   *   *   2   *
-c      *         * *         *
-c      *************************
-c
-c--------------------------------------------------------------------
-c
-      subroutine setcnstr_trian_iso_h1
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - setcnstr_trian_iso_h1
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2024
+!
+!   purpose            - routine evaluates constraint coefficients for
+!                        the triangular master element
+!
+!   arguments          - none
+!
+!----------------------------------------------------------------------
+!
+!    'big' parent nodes
+!
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *       *
+!      *         *
+!    4 *           *  3
+!      *             *
+!      *       1       *
+!      *                 *
+!      *                   *
+!      *                     *
+!      *************************
+!                  2
+!
+!      'small' constrained nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *   3   *
+!      *         *
+!      ******7******
+!      * *         * *
+!      *   *   4   *   *
+!      *     5     6     *
+!      *   1   *   *   2   *
+!      *         * *         *
+!      *************************
+!
+!--------------------------------------------------------------------
+!
+   subroutine setcnstr_trian_iso_h1
+!
       use parameters
       use constraints
-c
+!
       implicit none
-c
-      real(8) :: shapsma(MAXtriaH,MAXtriaH),
-     .           shapbig(MAXtriaH,MAXtriaH),
-     .           void(2,MAXtriaH),
-     .           r(MAXtriaH,MAXtriaH),xibig(2),xisma(2)
+!
+      real(8) :: shapsma(MAXtriaH,MAXtriaH), &
+                 shapbig(MAXtriaH,MAXtriaH), &
+                 void(2,MAXtriaH),           &
+                 r(MAXtriaH,MAXtriaH),xibig(2),xisma(2)
[…]

[…]
