+++ trunk/src/constrs/logic_nodes.F
@@ -1,349 +1,0 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - logic_nodes
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine establishes list of nodes
-c                        for a modified element
-c
-c   arguments :
-c     in:
-c            Mdle      - an element number, same as the middle node
-c                        number
-c            Nodesl    - (local) element nodes as returned by
-c                        get_connect_info
-c     out:
-c            Nodm      - actual (unconstrained) nodes returned in the
-c                        standard order: vertex, mid-edge, mid-face
-c                        and middle nodes
-c            Nrnodm    - number of nodes for the modified element
-c
-c----------------------------------------------------------------------
-c
-      subroutine logic_nodes(Mdle,Nodesl, Nodm,Nrnodm)
-c
-      use element_data
-      use data_structure3D
-      use constrained_nodes
-c
-      implicit none
-c
-      integer, intent(in)  :: Mdle
-      integer, intent(in)  :: Nodesl(27)
-      integer, intent(out) :: Nodm(MAXNODM)
-      integer, intent(out) :: Nrnodm
-c
-c  ...element type
-      integer :: ntype
-c
-c  ...local lists of nodes
-      integer :: list_v(16),list_e(24),list_f(12)
-c
-c  ...miscellanea
-      integer :: ie,ip,iv,jv,icv,ice,icf,icase,j
-      integer :: nvoid,nrnodl,nod,nodp,nc,nce
-c
-#if HP3D_DEBUG
-      integer :: iprint
-      iprint=0
-#endif
-c
-c----------------------------------------------------------------------
-c
-c  ...initialize output
-      Nodm(1:MAXNODM) = 0
-c
-      ntype = NODES(Mdle)%ntype
-c
-c  ...number of (local) nodes for the element (- middle node)
-      nrnodl = Nvert(ntype)+Nedge(ntype)+Nface(ntype)
-c
-c  ...establish lists of modified element vertex, edge, and face
-c     nodes
-      icv=0; ice=0; icf=0
-      do j=1,nrnodl
-        nod = Nodesl(j)
-c
-c  .....if nod is active
-        if ((Is_active(nod)).or.
-     .      (Is_inactive(nod) .and. NODES(nod)%ref_kind.ne.0)) then
-c
-          if (j.le.Nvert(ntype)) then
-            call add_to_list(list_v,16,icv,nod)
-          elseif (j.le.Nvert(ntype)+Nedge(ntype)) then
-            call add_to_list(list_e,24,ice,nod)
-          else
-            call add_to_list(list_f,12,icf,nod)
-          endif
-c
-c  .....inactive, i.e. constrained node
-        else
-c
-c  .......identify the constraint case
-          call decode2(NODES_CONSTR(j), nc,icase)
-c
-          select case(icase)
-c
-c  .......first and second mid-edge node constrained by an edge.....
-          case(11,12, 37,38, 47,48)
-c
-c  .........parent mid-edge node
-            nodp = NEDGC(nc)
-            call add_to_list(list_e,24,ice,nodp)
-c
-c  .......vertex node constrained by an edge.....
-          case(13,39,49)
-c
-c  .........parent mid-edge node
-            nodp = NEDGC(nc)
-            call add_to_list(list_e,24,ice,nodp)
-c
-c  .........loop through parent vertices
-            do iv=1,2
-              nodp = NEDG_CONS(iv,nc)
-c
-c  ...........active (unconstrained) vertex node
-              if (nodp.gt.0) then
[…]
