### 2024-02-27 7a7fa9d
update numeric legacy code
