+++ trunk/src/numeric/blas.F90
@@ -1,21 +1,21 @@
-c-----------------------------------------------------------------------
-c> @date Feb 2023
-c-----------------------------------------------------------------------
-      subroutine decomp(n,na,a,ip,Iflag)
+!-----------------------------------------------------------------------
+!> @date Feb 2024
+!-----------------------------------------------------------------------
+   subroutine decomp(n,na,a,ip,Iflag)
       implicit none
-c
+!
       integer :: n,na
       real(8) :: a(na,n)
       integer :: ip(n)
       integer :: Iflag
-c
+!
       integer :: j,k,imax,inc
-      real(8) :: c,pivmin
-c
+      real(8) :: cv,pivmin
+!
       real(8), parameter :: eps = 1.0d-15
-c
+!
       integer, external :: idamax
-c
+!
 #if HP3D_DEBUG
       integer :: i
       integer :: iprint
@@ -30,12 +30,12 @@ c
         call pause
       endif
 #endif
-c
+!
       pivmin = 1.d30
-c
+!
       Iflag=0
       inc=1
-c
+!
       do k=1,n
         imax = idamax(n-k+1,a(k,k),inc)
         ip(k) = k+imax-1
@@ -48,14 +48,14 @@ c
         if (dabs(a(ip(k),k)) > eps) go to 20
         Iflag=1
         return
-c
+!
  20     continue
-        c = a(ip(k),k)
+        cv = a(ip(k),k)
         a(ip(k),k) = a(k,k)
-        a(k,k) = c
-        c=1.d0/c
-c
-        call dscal(n-k,c,a(k+1,k),inc)
+        a(k,k) = cv
+        cv=1.d0/cv
+!
+        call dscal(n-k,cv,a(k+1,k),inc)
 #if HP3D_DEBUG
         if (iprint.eq.1) then
           write(*,7003) k
@@ -66,37 +66,37 @@ c
           call pause
         endif
 #endif
-c
+!
         do j=k+1,n
-          c = a(ip(k),j)
+          cv = a(ip(k),j)
           a(ip(k),j) = a(k,j)
-          a(k,j) = c
+          a(k,j) = cv
           call daxpy(n-k ,-a(k,j),a(k+1,k),inc,a(k+1,j),inc)
         enddo
-c
+!
       enddo
-c
-      end subroutine decomp
-c
-c-----------------------------------------------------------------------
-c> @date Feb 2023
-c-----------------------------------------------------------------------
-      subroutine gauss2(n,na,a,ip,b,x)
-c
+!
+   end subroutine decomp
+!
+!-----------------------------------------------------------------------
+!> @date Feb 2024
+!-----------------------------------------------------------------------
+   subroutine gauss2(n,na,a,ip,b,x)
+!
       implicit none
-c
+!
       integer :: n,na
       real(8) :: a(na,n),b(n),x(n)
       integer :: ip(n)
-c
+!
       integer :: inc,j,k
-      real(8) :: c
-c
+      real(8) :: cv
+!
 #if HP3D_DEBUG
       integer :: i
       integer :: iprint
       iprint = 0
-c
+!
       if (iprint.eq.1) then
         write(*,7001)
  7001   format('gauss2: n,na = ',2i4)
@@ -113,13 +113,13 @@ c
         call pause
       endif
 #endif
-c
+!
       inc=1
-c
+!
       do k=1,n-1
-        c=b(ip(k))
+        cv=b(ip(k))
         b(ip(k))=b(k)
-        b(k)=c
+        b(k)=cv
         call daxpy(n-k,-b(k),a(k+1,k),inc,b(k+1),inc)
       enddo
 
@@ -127,20 +127,20 @@ c
         x(j)=b(j)/a(j,j)
[…]

[…]
