### 2024-02-27 79dfcbf
updating legacy code in element/util
