+++ trunk/src/element/util/invmap.F
@@ -1,185 +1,0 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - invmap
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - given physical coordinates Xp of a point in
-c                        space and geometry dof for an element,
-c                        routine determines whether the point lies within
-c                        the element and returns its master element
-c                        coordinates
-c
-c   arguments :
-c     in:
-c             Ntype    - element type
-c             Xp       - physical coordinates of a point
-c             Norder   - element order
-c             Nedge_orient, Nface_orient - edge and face orientations
-c             Xnod     - element geometry dof
-c     out:
-c             Idec     = 1 if the point is within the element
-c                        0 otherwise or NR iterations have not converged
-c             Xi       - master element coordinates of Xp
-c
-c----------------------------------------------------------------------
-c
-      subroutine invmap(Ntype,Xp,Norder,Nedge_orient,Nface_orient,
-     .                  Xnod, Idec,Xi)
-c
-      use data_structure3D
-      use element_data
-      use control
-c
-      implicit none
-c
-      integer, intent(in)  :: Ntype,Norder(19)
-      integer, intent(in)  :: Nedge_orient(12),Nface_orient(6)
-      real(8), intent(in)  :: Xnod(3,MAXbrickH)
-      integer, intent(out) :: Idec
-      real(8), intent(out) :: Xi(3)
-c
-      real(8) :: Xp(3)
-c
-c  ...shape functions and their derivatives wrt master coordinates
-      real(8) :: shapH(MAXbrickH),gradH(3,MAXbrickH)
-      integer :: nrdofH
-c
-c  ...geometry
-      real(8) :: dxdxi(3,3),dxidx(3,3)
-c
-c  ...increment in xi
-      real(8) :: dxi(3)
-c
-      real(8) :: x(3),rjac,dmax
-      integer :: i,iter,k
-      integer :: iflag
-c
-#if HP3D_DEBUG
-      integer :: iprint
-#endif
-c
-      real(8), parameter :: eps = 1.d-8
-c
-c-----------------------------------------------------------------------
-c
-#if HP3D_DEBUG
-      iprint=0
-c
-c  ...initiate the NR iterations
-      if (iprint.eq.1) then
-        write(*,*)'***************************'
-        write(*,*)'invmap: NEWTON-RAPHSON LOOP'
-      endif
-#endif
-c
-c  ...use element centroid as initial guess
-      Xi(1:3)=0.d0
-      do i=1,nvert(Ntype)
-        select case(Ntype)
-        case(MDLB,BRIC) ; Xi(1:3) = Xi(1:3) + BRICK_COORD(1:3,i)
-        case(MDLN,TETR) ; Xi(1:3) = Xi(1:3) + TETRA_COORD(1:3,i)
-        case(MDLP,PRIS) ; Xi(1:3) = Xi(1:3) + PRISM_COORD(1:3,i)
-        case(MDLD,PYRA) ; Xi(1:3) = Xi(1:3) + PYRAM_COORD(1:3,i)
-        endselect
-      enddo
[…]
