+++ trunk/src/element/util/geom3D.F
@@ -1,397 +1,0 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - geom3D
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - element geometry routine
-c
-c   arguments :
-c     in:
-c             Mdle     - element middle node
-c             Xi       - master element coordinates of a point
-c             Xnod     - element geometry dof
-c             ShapH    - values of shape functions at a point
-c             GradH    - derivatives of the shape functions wrt
-c                        master coordinates at the point
-c             NrdofH   - number of element H1 dof
-c     out:
-c             X        - physical coordinates of the point
-c             Dxdxi    - Jacobian matrix
-c             Dxidx    - inverse Jacobian matrix
-c             Rjac     - Jacobian (determinant of the Jacobian matrix)
-c             Iflag    = 0 OK
-c                        1 negative Jacobian
-c
-c----------------------------------------------------------------------
-c
-      subroutine geom3D(Mdle,Xi,Xnod,ShapH,GradH,NrdofH,
-     .                  X,Dxdxi,Dxidx,Rjac,Iflag)
-c
-      use control
-      use data_structure3D
-      implicit none
-c
-      integer, intent(in)  :: Mdle,NrdofH
-      real(8), intent(out) :: Rjac
-      integer, intent(out) :: Iflag
-c
-      real(8) :: Xi(3),Xnod(3,MAXbrickH),
-     .           ShapH(MAXbrickH),GradH(3,MAXbrickH),
-     .           X(3),Dxdxi(3,3),Dxidx(3,3)
-c
-      integer :: i,k
-c
-#if HP3D_DEBUG
-      integer :: iprint
-      iprint=0
-      if (iprint.eq.1) then
-        write(*,7001) Mdle,Xi(1:3),EXGEOM
- 7001   format('geom3D: Mdle,Xi = ',i6,3x,3f8.3,' EXGEOM = ',i3)
-      endif
-#endif
-c
-c  ...determine physical coordinates and the derivatives of
-c     the physical coordinates wrt master element coordinates
-      select case(EXGEOM)
-c
-c  ...isoparametric element
-      case(0)
-#if HP3D_DEBUG
-        if (iprint.eq.1) then
-          write(*,*) 'geom3D: Xnod  = '
-          do i=1,3
-            write(*,7010) Xnod(i,1:NrdofH)
- 7010       format(10(27f7.2,1x,/))
-          enddo
-          write(*,*) '        ShapH = '
-          write(*,7010) ShapH(1:NrdofH)
-          write(*,*) '        GradH = '
-          do i=1,3
-            write(*,7010) GradH(i,1:NrdofH)
-          enddo
-        endif
-#endif
-        X(1:3)=0.d0 ; Dxdxi(1:3,1:3)=0.d0
-        do k=1,NrdofH
-          X(1:3) = X(1:3) + Xnod(1:3,k)*ShapH(k)
-          do i=1,3
-            Dxdxi(1:3,i) = Dxdxi(1:3,i) + Xnod(1:3,k)*GradH(i,k)
-          enddo
-        enddo
-c
-c  ...exact geometry element
-      case(1)
-        call exact_geom(Mdle,Xi, X,Dxdxi)
-      end select
-c
-c  ...evaluate the inverse Jacobian of the geometry map
-      call geom(Dxdxi, Dxidx,Rjac,Iflag)
-c
-      if (Iflag .ne. 0) then
-        !$OMP CRITICAL
-        write(*,*) 'geom3D: negative Jacobian!'
-        write(*,*) 'Mdle  = ', Mdle
-        write(*,*) 'Rjac  = ', Rjac
[…]

[…]
