### 2024-02-27 4db48fd
replacing debug and c_mode flags in mgcg
