+++ trunk/src/solver/MGCG/direct_solve/celem_mg.F90
@@ -108,7 +108,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
    integer :: nrdoflH,nrdoflE,nrdoflV,nrdoflQ
    integer :: nrdoflHi,nrdoflEi,nrdoflVi
 !
-#if DEBUG_MODE
+#if HP3D_DEBUG
    integer :: ians,ibeg,iend,jbeg,jend,kbeg,kend
    integer :: iprint=0
 #endif
@@ -135,7 +135,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
       end select
       call celndof(type,norder, nrdoflH,nrdoflE,nrdoflV,nrdoflQ)
       call celndof(type,norderi, nrdoflHi,nrdoflEi,nrdoflVi,nvoid)
-#if DEBUG_MODE
+#if HP3D_DEBUG
       if (iprint.eq.1) then
         write(*,5001) Mdle
  5001   format(' celem_mg: DEBUGGING for Mdle = ',i8)
@@ -172,7 +172,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
       Nrnodm = Nrnodm-1
 !
 !  ...printing
-#if DEBUG_MODE
+#if HP3D_DEBUG
       if ((iprint.ge.1).and.(Idec.ne.1)) then
         write(*,*) ' - H1 - '
         do k=1,nrdoflH
@@ -236,7 +236,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
       nrdofmHEV = nrdofmHE  + nrdofmV
       Nrdofm    = nrdofmHEV
 !
-#if DEBUG_MODE
+#if HP3D_DEBUG
       if (iprint .eq. 1) then
         write(*,6001) nrdoflH,nrdoflE,nrdoflV,nrdoflQ
  6001   format('celem: nrdoflH,nrdoflE,nrdoflV,nrdoflQ = ',4i4)
@@ -266,7 +266,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
          case default
       end select
    enddo
-#if DEBUG_MODE
+#if HP3D_DEBUG
    if (iprint .eq. 1) then
       write(*,*) 'Physics variables:'
       write(*,*) 'nrPhysH = ', nrPhysH
@@ -326,7 +326,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
                case(2)
                   k=k+1
                   NEXTRACT(k) = l
-#if DEBUG_MODE
+#if HP3D_DEBUG
                case default
                   write(*,*) 'celem_mg: nod,ii,index(ii) = ',nod,ii,index(ii)
                   write(*,*) '                    in H1: Mdle = ', Mdle, '. stop.'
@@ -392,7 +392,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
                case(4)
                   k=k+1
                   NEXTRACT(k) = l
-#if DEBUG_MODE
+#if HP3D_DEBUG
                case default
                   write(*,*) 'celem_mg: nod,ii,index(ii) = ',nod,ii,index(ii)
                   write(*,*) '               in H(curl): Mdle = ', Mdle, '. stop.'
@@ -458,7 +458,7 @@ subroutine celem_mg(Ielc,Iel,Mdle,Idec,                       &
                case(6)
                   k=k+1
                   NEXTRACT(k) = l
-#if DEBUG_MODE
+#if HP3D_DEBUG
                case default
                   write(*,*) 'celem_mg: nod,ii,index(ii) = ',nod,ii,index(ii)
[…]
