### 2024-02-27 00213fc
Renaming `DEBUG_MODE` and `C_MODE` preprocessor flags (#172)
* renaming DEBUG and COMPLEX preprocessor flags

* Adding CHANGELOG

* updating DEBUG and COMPLEX flags everywhere

* define variables for MPI launch

* minor fixes
