+++ trunk/src/numeric/blas.F
@@ -16,7 +16,7 @@ c
 c
       integer, external :: idamax
 c
-#if DEBUG_MODE
+#if HP3D_DEBUG
       integer :: i
       integer :: iprint
       iprint=0
@@ -56,7 +56,7 @@ c
         c=1.d0/c
 c
         call dscal(n-k,c,a(k+1,k),inc)
-#if DEBUG_MODE
+#if HP3D_DEBUG
         if (iprint.eq.1) then
           write(*,7003) k
  7003     format('decomp: k = ',i4)
@@ -92,7 +92,7 @@ c
       integer :: inc,j,k
       real(8) :: c
 c
-#if DEBUG_MODE
+#if HP3D_DEBUG
       integer :: i
       integer :: iprint
       iprint = 0
