### 2024-03-01 d6d9e4c
Update legacy code in constrained approximation routines (#174)
* update constrs/util/hdiv legacy

* update constrs/util/hcurl legacy

* update constrs/util/h1 legacy

* update constrs legacy code
