+++ trunk/src/constrs/util/h1/setcnstr_trian_iso_h1.F90
@@ -1,0 +1,823 @@
+!----------------------------------------------------------------------
+!
+!   routine name       - setcnstr_trian_iso_h1
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2024
+!
+!   purpose            - routine evaluates constraint coefficients for
+!                        the triangular master element
+!
+!   arguments          - none
+!
+!----------------------------------------------------------------------
+!
+!    'big' parent nodes
+!
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *       *
+!      *         *
+!    4 *           *  3
+!      *             *
+!      *       1       *
+!      *                 *
+!      *                   *
+!      *                     *
+!      *************************
+!                  2
+!
+!      'small' constrained nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *   3   *
+!      *         *
+!      ******7******
+!      * *         * *
+!      *   *   4   *   *
+!      *     5     6     *
+!      *   1   *   *   2   *
+!      *         * *         *
+!      *************************
+!
+!--------------------------------------------------------------------
+!
+   subroutine setcnstr_trian_iso_h1
+!
+      use parameters
+      use constraints
+!
+      implicit none
+!
+      real(8) :: shapsma(MAXtriaH,MAXtriaH), &
+                 shapbig(MAXtriaH,MAXtriaH), &
+                 void(2,MAXtriaH),           &
+                 r(MAXtriaH,MAXtriaH),xibig(2),xisma(2)
+      integer :: ip(MAXtriaH)
+!
+!  ...order and edge orientations
+      integer :: norder(4),norient(3),nsize(2)
+!
+      real(8) :: aux
+      integer :: i,ipp,j,jbeg,jp,jpbeg,jpend,icl,info
+      integer :: n,nrb,nord,nrdof,nrs
+!
+#if HP3D_DEBUG
+      real(8) :: val
+      integer :: ians,k,nel
+      integer :: iprint
+      iprint = 0
+#endif
+!
+      nord = MAXP
+!
+!  ...set uniform order MAXP:
+      norder  = nord
+      norient = 0
+      nsize   = (/MAXP,MAXtriaH/)
+!
+!  ...initiate the matrix of coefficients
+!!!!      RRTH = 0.d0
+!
+!*********************************************************************
+!
+!  ...small triangle 1 -- small nodes 1 and 5:
+!
+!*********************************************************************
+!
+      icl = 0
+!
+!  ...loop through collocation points:
+      do i=1,nord+1
+!
+        xibig(1) = (i-1)*0.5d0/nord
+        xisma(1) = (i-1)*1.0d0/nord
+!
+        do j=1,nord+1 -(i-1)
+          icl = icl+1
+!
+          xibig(2) = 0.5d0/nord*(j-1)
+          xisma(2) = 1.0d0/nord*(j-1)
+!
+!  .......find small shape functions at this point:
+!          call shapeHt(xisma,norder,norient, &
+!                        nrdof,shapsma(1:MAXtriaH,icl),void)
[…]
