+++ trunk/src/constrs/util/h1/setcnstr_trian_aniso_h1.F90
@@ -1,0 +1,273 @@
+!----------------------------------------------------------------------
+!
+!   routine name       - setcnstr_trian_aniso_h1
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2024
+!
+!   purpose            - routine evaluates constraint coefficients for
+!                        the triangular master element
+!
+!----------------------------------------------------------------------
+!
+!     'big' parent nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *       *
+!      *         *
+!    4 *           *  3
+!      *             *
+!      *       1       *
+!      *                 *
+!      *                   *
+!      *                     *
+!      *************************
+!                  2
+!
+!      Iref = 2; 'small' constrained nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *       *
+!      *         *
+!      *           *
+!      * *      1    *
+!      *   *           *
+!      *     ?           *
+!      *   ?   *           *
+!      *         *           *
+!      *************************
+!
+!      Iref = 3; 'small' constrained nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *       *
+!      *         *
+!      *         * *
+!      *    1    *   *
+!      *         *     *
+!      *         ?   ?   *
+!      *         *         *
+!      *         *           *
+!      *************************
+!
+!      Iref = 4; 'small' constrained nodes
+!      *
+!      * *
+!      *   *
+!      *     *
+!      *   ?   *
+!      *         *
+!      *           *
+!      * * * ? * * * *
+!      *               *
+!      *        1        *
+!      *                   *
+!      *                     *
+!      *************************
+!
+!--------------------------------------------------------------------
+!
+      subroutine setcnstr_trian_aniso_h1(Iref)
+!
+      use parameters
+      use constraints
+!
+      implicit none
+!
+      integer, intent(in) :: Iref
+!
+      real(8) :: shapsma(MAXquadH,MAXquadH), &
+                 shapbig(MAXquadH,MAXquadH), &
+                 void(2,MAXquadH),           &
+                 r(MAXquadH,MAXquadH),xibig(2),xisma(2)
+      integer :: ip(MAXquadH)
+!
+!  ...order and edge orientations
+      integer :: norder(5),norient(4)
+!
+      real(8) :: aux
+      integer :: i,ie,i1,i2,i1beg,ipp,jp,jpbeg,jpend,j,icl
+      integer :: n,na,nord,nrb,nrdof,nrs,info
+!
+#if HP3D_DEBUG
+      integer :: iprint
+      iprint = 0
+#endif
+!
+      nord = MAXP
+!
+!  ...set uniform order MAXP:
+      norder(1:4) = nord
+      norder(5) = nord*10 + nord
+!
+!  ...set orientations
+      norient = 0
+!
+!*********************************************************************
+!
+      icl = 0
+!
+!  ...loop through collocation points:
+      do j = 1,nord+1
+        do i = 1,nord+1
+          xisma(1) = (i-1)*1.0d0/nord
+          xisma(2) = (j-1)*1.0d0/nord
+          call map_quad(Iref,xisma(1:2), xibig(1:2))
+!
[…]

[…]
