### 2024-03-01 3a8529b
Update legacy code in numeric routines  (#176)
