### 2024-03-01 7eaae3c
update legacy code in gauss_quadrature
