+++ trunk/src/element/quadrature/gauss_quadrature.F
@@ -1,851 +1,0 @@
-      module gauss_quadrature
-
-      implicit none
-c
-      logical :: INITIALIZED = .FALSE.
-c$OMP THREADPRIVATE (INITIALIZED)
-c
-c======================================================================
-c  Constants for normalized integrated Legendre polynomials           |
-c======================================================================
-c
-      real(8), dimension(9,9) :: XLEG0
-c$OMP THREADPRIVATE (XLEG0)
-c
-c
-c======================================================================
-c  Integration points and corresponding weights for Gaussian          |
-c  integration on interal [-1,1].                                     |
-c  Remark: they are rescaled to interval [0,1] by routine             |
-c  init_gauss_quadrature                                              |
-c======================================================================
-c
-      real(8), dimension(10,10) :: XIGAUS1
-c$OMP THREADPRIVATE (XIGAUS1)
-      real(8), dimension(10,10) :: WAGAUS1
-c$OMP THREADPRIVATE (WAGAUS1)
-c
-c
-c======================================================================
-c  Gauss-Jacobi points and weights                                    |
-c======================================================================
-c
-      real(8), dimension(10,10) :: XIGAUSJ
-c$OMP THREADPRIVATE (XIGAUSJ)
-      real(8), dimension(10,10) :: WAGAUSJ
-c$OMP THREADPRIVATE (WAGAUSJ)
-c
-c
-c======================================================================
-c  Integration points and corresponding weights for 2-D Gaussian      |
-c  integration on a triangle                                          |
-c                                                                     |
-c  XIGAUSS(i,j,k) = i-th affine coordinate of j-th point of k-th      |
-c                   points set                                        |
-c  WAGAUSS(j,k)   = weight at j-th point of k-th points set           |
-c  NRGAUPO(k)     = number of Gauss points in the k-th points set     |
-c  NSELECT(k)     = number of points set to integrate exactly         |
-c                   mass matrix for an affine element and order k     |
-c======================================================================
-c
-      real(8), dimension(3,79,20) :: XIGAUSS
-c$OMP THREADPRIVATE (XIGAUSS)
-      real(8), dimension(  79,20) :: WAGAUSS
-c$OMP THREADPRIVATE (WAGAUSS)
-c
-      integer, parameter, dimension(20) :: NRGAUPO =
-     . (/ 1,3,4,6,7,12,13,16,19,25,27,33,37,42,48,52,61,70,73,79 /)
-c
-      integer, parameter, dimension( 9) :: NSELECT =
-     . (/ 2,4,6,8,10,12,14,17,19 /)
-c
-c
-c======================================================================
-c  Integration points and corresponding weights for 3-D Gaussian      |
-c  integration on the master tetrahedron                              |
-c                                                                     |
[…]
