### 2024-03-01 979c063
update legacy code in broken shape function routines
