+++ trunk/src/element/shape_1/broken/BrokenTetrahedron.F
@@ -1,35 +1,35 @@
-c Routines:
-c  - shape3DHBrokenTet
-c  - shape3DEBrokenTet
-c  - shape3DVBrokenTet
-c  - shape3DQBrokenTet
-c--------------------------------------------------------------------
-c
-c     routine name      - shape3DHBrokenTet
-c
-c--------------------------------------------------------------------
-c
-c     latest revision:  - Apr 17
-c
-c     purpose:          - routine returns values of 3D tetrahedron
-c                         BROKEN H1 shape functions
-c
-c     arguments:
-c
-c     in:
-c          X            - master tetrahedron coordinates from (0,1)^3
-c          NordM        - polynomial order for middle node (H1 sense)
-c          Nsize        - relevant sizes of local arrays
-c
-c     out:
-c          NrdofH       - number of dof
-c          ShapH        - values of the shape functions at the point
-c          GradH        - gradients of the shape functions
-c
-c-----------------------------------------------------------------------
-c
+! Routines:
+!  - shape3DHBrokenTet
+!  - shape3DEBrokenTet
+!  - shape3DVBrokenTet
+!  - shape3DQBrokenTet
+!--------------------------------------------------------------------
+!
+!     routine name      - shape3DHBrokenTet
+!
+!--------------------------------------------------------------------
+!
+!     latest revision:  - Apr 17
+!
+!     purpose:          - routine returns values of 3D tetrahedron
+!                         BROKEN H1 shape functions
+!
+!     arguments:
+!
+!     in:
+!          X            - master tetrahedron coordinates from (0,1)^3
+!          NordM        - polynomial order for middle node (H1 sense)
+!          Nsize        - relevant sizes of local arrays
+!
+!     out:
+!          NrdofH       - number of dof
+!          ShapH        - values of the shape functions at the point
+!          GradH        - gradients of the shape functions
+!
+!-----------------------------------------------------------------------
+!
       subroutine shape3DHBrokenTet(Xi,NordM,Nsize, NrdofH,ShapH,GradH)
-c
+!
       implicit none
       integer, intent(in)  :: NordM,Nsize(2)
       integer, intent(out) :: NrdofH
@@ -37,48 +37,48 @@ c
       double precision, intent(in)  :: Xi(3)
       double precision, intent(out) :: ShapH(Nsize(2))
       double precision, intent(out) :: GradH(1:3,Nsize(2))
-c
-c  ...Option 1: Simply call the usual shape functions with enrichment
+!
+!  ...Option 1: Simply call the usual shape functions with enrichment
       noriE(1:6)=0
       noriF(1:4)=0
       norder(1:11)=NordM
-      call shape3DHTet(Xi,norder,noriE,noriF,Nsize,
-     .                                             NrdofH,ShapH,GradH)
-c
-c
-c  ...Option 2: Write more efficient routine for enriched functions
-c
-c
+      call shape3DHTet(Xi,norder,noriE,noriF,Nsize, &
+                                                   NrdofH,ShapH,GradH)
+!
+!
+!  ...Option 2: Write more efficient routine for enriched functions
+!
+!
       end subroutine shape3DHBrokenTet
-c
-c
-c--------------------------------------------------------------------
-c
-c     routine name      - shape3DEBrokenTet
-c
-c--------------------------------------------------------------------
-c
-c     latest revision:  - Apr 17
-c
-c     purpose:          - routine returns values of 3D tetrahedron
-c                         BROKEN H(curl) shape functions
-c
-c     arguments:
-c
-c     in:
-c          X            - master tetrahedron coordinates from (0,1)^3
-c          NordM        - polynomial order for middle node (H1 sense)
-c          Nsize        - relevant sizes of local arrays
-c
-c     out:
[…]
