+++ trunk/src/element/shape_1/broken/BrokenPrism.F
@@ -1,74 +1,74 @@
-c Routines:
-c  - shape3DHBrokenPris
-c  - shape3DEBrokenPris
-c  - shape3DVBrokenPris
-c  - shape3DQBrokenPris
-c--------------------------------------------------------------------
-c
-c     routine name      - shape3DHBrokenPris
-c
-c--------------------------------------------------------------------
-c
-c     latest revision:  - Apr 17
-c
-c     purpose:          - routine returns values of 3D prism
-c                         BROKEN H1 shape functions
-c
-c     arguments:
-c
-c     in:
-c          X            - master prism coordinates from (0,1)^3
-c          NordM        - polynomial order for middle node (H1 sense)
-c          Nsize        - relevant sizes of local arrays
-c
-c     out:
-c          NrdofH       - number of dof
-c          ShapH        - values of the shape functions at the point
-c          GradH        - gradients of the shape functions
-c
-c-----------------------------------------------------------------------
-c
+! Routines:
+!  - shape3DHBrokenPris
+!  - shape3DEBrokenPris
+!  - shape3DVBrokenPris
+!  - shape3DQBrokenPris
+!--------------------------------------------------------------------
+!
+!     routine name      - shape3DHBrokenPris
+!
+!--------------------------------------------------------------------
+!
+!     latest revision:  - Apr 17
+!
+!     purpose:          - routine returns values of 3D prism
+!                         BROKEN H1 shape functions
+!
+!     arguments:
+!
+!     in:
+!          X            - master prism coordinates from (0,1)^3
+!          NordM        - polynomial order for middle node (H1 sense)
+!          Nsize        - relevant sizes of local arrays
+!
+!     out:
+!          NrdofH       - number of dof
+!          ShapH        - values of the shape functions at the point
+!          GradH        - gradients of the shape functions
+!
+!-----------------------------------------------------------------------
+!
       subroutine shape3DHBrokenPris(Xi,NordM,Nsize, NrdofH,ShapH,GradH)
-c
+!
       use parameters , only : MODORDER
       use node_types
-c
+!
       implicit none
       integer, intent(in)  :: NordM,Nsize(2)
       integer, intent(out) :: NrdofH
       integer :: i12,i3,m
-c      integer :: noriE(9),noriF(5),norder(15)
+!      integer :: noriE(9),noriF(5),norder(15)
       integer :: nordB(2),ndofH12,ndofH3
       double precision, intent(in)  :: Xi(3)
       double precision, intent(out) :: ShapH(Nsize(2))
       double precision, intent(out) :: GradH(1:3,Nsize(2))
-      double precision :: shapH12((Nsize(1)+2)*(Nsize(1)+1)/2),
-     .                   dshapH12(1:2,(Nsize(1)+2)*(Nsize(1)+1)/2)
+      double precision :: shapH12((Nsize(1)+2)*(Nsize(1)+1)/2), &
+                         dshapH12(1:2,(Nsize(1)+2)*(Nsize(1)+1)/2)
       double precision :: shapH3(Nsize(1)+1),dshapH3(Nsize(1)+1)
-c
-c  ...Option 1: Simply call the usual shape functions with enrichment
-c      noriE(1:9)=0
-c      noriF(1:5)=0
-c      call decod(NordM,MODORDER,2, nordB)
-c      norder(1:6)=nordB(1)
-c      norder(7:9)=nordB(2)
-c      norder(10:11)=nordB(1)
-c      norder(12:14)=NordM
-c      norder(15)=NordM
[…]

[…]
