### 2024-03-01 7de54d1
update legacy code in shape function routines
