+++ trunk/src/element/shape_1/BlendProject.F90
@@ -1,611 +1,611 @@
-c Routines:
-c  - BlendSegV
-c  - BlendQuadV
-c  - BlendProjectQuadE
-c  - BlendTriV
-c  - ProjectTriE
-c  - BlendHexaV
-c  - BlendProjectHexaE
-c  - BlendProjectHexaF
-c  - BlendTetV
-c  - ProjectTetE
-c  - ProjectTetF
-c  - BlendPrisV
-c  - BlendProjectPrisME
-c  - BlendProjectPrisQE
-c  - BlendProjectPrisTF
-c  - ProjectPrisQF
-c  - BlendPyraV
-c  - BlendProjectPyraME
-c  - ProjectPyraTE
-c  - ProjectPyraQF
-c  - BlendProjectPyraTF
-c  - ProjectPyraLamTF
-c----------------------------------------------------------------------
-c  b=blending p=projecting V=vertex E=edge F=face M=mixed T=tri Q=quad
-c----------------------------------------------------------------------
-c----------------------------------------------------------------------
-c  Data consistent with module element_data:
-c     numbering of topological entities - vertices, edges, faces
-c     local orientations of edges and faces
-c----------------------------------------------------------------------
-      subroutine BlendSegV(Mu,DMu, MubV,DMubV)
-c
+! Routines:
+!  - BlendSegV
+!  - BlendQuadV
+!  - BlendProjectQuadE
+!  - BlendTriV
+!  - ProjectTriE
+!  - BlendHexaV
+!  - BlendProjectHexaE
+!  - BlendProjectHexaF
+!  - BlendTetV
+!  - ProjectTetE
+!  - ProjectTetF
+!  - BlendPrisV
+!  - BlendProjectPrisME
+!  - BlendProjectPrisQE
+!  - BlendProjectPrisTF
+!  - ProjectPrisQF
+!  - BlendPyraV
+!  - BlendProjectPyraME
+!  - ProjectPyraTE
+!  - ProjectPyraQF
+!  - BlendProjectPyraTF
+!  - ProjectPyraLamTF
+!----------------------------------------------------------------------
+!  b=blending p=projecting V=vertex E=edge F=face M=mixed T=tri Q=quad
+!----------------------------------------------------------------------
+!----------------------------------------------------------------------
+!  Data consistent with module element_data:
+!     numbering of topological entities - vertices, edges, faces
+!     local orientations of edges and faces
+!----------------------------------------------------------------------
+   subroutine BlendSegV(Mu,DMu, MubV,DMubV)
+!
       implicit none
       integer :: N,v
       double precision, intent(in)  :: Mu(0:1),DMu(0:1)
       double precision, intent(out) :: MubV(1:2),DMubV(1:2)
-c
-c  ...Info from module element_data - coordinates,connectivities:
-c         V=((0),(1))=>(v0,v1)
-c
+!
+!  ...Info from module element_data - coordinates,connectivities:
+!         V=((0),(1))=>(v0,v1)
+!
       N=1
-c
-c  ...2 vertices, each with one blending function
-c
-c  ...v=1 --> v0=(0)
+!
[…]
