+++ trunk/src/element/shape_1/ContExactSequence.F90
@@ -1,0 +1,634 @@
+! Routines:
+!  - shape1DH
+!  - shape1DQ
+!  - shape2DH
+!  - shape2DE
+!  - shape2DV
+!  - shape2DQ
+!  - shape3DH
+!  - shape3DE
+!  - shape3DV
+!  - shape3DQ
+!
+!----------------------------------------------------------------------
+!----------------------------------------------------------------------
+!
+!                            1D: H1--->L2
+!
+!----------------------------------------------------------------------
+!----------------------------------------------------------------------
+!
+!----------------------------------------------------------------------
+!                                 1D H1
+!----------------------------------------------------------------------
+!
+!     routine name      - shape1DH
+!
+!----------------------------------------------------------------------
+!
+!     latest revision:  - Feb 2023
+!
+!     purpose:          - evaluate 1D H1 shape functions
+!
+!     arguments:
+!
+!     in:
+!       Xi              - master element coordinates
+!       Nord            - polynomial order of edge node (H1 sense)
+!
+!     out:
+!       NrdofH          - number of shape functions
+!       ShapH           - values of shape functions
+!       GradH           - gradients of shape functions
+!
+!----------------------------------------------------------------------
+!
+   subroutine shape1DH(Xi,Nord, NrdofH,ShapH,GradH)
+!
+      use parameters , only : MAXP
+      use node_types
+      use physics
+!
+      implicit none
+      integer, intent(in)  :: Nord
+      integer, intent(out) :: NrdofH
+      double precision, intent(in)  :: Xi
+      double precision, intent(out) :: ShapH(MAXP+1)
+      double precision, intent(out) :: GradH(MAXP+1)
+!
+      integer :: nsize(2),norder(19)
+      norder = 0; norder(1) = Nord
+!
+      call checkorder(SEGM,CONTIN,norder,MAXP, nsize)
+      call shape1DHSeg(Xi,Nord,nsize, NrdofH,ShapH,GradH)
+!
+   end subroutine shape1DH
+!
+!----------------------------------------------------------------------
+!                                 1D L2
+!----------------------------------------------------------------------
+!
+!     routine name      - shape1DQ
+!
+!----------------------------------------------------------------------
+!
+!     latest revision:  - Feb 2023
+!
+!     purpose:          - evaluate 1D L2 shape functions
+!
+!     arguments:
+!
+!     in:
+!       Xi              - master element coordinates
+!       Nord            - polynomial order of edge node (H1 sense)
+!
+!     out:
+!       NrdofQ          - number of shape functions
+!       ShapQ           - values of shape functions
+!
+!----------------------------------------------------------------------
+!
+   subroutine shape1DQ(Xi,Nord, NrdofQ,ShapQ)
+!
+      use parameters , only : MAXP
+      use node_types
+      use physics
+!
+      implicit none
+      double precision, intent(in)  :: Xi
+      integer         , intent(in)  :: Nord
+      integer         , intent(out) :: NrdofQ
+      double precision, intent(out) :: ShapQ(MAXP)
+!
+      integer :: nsize(2),norder(19)
+      norder = 0; norder(1) = Nord
+!
+      call checkorder(SEGM,DISCON,norder,MAXP, nsize)
+      call shape1DQSeg(Xi,Nord,nsize, NrdofQ,ShapQ)
+!
+   end subroutine shape1DQ
+!
+!----------------------------------------------------------------------
+!----------------------------------------------------------------------
+!
+!                        2D: H1--->Hcurl--->L2
+!                            H1--->Hdiv --->L2 (rotated)
+!
+!----------------------------------------------------------------------
[…]
