+++ trunk/src/element/shape_1/Hexahedron.F90
@@ -1,40 +1,40 @@
-c Routines:
-c  - shape3DHHexa
-c  - shape3DEHexa
-c  - shape3DVHexa
-c  - shape3DQHexa
-c--------------------------------------------------------------------
-c
-c     routine name      - shape3DHHexa
-c
-c--------------------------------------------------------------------
-c
-c     latest revision:  - Oct 14, Apr 17, Jul 21
-c
-c     purpose:          - routine returns values of 3D hexahedron element
-c                         H1 shape functions and their derivatives
-c
-c     arguments:
-c
-c     in:
-c          X            - master hexahedron coordinates from (0,1)^3
-c          Nord         - polynomial order for the nodes (H1 sense)
-c          NoriE        - edge orientation
-c          NoriF        - face orientation
-c          Nsize        - relevant sizes of local arrays
-c
-c     out:
-c          NrdofH       - number of dof
-c          ShapH        - values of the shape functions at the point
-c          GradH        - gradients of the shape functions
-c
-c-----------------------------------------------------------------------
-c
-      subroutine shape3DHHexa(Xi,Nord,NoriE,NoriF,Nsize,
-     .                                              NrdofH,ShapH,GradH)
-c
+! Routines:
+!  - shape3DHHexa
+!  - shape3DEHexa
+!  - shape3DVHexa
+!  - shape3DQHexa
+!--------------------------------------------------------------------
+!
+!     routine name      - shape3DHHexa
+!
+!--------------------------------------------------------------------
+!
+!     latest revision:  - Oct 14, Apr 17, Jul 21
+!
+!     purpose:          - routine returns values of 3D hexahedron element
+!                         H1 shape functions and their derivatives
+!
+!     arguments:
+!
+!     in:
+!          X            - master hexahedron coordinates from (0,1)^3
+!          Nord         - polynomial order for the nodes (H1 sense)
+!          NoriE        - edge orientation
+!          NoriF        - face orientation
+!          Nsize        - relevant sizes of local arrays
+!
+!     out:
+!          NrdofH       - number of dof
+!          ShapH        - values of the shape functions at the point
+!          GradH        - gradients of the shape functions
+!
+!-----------------------------------------------------------------------
+!
+   subroutine shape3DHHexa(Xi,Nord,NoriE,NoriF,Nsize, &
+                                                    NrdofH,ShapH,GradH)
+!
       use parameters , only : MODORDER
-c
+!
       implicit none
       integer, intent(in)  :: Nord(19),NoriE(12),NoriF(6),Nsize(2)
       integer, intent(out) :: NrdofH
@@ -55,116 +55,116 @@ c
       double precision :: phiE(2:Nsize(1)),DphiE(1:3,2:Nsize(1))
       double precision :: phiQuad(2:Nsize(1),2:Nsize(1))
       double precision :: DphiQuad(1:3,2:Nsize(1),2:Nsize(1))
-c
+!
 #if HP3D_DEBUG
-c  ...debugging flag
+!  ...debugging flag
       integer :: iprint
       iprint=0
 #endif
-c
-c  ...spatial dimensions
+!
+!  ...spatial dimensions
       N=3
-c  ...initiate counter for shape functions
+!  ...initiate counter for shape functions
[…]

[…]
