+++ trunk/src/modules/element_data.F90
@@ -1,310 +1,310 @@
-c----------------------------------------------------------------------
-c> @brief   element data for elements of different type
-c> @date    Feb 2023
-c----------------------------------------------------------------------
-c
-      module element_data
-c
+!----------------------------------------------------------------------
+!> @brief   element data for elements of different type
+!> @date    Feb 2024
+!----------------------------------------------------------------------
+!
+   module element_data
+!
       use node_types
-c
+!
       implicit none
-c
-c----------------------------------------------------------------------
-c  MASTER ELEMENTS VERTEX COORDINATES (vertex enumeration)            |
-c----------------------------------------------------------------------
-      real(8), parameter, dimension(2,3) :: TRIAN_COORD =
-     .reshape(
-     .(/0.d0,0.d0, 1.d0,0.d0, 0.d0,1.d0/)
-     .,(/2,3/))
-c
-      real(8), parameter, dimension(2,4) :: QUADR_COORD =
-     .reshape(
-     .(/0.d0,0.d0, 1.d0,0.d0, 1.d0,1.d0, 0.d0,1.d0/)
-     .,(/2,4/))
-c
-      real(8), parameter, dimension(3,6) :: PRISM_COORD =
-     .reshape(
-     .(/0.d0,0.d0,0.d0, 1.d0,0.d0,0.d0, 0.d0,1.d0,0.d0,
-     .  0.d0,0.d0,1.d0, 1.d0,0.d0,1.d0, 0.d0,1.d0,1.d0/)
-     .,(/3,6/))
-c
-      real(8), parameter, dimension(3,8) :: BRICK_COORD =
-     .reshape(
-     .(/0.d0,0.d0,0.d0, 1.d0,0.d0,0.d0, 1.d0,1.d0,0.d0, 0.d0,1.d0,0.d0,
-     .  0.d0,0.d0,1.d0, 1.d0,0.d0,1.d0, 1.d0,1.d0,1.d0, 0.d0,1.d0,1.d0/)
-     .,(/3,8/))
-c
-      real(8), parameter, dimension(3,4) :: TETRA_COORD =
-     .reshape(
-     .(/0.d0,0.d0,0.d0, 1.d0,0.d0,0.d0, 0.d0,1.d0,0.d0,
-     .  0.d0,0.d0,1.d0/)
-     .,(/3,4/))
-c
-      real(8), parameter, dimension(3,5) :: PYRAM_COORD =
-     .reshape(
-     .(/0.d0,0.d0,0.d0, 1.d0,0.d0,0.d0, 1.d0,1.d0,0.d0, 0.d0,1.d0,0.d0,
-     .  0.d0,0.d0,1.d0/)
-     .,(/3,5/))
-c
-c----------------------------------------------------------------------
-c  EDGE_2_VERT CONNECTIVITIES (edge enumeration & orientation)        |
-c----------------------------------------------------------------------
-      integer, parameter, dimension(2,3) :: TRIAN_EDGE_TO_VERT =
-     .reshape(
-     .(/1,2, 2,3, 1,3/)
-     .,(/2,3/))
-c
-      integer, parameter, dimension(2,4) :: QUADR_EDGE_TO_VERT =
-     .reshape(
-     .(/1,2, 2,3, 4,3, 1,4/)
-     .,(/2,4/))
-c
-      integer, parameter, dimension(2,9) :: PRISM_EDGE_TO_VERT =
-     .reshape(
[…]
