### 2024-03-01 df28790
update legacy code in frontal solver interf
