+++ trunk/src/solver/frontal/interf/solin1.F90
@@ -1,50 +1,50 @@
-c------------------------------------------------------------------
-c
-c   routine name       - solin1
-c
-c------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - a frontal solver interface routine
-c
-c   arguments :
-c     in:
-c              Iel     - an element number
-c     out:
-c              Numdes  - number of destination vectors
-c              Zaldest - destination vectors
-c
-c----------------------------------------------------------------------
+!------------------------------------------------------------------
+!
+!   routine name       - solin1
+!
+!------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - a frontal solver interface routine
+!
+!   arguments :
+!     in:
+!              Iel     - an element number
+!     out:
+!              Numdes  - number of destination vectors
+!              Zaldest - destination vectors
+!
+!----------------------------------------------------------------------
 #include "typedefs.h"
       subroutine solin1(Iel, Numdes,Zaldest)
-c
+!
       use frsolmod , ONLY: IDESVE, NDESVE
       implicit none
-c
+!
       integer, intent(in)  :: Iel
       integer, intent(out) :: Numdes
-c
+!
       VTYPE  , intent(out) :: Zaldest(*)
-c
+!
       integer :: i,iaux,naux,np
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint=0
 #endif
-c
+!
       Numdes = NDESVE(1,Iel)
       np = NDESVE(2,Iel)
-c
+!
 #if HP3D_DEBUG
       if (iprint.eq.1) then
         write(*,*) 'solin1: Numdes,np = ',Numdes,np
-        write(*,*) 'solin1: NDESVE(1,2403),NDESVE(2,2403) = ',
-     .                      NDESVE(1,2403),NDESVE(2,2403)
+        write(*,*) 'solin1: NDESVE(1,2403),NDESVE(2,2403) = ', &
+                            NDESVE(1,2403),NDESVE(2,2403)
       endif
 #endif
-c
+!
       do i=1,Numdes
         iaux = IDESVE(np+i-1)
 #if HP3D_COMPLEX
@@ -52,16 +52,16 @@ c
 #else
         Zaldest(i) = dfloat(IDESVE(np+i-1))
 #endif
-c
-c  .....double check
+!
+!  .....double check
         naux = INT(Zaldest(i))
         if (naux.ne.IDESVE(np+i-1)) then
-          write(*,*) 'solin1: IDESVE(np+i-1),Zaldest(i) = ',
-     .                        IDESVE(np+i-1),Zaldest(i)
+          write(*,*) 'solin1: IDESVE(np+i-1),Zaldest(i) = ', &
+                              IDESVE(np+i-1),Zaldest(i)
           stop 1
         endif
       enddo
-c
+!
 #if HP3D_DEBUG
       if (iprint.eq.1) then
         write(*,7001) Iel
@@ -77,5 +77,5 @@ c
         call pause
       endif
 #endif
-c
+!
       end subroutine solin1
