+++ trunk/src/solver/frontal/interf/solve1.F90
@@ -1,23 +1,23 @@
-c------------------------------------------------------------------
-c
-c   routine name       - solve1
-c
-c------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine solves a user defined system
-c                        of equations interfacing with the
-c                        frontal solver routines
-c
-c   arguments
-c   in:
-c        Number_of_RHS  - number of right-hand sides (load vectors)
-c
-c------------------------------------------------------------------
+!------------------------------------------------------------------
+!
+!   routine name       - solve1
+!
+!------------------------------------------------------------------
+!
+!   latest revision    - Feb 2023
+!
+!   purpose            - routine solves a user defined system
+!                        of equations interfacing with the
+!                        frontal solver routines
+!
+!   arguments
+!   in:
+!        Number_of_RHS  - number of right-hand sides (load vectors)
+!
+!------------------------------------------------------------------
 #include "typedefs.h"
       subroutine solve1(Number_of_RHS)
-c
+!
       use data_structure3D
       use element_data
       use assembly
@@ -27,104 +27,104 @@ c
       use stc,         only: stc_alloc, stc_dealloc, stc_get_nrdof
       use par_mesh,    only: DISTRIBUTED,HOST_MESH
       use mpi_param,   only: RANK,ROOT
-c
-c  ...frontal solver common blocks
+!
+!  ...frontal solver common blocks
       use surfsc1
       use surfsc2
-c
+!
       implicit none
-c
+!
       integer, intent(in) :: Number_of_RHS
-c
-c  ...nodes for a modified element and the corresponding number
-c     of H1,H(curl),H(div) and L2 dof
-      integer :: nodm(MAXNODM),ndofmH(MAXNODM),ndofmE(MAXNODM),
-     .                         ndofmV(MAXNODM),ndofmQ(MAXNODM)
-c
-c  ...number of variables for each physics attribute for an element
+!
+!  ...nodes for a modified element and the corresponding number
+!     of H1,H(curl),H(div) and L2 dof
+      integer :: nodm(MAXNODM),ndofmH(MAXNODM),ndofmE(MAXNODM), &
+                               ndofmV(MAXNODM),ndofmQ(MAXNODM)
+!
+!  ...number of variables for each physics attribute for an element
       integer :: nrdofs(NR_PHYSA)
-c
-c  ...element geometry dof, direction vector
+!
+!  ...element geometry dof, direction vector
       real(8) :: xnod(3,MAXbrickH),xc(3),direction(3)
-c
-c  ...number of local element dof for each physics variable
+!
+!  ...number of local element dof for each physics variable
       integer :: nrdofi(NR_PHYSA),nrdofb(NR_PHYSA)
-c
+!
       integer :: inick,iel,iel1,i,j,iv,kk,kel
       integer :: max_active_node_no,mdle,mr,ms,mu,nod,nrnodm
       integer :: nrnick,nrdof,nrdofm,nrdofc,nrnod1,nrdof_mdl
       VTYPE   :: zvoid
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint=0
 #endif
-c
-c-----------------------------------------------------------------------
-c
+!
[…]
