### 2024-03-01 7be3397
update legacy code in frontal solver real
