+++ trunk/src/solver/frontal/base/real/bckwrdR.F90
@@ -1,84 +1,84 @@
-c***===***===***===***===***===***===***===***===***===***===***===***==
-c FUNCTION:
-c     Backsubstitution
-c     -----------------
-c     calls SOLIN1 for dest. vectors
-c     passes Elemental solutions to SOLOUT
-**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==
-c ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
-c
-c Typ Name      Function
-c W  Alelm   final length of the current front as we process each Elem
-c W  Aldest  nodal destination vectors
-c W  Amdest  dof destination vectors
-c W  Andest  Destination in the front where the remaining values
-c            in the front will transfer to as each exiting dof is elimin
-c
-c W  Elem    Temp storage for the holding the Element solution
-c             (ie: used to transfer x from the solution front to the
-c                  local Element numbering based on the dest vecs)
-c
-c W  Frnt    Temp storage for the solutions
-c            Note: on entry, this contains the previous values in the fr
-c                   the values at the end of sub.frwcp or sub.frwrs
-c                   but this info isnt used for anything
-c
-c W  Ubuf    The eliminated lhs buffered equations
-c W  Bbuf    The eliminated rhs buffered equations
-c
-c*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
-c LATEST REVISION: Mar 2023
-c++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
-c NAMING CONVENTIONS:
-c     AAAAAAAA    Variables in COMMON & PARAMETERS
-c     Aaaaaaaa    Variables as ARGUMENTS
-c     aaaaaaaa    LOCAL Variables
-c         7xxx    FORMAT Statements
-c         9xxx    ERROR Handling
-c+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
-      subroutine bckwrd (Alelm, Aldest, Amdest,
+!***===***===***===***===***===***===***===***===***===***===***===***==
+! FUNCTION:
+!     Backsubstitution
+!     -----------------
+!     calls SOLIN1 for dest. vectors
+!     passes Elemental solutions to SOLOUT
+!**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**=
+! ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
+!
+! Typ Name      Function
+! W  Alelm   final length of the current front as we process each Elem
+! W  Aldest  nodal destination vectors
+! W  Amdest  dof destination vectors
+! W  Andest  Destination in the front where the remaining values
+!            in the front will transfer to as each exiting dof is elimin
+!
+! W  Elem    Temp storage for the holding the Element solution
+!             (ie: used to transfer x from the solution front to the
+!                  local Element numbering based on the dest vecs)
+!
+! W  Frnt    Temp storage for the solutions
+!            Note: on entry, this contains the previous values in the fr
+!                   the values at the end of sub.frwcp or sub.frwrs
+!                   but this info isnt used for anything
+!
+! W  Ubuf    The eliminated lhs buffered equations
+! W  Bbuf    The eliminated rhs buffered equations
+!
+!*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
+! LATEST REVISION: Mar 2023
[…]
