+++ trunk/src/solver/frontal/base/real/unsasmR.F90
@@ -1,110 +1,110 @@
-c***===***===***===***===***===***===***===***===***===***===***===***==
-c FUNCTION: Assembles lhs into the front for unsymmetric matrices
-c**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**=
-c ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
-c
-c Typ Name      Function
-c
-c I   Amdest:  dof destination vectors
-c I   Ellhs :  element lhs
-c IO  Flhs  :  front assembled lhs
-c               note: for unsymmetric, lhs is stored as:
-c               a11 a12 a13 ...
-c               a21 a22 a23 ...   ==>  [a11,a12,a13,..., a21,a22,a23, ..
-c               a31 a32 a33 ...
-c                :   :   :
-c*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
-c LATEST REVISION: Mar 2023
-c++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
-c NAMING CONVENTIONS:
-c     AAAAAAAA    Variables in COMMON & PARAMETERS
-c     Aaaaaaaa    Variables as ARGUMENTS
-c     aaaaaaaa    LOCAL Variables
-c         7xxx    FORMAT Statements
-c         9xxx    ERROR Handling
-c+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
-c
+!***===***===***===***===***===***===***===***===***===***===***===***==
+! FUNCTION: Assembles lhs into the front for unsymmetric matrices
+!**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**=
+! ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
+!
+! Typ Name      Function
+!
+! I   Amdest:  dof destination vectors
+! I   Ellhs :  element lhs
+! IO  Flhs  :  front assembled lhs
+!               note: for unsymmetric, lhs is stored as:
+!               a11 a12 a13 ...
+!               a21 a22 a23 ...   ==>  [a11,a12,a13,..., a21,a22,a23, ..
+!               a31 a32 a33 ...
+!                :   :   :
+!*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
+! LATEST REVISION: Mar 2023
+!++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
+! NAMING CONVENTIONS:
+!     AAAAAAAA    Variables in COMMON & PARAMETERS
+!     Aaaaaaaa    Variables as ARGUMENTS
+!     aaaaaaaa    LOCAL Variables
+!         7xxx    FORMAT Statements
+!         9xxx    ERROR Handling
+!+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
+!
       subroutine unsasm (Amdest, Ellhs, Flhs)
-c
+!
       use surfsc1
       use surfsc2
-c
+!
       implicit none
-c
+!
       real(8) :: Amdest(*), Ellhs(*), Flhs(*)
       integer :: ntemp(324)
-c
+!
       integer :: n,i,j,mdi,mdj,ml
-c
+!
       n = 1
-c
+!
[…]

[…]
