+++ trunk/src/solver/frontal/base/complex/frwcp.F90
@@ -1,171 +1,171 @@
-c***===***===***===***===***===***===***===***===***===***===***===***==
-c FUNCTION:
-c Forward elimination of both lhs and rhs
-c  calls solin1 for dest. vectors
-c  calls solin2 for lhs and rhs's
-c
-c**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**=
-c ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
-c
-c Typ Name      Function
-c W   Alelm   length of the current front as we process each Elem
-c W   Aldest  nodal destination vectors
-c W   Amdest  dof destination vectors
-c W   Andest  destination in the front where the remaining values
-c              in the front will transfer to as each exiting dof is elim
-c W   Elem    Element lhs + rhs
-c W   Frnt    equations in the front
-c W   Buf     eliminated equations
-c
-c*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
-c LATEST REVISION: Mar 2023
-c++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
-c NAMING CONVENTIONS:
-c     AAAAAAAA    Variables in COMMON & PARAMETERS
-c     Aaaaaaaa    Variables as ARGUMENTS
-c     aaaaaaaa    LOCAL Variables
-c         7xxx    FORMAT Statements
-c         9xxx    ERROR Handling
-c+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
-c
+!***===***===***===***===***===***===***===***===***===***===***===***==
+! FUNCTION:
+! Forward elimination of both lhs and rhs
+!  calls solin1 for dest. vectors
+!  calls solin2 for lhs and rhs's
+!
+!**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**=
+! ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
+!
+! Typ Name      Function
+! W   Alelm   length of the current front as we process each Elem
+! W   Aldest  nodal destination vectors
+! W   Amdest  dof destination vectors
+! W   Andest  destination in the front where the remaining values
+!              in the front will transfer to as each exiting dof is elim
+! W   Elem    Element lhs + rhs
+! W   Frnt    equations in the front
+! W   Buf     eliminated equations
+!
+!*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
+! LATEST REVISION: Mar 2023
+!++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
+! NAMING CONVENTIONS:
+!     AAAAAAAA    Variables in COMMON & PARAMETERS
+!     Aaaaaaaa    Variables as ARGUMENTS
+!     aaaaaaaa    LOCAL Variables
+!         7xxx    FORMAT Statements
+!         9xxx    ERROR Handling
+!+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
+!
       subroutine frwcp (Alelm, Aldest, Amdest, Andest, Elem, Frnt, Buf)
-c
+!
       use surfsc1
       use surfsc2
-c
+!
       implicit none
-c
-      complex(8) :: Aldest(*),Amdest(*),Andest(*),Elem(*),Frnt(*),
[…]
