+++ trunk/src/solver/frontal/base/complex/zdiriodum.F90
@@ -1,139 +1,139 @@
-c***===***===***===***===***===***===***===***===***===***===***===***==
-c FUNCTION: This routine preforms direct access I/O for scratch tapes
-c**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**=
-c ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
-c
-c Typ Name      Function
-c I   Unname - name of the unit to operate on
-c               (e.g. for surfs: 'U', 'B', or 'L')
-c I   Commnd - character variable that tells what i/o action to take
-c               it may be equal to 'OPEN', 'CLOSE', 'READ', or 'WRITE'
-c I   Irec   - logical record number to read or write (not used for open
-c
-c I   Len    - for write: the number of words to write
-c O   Len      for read: the number of words read
-c I   Len      for open: the maximum length in words for a record
-c
-c I   Sbuf   - for write: array for writing out to
-c O   Sbuf   - for read:  array for reading into
-c
-c O   Jerr   - error flag set to 0 for no error
-c*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
-c LATEST REVISION: Mar 2023
-c++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
-c NAMING CONVENTIONS:
-c     AAAAAAAA    Variables in COMMON & PARAMETERS
-c     Aaaaaaaa    Variables as ARGUMENTS
-c     aaaaaaaa    LOCAL Variables
-c         7xxx    FORMAT Statements
-c         9xxx    ERROR Handling
-c+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
-c
+!***===***===***===***===***===***===***===***===***===***===***===***==
+! FUNCTION: This routine preforms direct access I/O for scratch tapes
+!**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**=
+! ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
+!
+! Typ Name      Function
+! I   Unname - name of the unit to operate on
+!               (e.g. for surfs: 'U', 'B', or 'L')
+! I   Commnd - character variable that tells what i/o action to take
+!               it may be equal to 'OPEN', 'CLOSE', 'READ', or 'WRITE'
+! I   Irec   - logical record number to read or write (not used for open
+!
+! I   Len    - for write: the number of words to write
+! O   Len      for read: the number of words read
+! I   Len      for open: the maximum length in words for a record
+!
+! I   Sbuf   - for write: array for writing out to
+! O   Sbuf   - for read:  array for reading into
+!
+! O   Jerr   - error flag set to 0 for no error
+!*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
+! LATEST REVISION: Mar 2023
+!++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
+! NAMING CONVENTIONS:
+!     AAAAAAAA    Variables in COMMON & PARAMETERS
+!     Aaaaaaaa    Variables as ARGUMENTS
+!     aaaaaaaa    LOCAL Variables
+!         7xxx    FORMAT Statements
[…]

[…]
