+++ trunk/src/solver/frontal/base/real/zerorR.F90
@@ -16,7 +16,7 @@
 !         9xxx    ERROR Handling
 !+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
 !
-      subroutine zeror (Frhs)
+   subroutine zeror (Frhs)
 !
       use surfsc1
       use surfsc2
@@ -51,4 +51,4 @@
 120   continue
 !
 !
-      end subroutine zeror
+   end subroutine zeror
