+++ trunk/src/solver/frontal/base/complex/zeros.F90
@@ -22,7 +22,7 @@
 !         9xxx    ERROR Handling
 !+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
 !
-      subroutine zeros (Flhs)
+   subroutine zeros (Flhs)
 !
       use surfsc2
 !
@@ -54,4 +54,4 @@
       enddo
 !
 !
-      end subroutine zeros
+   end subroutine zeros
