+++ trunk/src/solver/frontal/base/complex/zeror.F90
@@ -50,7 +50,7 @@
          m = LFW + 1
 !
 ! ALLIANT directives
-!cvd$ select (vector)
+!vd$ select (vector)
 ! ARDENT directives
 !$doit VBEST
 !
