+++ trunk/src/solver/frontal/base/real/zerorR.F90
@@ -40,7 +40,7 @@
          m = LFW + 1
 !
 ! ALLIANT directives
-!cvd$ select (vector)
+!vd$ select (vector)
 ! ARDENT directives
 !$doit VBEST
 !

[…]
