### 2024-03-01 26ca0cd
update legacy code in GMP module
