+++ trunk/src/GMP/module/GMP.F90
@@ -1,160 +1,160 @@
-c----------------------------------------------------------------------
-c> @brief Data structure arrays for Geometry Modeling Package (GMP)
-c
-c> @date Mar 2023
-c----------------------------------------------------------------------
-c
-      module GMP
-c
+!----------------------------------------------------------------------
+!> @brief Data structure arrays for Geometry Modeling Package (GMP)
+!
+!> @date Mar 2023
+!----------------------------------------------------------------------
+!
+   module GMP
+!
       implicit none
-c
-c----------------------------------------------------------------------
-c  PARAMETERS                                                         |
-c----------------------------------------------------------------------
-c     dimension of the problem (2, 3)
+!
+!----------------------------------------------------------------------
+!  PARAMETERS                                                         |
+!----------------------------------------------------------------------
+!     dimension of the problem (2, 3)
       integer, save :: NDIM
-c     dimension of the manifold (2, 3)
+!     dimension of the manifold (2, 3)
       integer, save :: MANDIM
-c     number of algebraic surfaces
+!     number of algebraic surfaces
       integer, save :: NRSURFS
-c     number of points
+!     number of points
       integer, save :: NRPOINT
-c     number of curves
+!     number of curves
       integer, save :: NRCURVE
-c     number of triangles
+!     number of triangles
       integer, save :: NRTRIAN
-c     number of rectangles
+!     number of rectangles
       integer, save :: NRRECTA
-c     number of hexahedra
+!     number of hexahedra
       integer, save :: NRHEXAS
-c     number of tetrahedra
+!     number of tetrahedra
       integer, save :: NRTETRA
-c     number of prisms
+!     number of prisms
       integer, save :: NRPRISM
-c     number of pyramids
+!     number of pyramids
       integer, save :: NRPYRAM
-c     number of subdomains
+!     number of subdomains
       integer, save :: NRDOMAIN
-c     number of fine grid points
+!     number of fine grid points
       integer, save :: NRPO_FG
-c     fine grid scale
+!     fine grid scale
       real(8), save :: SCALE_FG
-c
-c     maximum number of algebraic surfaces
+!
+!     maximum number of algebraic surfaces
       integer, save :: MAXSU
-c     maximum number of points
+!     maximum number of points
       integer, save :: MAXNP
-c     maximum number of curves
+!     maximum number of curves
       integer, save :: MAXNC
-c     maximum number of triangles
+!     maximum number of triangles
       integer, save :: MAXTR
-c     maximum number of rectangles
+!     maximum number of rectangles
       integer, save :: MAXRE
-c     maximum number of prismatic blocks
+!     maximum number of prismatic blocks
       integer, save :: MAXBT
-c     maximum number of hexahedral blocks
+!     maximum number of hexahedral blocks
       integer, save :: MAXHE
-c     maximum number of tetrahedral blocks
+!     maximum number of tetrahedral blocks
       integer, save :: MAXTE
-c     maximum number of pyramids
+!     maximum number of pyramids
       integer, save :: MAXPY
-c     maximum number of fine grid points
+!     maximum number of fine grid points
       integer, save :: MAXNP_FG
-c
-c
-c----------------------------------------------------------------------
-c  SURFACE                                                            |
-c----------------------------------------------------------------------
-c
-c  Type =
[…]
