### 2024-03-01 b2cfac9
update legacy code in GMP shape0
