+++ trunk/src/GMP/shape_0/projections.F
@@ -1,315 +1,0 @@
-c----------------------------------------------------------------------
-c> Purpose : triangle to edge projection
-c
-c  @param[in ] Iv1,2   - vertices defining the edge
-c  @param[in ] Vshape  - master triangle vertex shape functions,
-c                        computed at the point of interest
-c  @param[in ] Dvshape - derivatives of vertex shape functions
-c  @param[out] T       - edge coordinate
-c  @param[out] Dtdxi   - derivative of edge coordinate
-c
-c  @revision Nov 12
-c----------------------------------------------------------------------
-      subroutine proj_t2e(Iv1,Iv2,Vshape,Dvshape, T,Dtdxi)
-c
-      implicit none
-      integer               ,intent(in ) :: Iv1,Iv2
-      real(8),dimension(  3),intent(in ) :: Vshape
-      real(8),dimension(2,3),intent(in ) :: Dvshape
-      real(8)               ,intent(out) :: T
-      real(8),dimension(2)  ,intent(out) :: Dtdxi
-c----------------------------------------------------------------------
-c
-      T = (Vshape(Iv2)-Vshape(Iv1)+1.d0)*.5d0
-      Dtdxi(1:2) = (Dvshape(1:2,Iv2)-Dvshape(1:2,Iv1))*.5d0
-c
-      end subroutine proj_t2e
-c
-c
-c----------------------------------------------------------------------
-c> Purpose : quad to edge projection
-c
-c  @param[in ] Iv1,2   - vertices defining the edge
-c  @param[in ] Vshape  - master quad vertex shape functions, computed
-c                        at the point of interest
-c  @param[in ] Dvshape - derivatives of vertex shape functions
-c  @param[out] T       - edge coordinate
-c  @param[out] Dtdxi   - derivative of edge coordinate
-c
-c  @revision Nov 12
-c----------------------------------------------------------------------
-      subroutine proj_r2e(Iv1,Iv2,Vshape,Dvshape, T,Dtdxi)
-c
-      use element_data , only : QUADR_EDGE_TO_VERT
-c
-      implicit none
-      integer               ,intent(in ) :: Iv1,Iv2
-      real(8),dimension(  4),intent(in ) :: Vshape
-      real(8),dimension(2,4),intent(in ) :: Dvshape
-      real(8)               ,intent(out) :: T
-      real(8),dimension(2)  ,intent(out) :: Dtdxi
-c
-      integer :: iedge,i,iv
-c----------------------------------------------------------------------
-c
-      call quad_aux(Iv1,Iv2, iedge)
-c
-c  ...loop over edge vertices
-      T=0.d0 ; Dtdxi(1:2)=0.d0
-      do i=1,2
-        iv=QUADR_EDGE_TO_VERT(i,iedge)
-        T          = T          +  Vshape(    iv)
-        Dtdxi(1:2) = Dtdxi(1:2) + Dvshape(1:2,iv)
-      enddo
-c
-      end subroutine proj_r2e
-c
-c
-c----------------------------------------------------------------------
-c> Purpose : tetrahedroN to edge projection
-c
-c  @param[in ] Iv1,2   - vertices defining the edge
-c  @param[in ] Vshape  - master tet vertex shape functions, computed
-c                        at the point of interest
-c  @param[in ] Dvshape - derivatives of vertex shape functions
-c  @param[out] T       - edge coordinate
-c  @param[out] Dtdxi   - derivative of edge coordinate
-c
-c  @revision Nov 12
[…]
