+++ trunk/src/GMP/shape_0/projections.F90
@@ -1,0 +1,315 @@
+!----------------------------------------------------------------------
+!> Purpose : triangle to edge projection
+!
+!  @param[in ] Iv1,2   - vertices defining the edge
+!  @param[in ] Vshape  - master triangle vertex shape functions,
+!                        computed at the point of interest
+!  @param[in ] Dvshape - derivatives of vertex shape functions
+!  @param[out] T       - edge coordinate
+!  @param[out] Dtdxi   - derivative of edge coordinate
+!
+!  @revision Nov 12
+!----------------------------------------------------------------------
+      subroutine proj_t2e(Iv1,Iv2,Vshape,Dvshape, T,Dtdxi)
+!
+      implicit none
+      integer               ,intent(in ) :: Iv1,Iv2
+      real(8),dimension(  3),intent(in ) :: Vshape
+      real(8),dimension(2,3),intent(in ) :: Dvshape
+      real(8)               ,intent(out) :: T
+      real(8),dimension(2)  ,intent(out) :: Dtdxi
+!----------------------------------------------------------------------
+!
+      T = (Vshape(Iv2)-Vshape(Iv1)+1.d0)*.5d0
+      Dtdxi(1:2) = (Dvshape(1:2,Iv2)-Dvshape(1:2,Iv1))*.5d0
+!
+      end subroutine proj_t2e
+!
+!
+!----------------------------------------------------------------------
+!> Purpose : quad to edge projection
+!
+!  @param[in ] Iv1,2   - vertices defining the edge
+!  @param[in ] Vshape  - master quad vertex shape functions, computed
+!                        at the point of interest
+!  @param[in ] Dvshape - derivatives of vertex shape functions
+!  @param[out] T       - edge coordinate
+!  @param[out] Dtdxi   - derivative of edge coordinate
+!
+!  @revision Nov 12
+!----------------------------------------------------------------------
+      subroutine proj_r2e(Iv1,Iv2,Vshape,Dvshape, T,Dtdxi)
+!
+      use element_data , only : QUADR_EDGE_TO_VERT
+!
+      implicit none
+      integer               ,intent(in ) :: Iv1,Iv2
+      real(8),dimension(  4),intent(in ) :: Vshape
+      real(8),dimension(2,4),intent(in ) :: Dvshape
+      real(8)               ,intent(out) :: T
+      real(8),dimension(2)  ,intent(out) :: Dtdxi
+!
+      integer :: iedge,i,iv
+!----------------------------------------------------------------------
+!
+      call quad_aux(Iv1,Iv2, iedge)
+!
+!  ...loop over edge vertices
+      T=0.d0 ; Dtdxi(1:2)=0.d0
+      do i=1,2
+        iv=QUADR_EDGE_TO_VERT(i,iedge)
+        T          = T          +  Vshape(    iv)
+        Dtdxi(1:2) = Dtdxi(1:2) + Dvshape(1:2,iv)
+      enddo
+!
+      end subroutine proj_r2e
+!
+!
+!----------------------------------------------------------------------
+!> Purpose : tetrahedroN to edge projection
+!
+!  @param[in ] Iv1,2   - vertices defining the edge
+!  @param[in ] Vshape  - master tet vertex shape functions, computed
+!                        at the point of interest
+!  @param[in ] Dvshape - derivatives of vertex shape functions
+!  @param[out] T       - edge coordinate
+!  @param[out] Dtdxi   - derivative of edge coordinate
+!
+!  @revision Nov 12
[…]
