### 2024-03-01 a7b2d9c
Update legacy code in shape functions package (#177)
* updating legacy code in element/util

* update legacy code in gauss_quadrature

* update legacy code in broken shape function routines

* update legacy code in shape function routines
