+++ trunk/src/element/shape_1/CheckOrder.F90
@@ -1,39 +1,39 @@
-c Routines:
-c  - checkorder
-c----------------------------------------------------------------------
-c
-c     routine name      - checkorder
-c
-c----------------------------------------------------------------------
-c
-c     latest revision:  - Feb 2023
-c
-c     purpose:          - routine checks whether polynomial orders of
-c                         edges, faces and bubbles are within
-c                         appropriate bounds. Returns values that help
-c                         to more efficiently size temporary arrays in
-c                         other routines in order to save memory.
-c
-c     arguments:
-c
-c     in:
-c       Ntype           - element type
-c       Dtype           - discretization type (H1,H(curl),H(div),L2)
-c       Norder          - polynomial order for the nodes (H1 sense)
-c       MaxOrd          - maximum polynomial order allowed
-c
-c     out:
-c       Nsize           - highest polynomial order in Norder and other
-c                         values that help appropriately size arrays
-c
-c----------------------------------------------------------------------
-c
-      subroutine checkorder(Ntype,Dtype,Norder,MaxOrd, Nsize)
-c
+! Routines:
+!  - checkorder
+!----------------------------------------------------------------------
+!
+!     routine name      - checkorder
+!
+!----------------------------------------------------------------------
+!
+!     latest revision:  - Feb 2023
+!
+!     purpose:          - routine checks whether polynomial orders of
+!                         edges, faces and bubbles are within
+!                         appropriate bounds. Returns values that help
+!                         to more efficiently size temporary arrays in
+!                         other routines in order to save memory.
+!
+!     arguments:
+!
+!     in:
+!       Ntype           - element type
+!       Dtype           - discretization type (H1,H(curl),H(div),L2)
+!       Norder          - polynomial order for the nodes (H1 sense)
+!       MaxOrd          - maximum polynomial order allowed
+!
+!     out:
+!       Nsize           - highest polynomial order in Norder and other
+!                         values that help appropriately size arrays
+!
+!----------------------------------------------------------------------
+!
+   subroutine checkorder(Ntype,Dtype,Norder,MaxOrd, Nsize)
+!
       use parameters , only : MODORDER
       use node_types
       use physics
-c
+!
       implicit none
       integer, intent(in)  :: Ntype
       integer, intent(in)  :: Dtype
@@ -42,23 +42,23 @@ c
       integer, intent(out) :: Nsize(2)
       integer :: minp,maxp,nps(27),m,f,nordF(2),nordB(3)
       logical :: confident
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint = 0
 #endif
-c
-c  ...initialize output variable
+!
+!  ...initialize output variable
       Nsize = 0
-c
-c  ...The flag confident determines whether you are confident enough to
-c     NOT make a check to Norder. In that case, it is assumed all orders
-c     are less than MaxOrd, and the size of the arrays assume the worst
-c     case scenario (with MaxOrd).
-c  ...If you are not confident enough (default), this routine will check
-c     whether the values in Norder lie below MaxOrd, and will determine
-c     the maximum order in Norder and size temporary arrays accordingly
-c     in the shape element routines.
+!
[…]
