+++ trunk/src/element/util/celndof.F90
@@ -1,60 +1,60 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - celndof
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Feb 2023
-c
-c   purpose            - routine determines number of single dof
-c                        for a 3D element
-c
-c   arguments :
-c     in:
-c             Ntype    - middle node type
-c             Norder   - order of approximation for the element
-c     out:
-c             NrdofH   - number of H1 dof
-c             NrdofE   - number of H(curl) dof
-c             NrdofV   - number of H(div) dof
-c             NrdofQ   - number of L2 dof
-c
-c----------------------------------------------------------------------
-c
-      subroutine celndof(Ntype,Norder, NrdofH,NrdofE,NrdofV,NrdofQ)
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - celndof
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Feb 2024
+!
+!   purpose            - routine determines number of single dof
+!                        for a 3D element
+!
+!   arguments :
+!     in:
+!             Ntype    - middle node type
+!             Norder   - order of approximation for the element
+!     out:
+!             NrdofH   - number of H1 dof
+!             NrdofE   - number of H(curl) dof
+!             NrdofV   - number of H(div) dof
+!             NrdofQ   - number of L2 dof
+!
+!----------------------------------------------------------------------
+!
+   subroutine celndof(Ntype,Norder, NrdofH,NrdofE,NrdofV,NrdofQ)
+!
       use element_data
-c
+!
       implicit none
-c
+!
       integer, intent(in)  :: Ntype,Norder(19)
       integer, intent(out) :: NrdofH,NrdofE,NrdofV,NrdofQ
-c
+!
       integer :: i,j
       integer :: nredg
       integer :: ndofH,ndofE,ndofV,ndofQ
-c
+!
 #if HP3D_DEBUG
       integer :: iprint
       iprint=0
 #endif
       NrdofH=0; NrdofE=0; NrdofV=0; NrdofQ=0
-c
-c  ...V E R T I C E S..................................................
+!
+!  ...V E R T I C E S..................................................
       do i=1,nvert(Ntype)
         call ndof_nod(VERT,1, ndofH,ndofE,ndofV,ndofQ)
 #if HP3D_DEBUG
         if (iprint.eq.1) then
           write(*,7001) 'vert',1,ndofH,ndofE,ndofV,ndofQ
- 7001     format('celndof: type,nord,ndofH,ndofE,ndofV,ndofQ = ',
-     .                     a5,i4,4i6)
+ 7001     format('celndof: type,nord,ndofH,ndofE,ndofV,ndofQ = ', &
+                           a5,i4,4i6)
         endif
 #endif
         NrdofH=NrdofH+ndofH; NrdofE=NrdofE+ndofE
         NrdofV=NrdofV+ndofV; NrdofQ=NrdofQ+ndofQ
       enddo
-c
-c  ...E D G E S........................................................
+!
+!  ...E D G E S........................................................
       nredg = nedge(Ntype)
       do i=1,nredg
[…]

[…]
