### 2024-03-01 071c797
update legacy code in GMP routines
