+++ trunk/src/GMP/graphics/GMPwireframe.F90
@@ -1,128 +1,128 @@
 #if HP3D_USE_X11
 
-c----------------------------------------------------------------------
-c
-c   routine name       - GMPwireframe
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Mar 2023
-c
-c   purpose            - routine displays the wireframe for a 3D manifold
-c
-c
-c   arguments :
-c     in:
-c               Iwind  - window number
-c
-c----------------------------------------------------------------------
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - GMPwireframe
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Mar 2023
+!
+!   purpose            - routine displays the wireframe for a 3D manifold
+!
+!
+!   arguments :
+!     in:
+!               Iwind  - window number
+!
+!----------------------------------------------------------------------
+!
       subroutine GMPwireframe(Iwind)
-c
+!
       use graphmod
       use GMP
-c
+!
       implicit none
-c
+!
       integer :: Iwind
-c
-c  ...list of segments
+!
+!  ...list of segments
       real(8), allocatable ::  SEGMENTS(:,:,:)
       integer, allocatable ::  COLORN(:)
-c
-c  ...list of rectangles to display
+!
+!  ...list of rectangles to display
       integer, parameter :: max_nr_rect=10
       integer :: nrect(max_nr_rect)
-c
-c  ...list of curves to display
+!
+!  ...list of curves to display
       integer, parameter :: max_nr_curv=10
       integer :: ncurv(max_nr_curv)
-c
-c  ...point, derivatives
+!
+!  ...point, derivatives
       real(8) :: x(3),dxdeta(3),xob(3),xyp(2,2)
-c
-c  ...bounds
+!
+!  ...bounds
       real(8) :: xmax(2),xmin(2)
       integer :: im(2)
-c
+!
       real(8) :: eta,psi,q,q1,rnsc,theta
       integer :: ic,ichoice,idec,idec_bundle,idec_curve,idec_edge
       integer :: idec_Hermite,idec_normal,idec_rectangle
       integer :: ii,ip,ir,j,k,lab,nc,ncol,np,npoint
       integer :: nr,nr_curv,nr_curves,nr_rect,nr_segments,nrsub1
-c
+!
       integer :: iprint
-c
+!
       real(8) :: bigp,bign,small,pi
       data bigp,bign,small /1.d30,-1.d30,1.d-14/
       data pi /3.14159265358979312d0/
-c
-c***********************************************************************
+!
+!***********************************************************************
       iprint=0
-c
-c  ...set up default values for control parameters
+!
+!  ...set up default values for control parameters
       idec_Hermite=0
       idec_rectangle=0
       idec_curve=0
       idec_bundle=0
       idec_normal=0
       idec_edge=0
-c
-c  ...set initial scaling constants
+!
+!  ...set initial scaling constants
       XCIM(1) = 0.d0
       XCIM(2) = 0.d0
       xmax(1:2) = bign
       xmin(1:2) = bigp
       XCWIN(1) = rmargin + xlength/2.d0
       XCWIN(2) = rmargin + ylength/2.d0
-c
-c  ...set up number of line segments for drawing a curve
+!
[…]
