+++ trunk/src/GMP/base/triangles/trian_CylTri.F90
@@ -1,72 +1,72 @@
-c-----------------------------------------------------------------------
-c
-c   routine name       - trian_CylTri
-c
-c-----------------------------------------------------------------------
-c
-c   latest revision    - Mar 2023
-c
-c   purpose            - routine evaluates physical coordinates
-c                        and its derivatives wrt to reference
-c                        coordinates for a point on the image
-c                        of a linear triangle through a global system
-c                        of coordinates: x,y=rcos(\theta),z=rsin(\theta)
-c                        and their  derivative wrt to reference
-c                        coordinates
-c
-c   arguments :
-c     in:
-c               No     - a GMP triangle number
-c               Eta    - reference coordinates of a point
-c                        in the triangle
-c     out:
-c               X      - physical coordinates of the point
-c               Dxdeta - derivatives of the physical coordinates wrt
-c                        to the parameters
-c
-c-----------------------------------------------------------------------
-c
+!-----------------------------------------------------------------------
+!
+!   routine name       - trian_CylTri
+!
+!-----------------------------------------------------------------------
+!
+!   latest revision    - Mar 2023
+!
+!   purpose            - routine evaluates physical coordinates
+!                        and its derivatives wrt to reference
+!                        coordinates for a point on the image
+!                        of a linear triangle through a global system
+!                        of coordinates: x,y=rcos(\theta),z=rsin(\theta)
+!                        and their  derivative wrt to reference
+!                        coordinates
+!
+!   arguments :
+!     in:
+!               No     - a GMP triangle number
+!               Eta    - reference coordinates of a point
+!                        in the triangle
+!     out:
+!               X      - physical coordinates of the point
+!               Dxdeta - derivatives of the physical coordinates wrt
+!                        to the parameters
+!
+!-----------------------------------------------------------------------
+!
       subroutine trian_CylTri(No,Eta, X,Dxdeta)
-c
+!
       use GMP
-c
+!
       implicit none
-c
+!
       integer :: No
       real(8) :: Eta(2),X(NDIM),Dxdeta(NDIM,2)
-c
-c  ...cylindrical coordinates of the endpoints of the triangle
+!
+!  ...cylindrical coordinates of the endpoints of the triangle
       real(8) :: xp(3,3)
-c
+!
       integer :: iv,np
       real(8) :: costhet,dr_deta1,dr_deta2,dthet_deta1,dthet_deta2
       real(8) :: pi,twopi,r,sinthet,theta
-c
+!
       integer :: iprint
       iprint=0
-c
+!
       if ((TRIANGLES(No)%Type.ne.'CylTri'.or.(NDIM.ne.3))) then
         write(*,7001) TRIANGLES(No)%Type
  7001   format('tria_CylTri: WRONG TRIANGLE TYPE = ',a10)
         stop 1
       endif
-c
+!
       if (iprint.eq.1) then
         write(*,7002) No,Eta
  7002   format('trian_CylTri: No,Eta = ',i4,2x,2f8.3)
       endif
-c
+!
       pi = acos(-1.d0)
[…]
