+++ trunk/src/GMP/base/curves/curve_HermCur.F90
@@ -1,58 +1,58 @@
-c----------------------------------------------------------------------
-c
-c   routine name       - curve_HermCur
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Mar 2023
-c
-c   purpose            - Hermite parametrization curve
-c
-c   arguments :
-c     in:
-c               No     - the curve number
-c               Eta    - reference coordinate  (between 0 and 1)
-c     out:
-c               X      - physical coordinates of the point
-c               Dxdeta - derivatives of the physical coordinates wrt
-c                        reference coordinate
-c
-c----------------------------------------------------------------------
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - curve_HermCur
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Mar 2023
+!
+!   purpose            - Hermite parametrization curve
+!
+!   arguments :
+!     in:
+!               No     - the curve number
+!               Eta    - reference coordinate  (between 0 and 1)
+!     out:
+!               X      - physical coordinates of the point
+!               Dxdeta - derivatives of the physical coordinates wrt
+!                        reference coordinate
+!
+!----------------------------------------------------------------------
+!
       subroutine curve_HermCur(No,Eta, X,Dxdeta)
-c
+!
       use GMP
       use parameters , only : MAXP
-c
+!
       implicit none
-c
+!
       integer :: No
       real(8) :: Eta,X(3),Dxdeta(3)
-c
-c  ...curve dof
+!
+!  ...curve dof
       real(8) :: gdof(1:3,6)
-c
-c  ...fifth order C^2-conforming shape functions
+!
+!  ...fifth order C^2-conforming shape functions
       real(8) :: vshap(MAXP+1),dvshap(MAXP+1),ddvshap(MAXP+1)
-c
+!
       integer :: i,k,np,nord
-c
+!
       integer :: iprint
       iprint=0
-c
+!
       if ((CURVES(No)%Type.ne.'HermCur').or.(NDIM.ne.3)) then
         write(*,7001)
  7001   format('curve_HermCur: WRONG INPUT')
         stop 1
       endif
-c
-c  ...get the curve endpoints
+!
+!  ...get the curve endpoints
       do i=1,2
         np = CURVES(No)%EndPoNo(i)
         call pointr(np, gdof(1:3,i))
       enddo
-c
-c  ...get the tangent vectors and second derivatives
+!
+!  ...get the tangent vectors and second derivatives
       gdof(1:3,3) = CURVES(No)%Rdata(1:3)
       gdof(1:3,4) = CURVES(No)%Rdata(4:6)
       gdof(1:3,5) = CURVES(No)%Rdata(7:9)
@@ -65,11 +65,11 @@ c  ...get the tangent vectors and second derivatives
         write(*,7003) (gdof(1:3,k),k=5,6)
  7003   format(6(3f8.3,2x))
       endif
-c
-c  ...determine C^2 polynomials
+!
+!  ...determine C^2 polynomials
       nord=5
       call Gshape1(nord,Eta, vshap,dvshap,ddvshap)
-c
+!
       X(1:3) = 0.d0; Dxdeta(1:3) = 0.d0
       do k=1,6
         X(1:3) = X(1:3) + gdof(1:3,k)*vshap(k)
@@ -81,6 +81,6 @@ c
         write(*,7003) X(1:3), Dxdeta(1:3)
         call pause
       endif
-c
-c
+!
+!
       end subroutine curve_HermCur

[…]
