+++ trunk/src/GMP/base/block_hexa.F90
@@ -296,7 +296,7 @@ end subroutine hexa_TraHex
 !
 !-----------------------------------------------------------------------
 !
-      subroutine hexa_CylHex(No,Eta, X,Dxdeta)
+   subroutine hexa_CylHex(No,Eta, X,Dxdeta)
 !
       use control
       use GMP          , only : HEXAS,POINTS,NDIM
@@ -394,4 +394,4 @@ end subroutine hexa_TraHex
       endif
 !
 !
-      end subroutine hexa_CylHex
+   end subroutine hexa_CylHex
