+++ trunk/src/GMP/base/triangles/trian_CylTri.F90
@@ -26,7 +26,7 @@
 !
 !-----------------------------------------------------------------------
 !
-      subroutine trian_CylTri(No,Eta, X,Dxdeta)
+   subroutine trian_CylTri(No,Eta, X,Dxdeta)
 !
       use GMP
 !
@@ -97,4 +97,4 @@
       Dxdeta(3,2) = dr_deta2*sinthet + r*costhet*dthet_deta2
 !
 !
-      end subroutine trian_CylTri
+   end subroutine trian_CylTri
