### 2024-03-01 d2b3b25
Update legacy code in frontal solver (#179)
