+++ trunk/src/solver/frontal/base/complex/bckwrd.F
@@ -1,366 +1,0 @@
-c***===***===***===***===***===***===***===***===***===***===***===***==
-c FUNCTION:
-c     Backsubstitution
-c     -----------------
-c     calls SOLIN1 for dest. vectors
-c     passes Elemental solutions to SOLOUT
-**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==**==
-c ARGUMENTS:  (I : input, O : output, IO : input & output, W : workspace
-c
-c Typ Name      Function
-c W  Alelm   final length of the current front as we process each Elem
-c W  Aldest  nodal destination vectors
-c W  Amdest  dof destination vectors
-c W  Andest  Destination in the front where the remaining values
-c            in the front will transfer to as each exiting dof is elimin
-c
-c W  Elem    Temp storage for the holding the Element solution
-c             (ie: used to transfer x from the solution front to the
-c                  local Element numbering based on the dest vecs)
-c
-c W  Frnt    Temp storage for the solutions
-c            Note: on entry, this contains the previous values in the fr
-c                   the values at the end of sub.frwcp or sub.frwrs
-c                   but this info isnt used for anything
-c
-c W  Ubuf    The eliminated lhs buffered equations
-c W  Bbuf    The eliminated rhs buffered equations
-c
-c*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*
-c LATEST REVISION: Mar 2023
-c++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++==++=
-c NAMING CONVENTIONS:
-c     AAAAAAAA    Variables in COMMON & PARAMETERS
-c     Aaaaaaaa    Variables as ARGUMENTS
-c     aaaaaaaa    LOCAL Variables
-c         7xxx    FORMAT Statements
-c         9xxx    ERROR Handling
-c+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+
-c
-      subroutine bckwrd (Alelm, Aldest, Amdest,
-     .                   Andest, Elem, Frnt, Bbuf, Ubuf)
-c
-      use surfsc1
-      use surfsc2
-c
-      implicit none
-c
-      complex(8) :: Aldest(*), Amdest(*), Andest(*), Elem(*),
-     .              Frnt(*), Bbuf(*), Ubuf(*), Alelm(*)
-c
-      integer :: iuu,jel,i,ie,iel,ifbi,ifui,ill
-      integer :: irout4,irout5,iuuold,j,jerr,k,l,lenx
-      integer :: md,n,ne,negiel,numdes
-c
-c  ...test......test......test......test......test......test......test...
-ccc      integer :: ndest1(100)
-c  ...test......test......test......test......test......test......test...
-
-c
-#if HP3D_DEBUG
-      integer :: iprint
-      iprint=0
-#endif
-c
-      iuu = 1
-      jel = NUMELM + 1
-c
-      IB = 0
-      ifui = IFU
-      ifbi = IFB
-c
-      IFU = IFU + 1
-      IFB = IFB + 1
-
-c
-c loop over the Elements (we are actually looping thru them backwards)
-c **********************
-      do 200 iel = 1,NUMELM
-c     *******************
-c set to loop backwards thru the Elems
-
-c
-         jel = jel - 1
-c
-         negiel = -jel
-c
-c pull up the Elements nodal destination vectors (Aldest)
-c
-
-         call solin1 (jel, numdes, Aldest)
[…]
