### 2024-03-01 efef205
Update legacy code in GMP (#180)
