+++ trunk/src/GMP/graphics/object3.F90
@@ -1,106 +1,106 @@
 #if HP3D_USE_X11
 
-c----------------------------------------------------------------------
-c
-c   routine name       - object3
-c
-c----------------------------------------------------------------------
-c
-c   latest revision    - Mar 2023
-c
-c   purpose            - routine displays a 3D manifold
-c
-c   arguments :
-c     in:
-c               Iwind  - window number
-c
-c----------------------------------------------------------------------
-c
-      subroutine object3(Iwind)
-c
+!----------------------------------------------------------------------
+!
+!   routine name       - object3
+!
+!----------------------------------------------------------------------
+!
+!   latest revision    - Mar 2023
+!
+!   purpose            - routine displays a 3D manifold
+!
+!   arguments :
+!     in:
+!               Iwind  - window number
+!
+!----------------------------------------------------------------------
+!
+   subroutine object3(Iwind)
+!
       use GMP
       use graphmod
 
       implicit none
-c
+!
       integer :: Iwind
-c
+!
       real(8) :: xcl(3),xxy(2)
       integer :: igv(20),im(2)
-c
+!
       integer :: i,ibl,iblno,ichoice,icount,inickbl,ivis,loc
       integer :: nel,nh,nick,nickbl,npri,npyr,nrsub1,ntet
       real(8) :: psi,q,q1,rnsc,theta
-c
+!
       real(8) :: bigp,bign,small,pi
       data bigp,bign,small /1.d30,-1.d30,1.d-14/
       data pi /3.14159265358979312d0/
-c
-c  ...set initial clipping plane for sections
+!
+!  ...set initial clipping plane for sections
       CLPL(1) = 0.d0
       CLPL(2) = 0.d0
       CLPL(3) = 0.d0
       CLPL(4) = 0.d0
-c
-c  ...initialize array of invisible blocks
+!
+!  ...initialize array of invisible blocks
       NRINVBL = 0
-ccc      IGINV(1) = 0
-c
-c  ...initiate the number of curvilinear blocks
+!!!      IGINV(1) = 0
+!
+!  ...initiate the number of curvilinear blocks
       NRCURVBL=0
-ccc      eta(1:2) = 0.5d0
-ccc      do nr=1,NRRECTA
-ccc        if (RECTANGLES(nr)%BlockNo(2).eq.0) then
-ccc          call recta_linear(nr,eta, x,dxdeta)
-ccc          if (x(3).gt.-29.d0) then
-ccc            NRCURVBL = NRCURVBL + 1
-ccc            NLINBLOCKS(NRCURVBL) = abs(RECTANGLES(nr)%BlockNo(1))
-ccc          endif
-ccc        endif
-ccc      enddo
-c
-c  ...set initial scaling constants
+!!!      eta(1:2) = 0.5d0
+!!!      do nr=1,NRRECTA
+!!!        if (RECTANGLES(nr)%BlockNo(2).eq.0) then
+!!!          call recta_linear(nr,eta, x,dxdeta)
+!!!          if (x(3).gt.-29.d0) then
+!!!            NRCURVBL = NRCURVBL + 1
+!!!            NLINBLOCKS(NRCURVBL) = abs(RECTANGLES(nr)%BlockNo(1))
+!!!          endif
+!!!        endif
+!!!      enddo
+!
+!  ...set initial scaling constants
       DIMIM = bigp
       XCIM(1) = 0.d0
[…]

[…]
