### 2024-03-01 3677ef7
removing remaining common blocks in graphics/x11
