+++ trunk/src/graphics/x11/module/wvscrn.F90
@@ -1,0 +1,7 @@
+!>@brief This module replaces common file "wvscrn.blk"
+!>@date  Feb 2024
+module wvscrn
+
+   integer :: MAXSCX,MAXSCY,MAXCOL,MAXDFX,MAXDFY
+
+end module wvscrn
