+++ trunk/problems/SHEATHED_HOSE/exact_error.F90
@@ -44,9 +44,9 @@ subroutine exact_error
 !
   select case(IERROR_PROB)
   case(IERROR_L2)
-    L2PROJ = .TRUE.
+    L2PROJ = .true.
   case(IERROR_NATURAL)
-    L2PROJ = .FALSE.
+    L2PROJ = .false.
   case default
     write(*,*) 'exact_error : Error calculation type not supported'
   end select
@@ -172,7 +172,7 @@ subroutine exact_error
 !     raise visitation flag
       ivis=ivis+1
 !
-IF (.NOT. QUIET_MODE) THEN
+if (.not. QUIET_MODE) then
 !
 !     check
       if (ivis > maxvis) then
@@ -189,7 +189,7 @@ IF (.NOT. QUIET_MODE) THEN
       rwork(ivis, 6)=rateHEVQ
       iwork(ivis, 1)=IERROR_ATTR
 !
-ENDIF
+endif
 !
 !     printing
 !
@@ -204,11 +204,11 @@ ENDIF
         endif
 !
 !       print header
-IF (.NOT. L2PROJ) THEN
+if (.not. L2PROJ) then
         write(nin,*)'-- Error Report --'
 ELSE
         write(nin,*)'-- Error Report (L2 only)--'
-ENDIF
+endif
         write(nin,9998)
  9998   format('             H1            //', &
                            ' H(curl)       //', &
@@ -234,15 +234,15 @@ ENDIF
  9999 format(1x,i6,' ; ',2x,5(e12.5,' ; ',2x),f9.6,' ; ',3x,i8)
 !
 !     print to screen
-IF (.NOT.QUIET_MODE) THEN ; write(*,*)''
-IF (.NOT. L2PROJ   ) THEN ; write(*,*)'-- Error Report --'
+if (.not.QUIET_MODE) then ; write(*,*)''
+if (.not. L2PROJ   ) then ; write(*,*)'-- Error Report --'
 ELSE                      ; write(*,*)'-- Error Report (L2 only)--'
-ENDIF
+endif
                             write(*,9998)
       do i=1,ivis         ; write(*,9999)i,rwork(i,1:6),iwork(i,1)
       enddo
                             write(*,*)''
-ENDIF
+endif
 !
 !     close file
       close(unit=nin,iostat=ic)
