+++ trunk/src/meshmod/nodmod.F90
@@ -20,7 +20,7 @@
 !
 !-----------------------------------------------------------------------
 !
-      subroutine nodmod(Nod,Nordn)
+   subroutine nodmod(Nod,Nordn)
 !
       use data_structure3D
       use mpi_param, only: RANK
@@ -272,7 +272,7 @@
 #endif
 !
 !
-      end subroutine nodmod
+   end subroutine nodmod
 !
 !---------------------------------------------------------------------
 !
@@ -295,7 +295,7 @@
 !
 !-----------------------------------------------------------------------
 !
-      subroutine copy_dofG(Ntype,Nordo,Nordn,NdofGo,NdofGn, &
+   subroutine copy_dofG(Ntype,Nordo,Nordn,NdofGo,NdofGn, &
                            Xnodo,Xnodn)
 !
       use node_types
@@ -356,7 +356,7 @@
 #endif
 !
 !
-      end subroutine copy_dofG
+   end subroutine copy_dofG
 !
 !---------------------------------------------------------------------
 !
@@ -380,7 +380,7 @@
 !
 !-----------------------------------------------------------------------
 !
-      subroutine copy_dofH(Ntype,Nordo,Nordn,NdofHo,NdofHn, &
+   subroutine copy_dofH(Ntype,Nordo,Nordn,NdofHo,NdofHn, &
                            NvarH,ZdofHo,ZdofHn)
 !
       use node_types
@@ -441,7 +441,7 @@
 #endif
 !
 !
-      end subroutine copy_dofH
+   end subroutine copy_dofH
 !
 !---------------------------------------------------------------------
 !
@@ -465,7 +465,7 @@
 !
 !-----------------------------------------------------------------------
 !
-      subroutine copy_dofE(Ntype,Nordo,Nordn,NdofEo,NdofEn, &
+   subroutine copy_dofE(Ntype,Nordo,Nordn,NdofEo,NdofEn, &
                            NvarE,ZdofEo,ZdofEn)
 !
       use node_types
@@ -636,7 +636,7 @@
 #endif
 !
 !
-      end subroutine copy_dofE
+   end subroutine copy_dofE
 !
 !---------------------------------------------------------------------
 !
@@ -660,7 +660,7 @@
 !
 !-----------------------------------------------------------------------
 !
-      subroutine copy_dofV(Ntype,Nordo,Nordn,NdofVo,NdofVn, &
+   subroutine copy_dofV(Ntype,Nordo,Nordn,NdofVo,NdofVn, &
                            NvarV,ZdofVo,ZdofVn)
 !
       use node_types
@@ -824,7 +824,7 @@
 #endif
 !
 !
-      end subroutine copy_dofV
+   end subroutine copy_dofV
 !
 !---------------------------------------------------------------------
 !
@@ -848,7 +848,7 @@
 !
 !-----------------------------------------------------------------------
 !
-      subroutine copy_dofQ(Ntype,Nordo,Nordn,NdofQo,NdofQn, &
+   subroutine copy_dofQ(Ntype,Nordo,Nordn,NdofQo,NdofQn, &
                            NvarQ,ZdofQo,ZdofQn)
 !
       use node_types
@@ -912,11 +912,11 @@
 #endif
 !
 !
-      end subroutine copy_dofQ
+   end subroutine copy_dofQ
 !
 !-----------------------------------------------------------------------
 !
-      subroutine copy_1array(M,A,Ia, B,Ib)
[…]
