+++ trunk/src/element/shape_1/TraceSequence.F90
@@ -120,7 +120,7 @@
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 !     Check polynomial order
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-     subroutine checkpolyorder(Norder)
+      subroutine checkpolyorder(Norder)
 !
         integer, intent(in) :: Norder
 !
@@ -132,7 +132,7 @@
           stop 1
         endif
 !
-     end subroutine checkpolyorder
+      end subroutine checkpolyorder
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    end subroutine TraceEshapeH
 !
@@ -237,7 +237,7 @@
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 !     Check polynomial order
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-     subroutine checkpolyorder(Norder)
+      subroutine checkpolyorder(Norder)
 !
         integer, intent(in) :: Norder
 !
@@ -249,7 +249,7 @@
           stop 1
         endif
 !
-     end subroutine checkpolyorder
+      end subroutine checkpolyorder
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    end subroutine TraceEshapeE
 !
@@ -386,7 +386,7 @@
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 !     Triangle traces
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-     subroutine traceTriFshapeH
+      subroutine traceTriFshapeH
 !
         integer :: i,j,nij,m,N,ndofF
         integer :: minI,minJ,minIJ,maxI,maxJ,maxIJ
@@ -459,11 +459,11 @@
         endif
 #endif
 !
-     end subroutine traceTriFshapeH
+      end subroutine traceTriFshapeH
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 !     Quadrilateral traces
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-     subroutine traceQuadFshapeH
+      subroutine traceQuadFshapeH
 !
         integer :: N,m,i,j,nordF(2),ndofF,minI,minJ,maxI,maxJ
         logical :: IdecF(2),GIdecF(2)
@@ -536,11 +536,11 @@
         endif
 #endif
 !
-     end subroutine traceQuadFshapeH
+      end subroutine traceQuadFshapeH
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 !     Check polynomial order
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-     subroutine checkpolyorder(Norder)
+      subroutine checkpolyorder(Norder)
 !
         integer, intent(in) :: Norder
 !
@@ -552,7 +552,7 @@
           stop 1
         endif
 !
-     end subroutine checkpolyorder
+      end subroutine checkpolyorder
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    end subroutine TraceFshapeH
 !
@@ -611,7 +611,7 @@
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 !     Triangle traces
 ! - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
[…]
