### 2024-03-01 4edcfc1
make some headers more consistent with doxygen
