+++ trunk/problems/POISSON/ULTRAWEAK_DPG/common/exact.F90
@@ -6,7 +6,7 @@
 !
 !     latest revision:  - May 2023
 !
-!     purpose:          - return exact solution value at a point X
+!> @brief         - return exact solution value at a point X
 !
 !     arguments:
 !        in:
