+++ trunk/problems/MAXWELL/ULTRAWEAK_DPG/common/set_environment.F90
@@ -6,7 +6,7 @@
 !
 !     latest revision:  - Sep 2023
 !
-!     purpose:          - define options for main file specific to the problem.
+!> @brief         - define options for main file specific to the problem.
 !                         These can be consulted with the -help option when running
 !                         the executable. It is especially important to define
 !                         the global environment variables in module/environment.
