+++ trunk/src/element/shape_1/Hexahedron.F90
@@ -11,7 +11,7 @@
 !
 !     latest revision:  - Oct 14, Apr 17, Jul 21
 !
-!     purpose:          - routine returns values of 3D hexahedron element
+!> @brief         - routine returns values of 3D hexahedron element
 !                         H1 shape functions and their derivatives
 !
 !     arguments:
@@ -218,7 +218,7 @@
 !
 !     latest revision:  - Oct 14, Apr 17, Jul 21
 !
-!     purpose:          - routine returns values of 3D hexahedron element
+!> @brief         - routine returns values of 3D hexahedron element
 !                         H(curl) shape functions and their derivatives
 !
 !     arguments:
@@ -447,7 +447,7 @@
 !
 !     latest revision:  - Oct 14, Apr 17, Jul 21
 !
-!     purpose:          - routine returns values of 3D hexahedron element
+!> @brief         - routine returns values of 3D hexahedron element
 !                         H(div) shape functions and their divergences
 !
 !     arguments:
@@ -615,7 +615,7 @@
 !
 !     latest revision:  - Oct 14, Apr 17
 !
-!     purpose:          - routine returns values of 3D hexahedron
+!> @brief         - routine returns values of 3D hexahedron
 !                         element L2 shape functions
 !
 !     arguments:

[…]
