### 2024-03-01 2339aa8
Remove legacy code common blocks (#181)
* remove obsolete TODO file

* delete obsolete syscom.blk file

* removing remaining common blocks in graphics/x11
