+++ trunk/src/graphics/x11/module/rastcom.F90
@@ -1,0 +1,7 @@
+!>@brief This module replaces common file "rastcom.blk"
+!>@date  Feb 2024
+module rastcom
+
+   integer :: IFILE,IDUMP,LASTCOL,LASTX,LASTY
+
+end module rastcom
