### 2024-03-01 9db1178
Make style more consistent, mostly in legacy code (#182)
* style: remove keyword capitalization

* fix some fixed-form indentation

* make some headers more consistent with doxygen

* update CHANGELOG

* minor
