+++ trunk/problems/SHEATHED_HOSE/module/common_prob_data.F90
@@ -1,5 +1,5 @@
-!> Purpose : Define problem specific parameters and constants
-!! @rev - Feb.2012
+!> @brief Define problem specific parameters and constants
+!> @rev - Feb.2012
 !------------------------------------------------------------------------------
 module common_prob_data
 !
@@ -71,10 +71,10 @@ module common_prob_data
 
 ! control variables/parameters of geometrical transformation on
 ! point coordinates
-  logical :: COORD_TRANS        = .FALSE.
-  logical :: COORD_TRANS_TRAS   = .FALSE.
-  logical :: COORD_TRANS_ROT    = .TRUE.
-  logical :: COORD_TRANS_SCAL   = .FALSE.
+  logical :: COORD_TRANS        = .false.
+  logical :: COORD_TRANS_TRAS   = .false.
+  logical :: COORD_TRANS_ROT    = .true.
+  logical :: COORD_TRANS_SCAL   = .false.
   double precision, dimension(3):: TRAS_VECTOR = (/1.d0,1.d0,1.d0/)
   double precision, dimension(3):: ROT_ANGLES = (/10.d0,45.d0,60.d0/)
 
