+++ trunk/src/solver/frontal/base/real/desvecR.F90
@@ -77,19 +77,19 @@
 !
 ! initialize the destination vector storage
 !
-      do 10 i = 1,NFN
+      do i = 1,NFN
          Ibb(i) = 0
-   10 continue
+      enddo
 !
 ! loop thru the elements
 !
-      do 100 iel = 1,NUMELM
+      do iel = 1,NUMELM
 !     **********************
 !
 ! if the element contaions no nodes or multipliers, then skip it
 !
          n = In(iel)
-         if (n .eq. 0) go to 100
+         if (n .eq. 0) goto 100
 !
 ! initialize
 !
@@ -101,7 +101,7 @@
 !
 ! loop thru the nicknames
 !
-         do 60 id = 1,n
+         do id = 1,n
 !        ***************
 !
             ifrst = 0
@@ -148,7 +148,7 @@
 !
 ! loop thru all the remaining nicknames and search for the next matching
 ! ----------------------------------------------------------------------
-               do 40 jjd = jp,nfn
+               do jjd = jp,nfn
 !              ******************
                   jd = jjd
 !
@@ -169,10 +169,10 @@
 !
 ! skip to the next nickname
 !
-                     go to 60
+                     goto 60
                   endif
 !
-   40          continue
+               enddo
 !  ********************
             endif
 !
@@ -198,7 +198,7 @@
 !
             ntt = ntt + NDOFM
 !
-   60    continue
+   60    enddo
 !  ***************
 !
 ! increment the total number of dof for the model
@@ -225,7 +225,7 @@
 !
 ! search thru the remaining nicknames
 !
-               do 80 jd = jp,jdn
+               do jd = jp,jdn
 !
 ! check if this nickname was given a front dest already
 !
@@ -236,14 +236,14 @@
 !
 ! look thru those front dests which are being eliminated
 !
-                     do 70 i = 1,ne
+                     do i = 1,ne
 !
 ! if the front dest of this nickname is ge than the one we are eliminati
 !  then increment the number of dof by that being eliminated
 !
                         if (Ibb(jd) .ge. Ic(ipc))  nt = nt + Ic(ipc+1)
                         ipc = ipc + 2
-   70                continue
+                     enddo
 !
 ! reduce the front dest of this nickname by the number of dof eliminated
 !  below it in the front
@@ -251,7 +251,7 @@
                      Ibb(jd) = Ibb(jd) - nt
                   endif
 !
-   80          continue
+               enddo
             endif
          endif
 !
@@ -268,7 +268,7 @@
                     (/,10x,10i8))
          endif
 !
-  100 continue
+  100 enddo
 !
 !
    end subroutine desvec
