+++ trunk/src/constrs/util/hcurl/setcnstr_edge_hcurl.F90
@@ -259,7 +259,7 @@
 !
       write(*,*) 'setcnstr_edge_hcurl: CONTINUE ?(1/0)'
       read(*,*) ians
-      if (ians.eq.1) go to 10
+      if (ians.eq.1) goto 10
 !
 #endif
 !
