+++ trunk/src/hpinterp/edge/dhpedgeH.F90
@@ -131,7 +131,7 @@ subroutine dhpedgeH(Mdle,Iflag,No,Etav,Ntype,Icase,Bcond,   &
   call homogenD(CONTIN,Icase,Bcond, is_homD,ncase,ibcnd)
   if (is_homD) then
     zuH = ZERO
-    go to 100
+    goto 100
   endif
 !
 ! if # of dof is zero, return, nothing to do

[…]
