### 2024-03-01 bbea5cd
removed obsolete common blocks from frontal solver
