+++ trunk/src/solver/frontal/base/complex/zdiriodum.F90
@@ -31,6 +31,7 @@
 !
    subroutine zdiriodum (Unname, Commnd, Irec, Len, Sbuf, Jerr)
 !
+      use czdirio
       use surfsc1
 !
       implicit none
@@ -43,13 +44,7 @@
       integer    :: iend,inow,ioffst,iostat,irecp,irecsv,iunit
       integer    :: lenop,lenr,lenw,ntape
       complex(8) :: slen
-!
-      integer    :: nbuf(9), lenf(9)
-      integer    :: lbuf(9), irsave(9)
-      complex(8) :: storage(100)
-!
-      common /czdirio1/ nbuf, lenf, irsave
-      common /czdirio2/ storage
+      integer    :: lbuf(9)
 !
 ! the following data statement is machine dependant
 ! ...................................................
