### 2024-03-01 2f109de
resolve some minor issues (f2018 warnings)
