+++ trunk/problems/LASER/UW_COUPLED/elem/elem_residual_maxwell.F90
@@ -163,6 +163,7 @@ subroutine elem_residual_maxwell(Mdle,Fld_flag,          &
    integer :: dom_flag
 !
    integer, external :: ij_upper_to_packed
+   logical, external :: dnear
 !
 !..timer
    !real(8) :: start_time,end_time
@@ -361,7 +362,7 @@ subroutine elem_residual_maxwell(Mdle,Fld_flag,          &
          endif
 !
 !     ...update background polarization with thermal perturbation
-         if(delta_n .ne. 0.d0) then
+         if(.not. dnear(delta_n,0.d0)) then
             call get_bgPol(dom_flag,Fld_flag,delta_n,x, bg_pol)
          endif
 !
