+++ trunk/problems/LASER/UW_COUPLED/elem/elem_maxwell_gram_hexa.F90
@@ -147,6 +147,7 @@
    real(8), dimension(MAXPP+1,MAXPP+1) :: sH2p,sH3p,dsH2p,dsH3p
 !
    integer, external :: ij_upper_to_packed
+   logical, external :: dnear
 !
    integer :: deltak(3,3)
 !
@@ -434,7 +435,7 @@
                endif
 !
 !           ...update background polarization with thermal perturbation
-               if(delta_n .ne. 0.d0) then
+               if(.not. dnear(delta_n,0.d0)) then
                   call get_bgPol(dom_flag,Fld_flag,delta_n,x, bg_pol)
                endif
 !
