+++ trunk/problems/LASER/UW_COUPLED/elem/elem_maxwell_gram_pris.F90
@@ -143,6 +143,7 @@
    integer, allocatable :: mapEE(:)
 !
    integer, external :: ij_upper_to_packed
+   logical, external :: dnear
 !
    integer, dimension(3,3) :: deltak
 !
@@ -438,7 +439,7 @@
             endif
 !
 !        ...update background polarization with thermal perturbation
-            if(delta_n .ne. 0.d0) then
+            if(.not. dnear(delta_n,0.d0)) then
                call get_bgPol(dom_flag,Fld_flag,delta_n,x, bg_pol)
             endif
 !
