### 2024-03-01 #5946 @jedbrown
Thanks for completing this. It looks good to me apart from some editorial details:

1. Use in-text `@citations` when the parentheses should be removed.
```diff
diff --git i/paper/paper.md w/paper/paper.md
index db63ec2..ed130f3 100644
--- i/paper/paper.md
+++ w/paper/paper.md
@@ -58,7 +58,7 @@ Some examples of 3D applications that have been implemented in $hp\mathrm{3D}$ i

 # Further reading

-Instructions on installing and using the code are available in the $hp\mathrm{3D}$ user manual [@hpUserManual]. $hp\mathrm{3D}$'s underlying algorithms are described in various published articles and books. Many parts of the current version of the $hp\mathrm{3D}$ software are based on the algorithms described in the two-volume $hp$ book series [@hpbook; @hpbook2] on the former 2D and 3D versions of the code (which were not published as open-source libraries). A third $hp$ book volume detailing the additions and modifications of the newest version, including its MPI/OpenMP parallel algorithms, will be published soon [@hpbook3]. The orientation-embedded shape functions package is described in [@fuentes2015shape]. Details on the FE methodology and conforming discretization of exact-sequence elements are given in [@demkowicz2023fem].
+Instructions on installing and using the code are available in the $hp\mathrm{3D}$ user manual [@hpUserManual]. $hp\mathrm{3D}$'s underlying algorithms are described in various published articles and books. Many parts of the current version of the $hp\mathrm{3D}$ software are based on the algorithms described in the two-volume $hp$ book series [@hpbook; @hpbook2] on the former 2D and 3D versions of the code (which were not published as open-source libraries). A third $hp$ book volume detailing the additions and modifications of the newest version, including its MPI/OpenMP parallel algorithms, will be published soon [@hpbook3]. The orientation-embedded shape functions package is described in @fuentes2015shape. Details on the FE methodology and conforming discretization of exact-sequence elements are given in @demkowicz2023fem.

 # Acknowledgements

```
2. The `journal` field should not be abbreviated according to APA style. Please update to spell out all those words (sorry).
3. You can use the [current `petsc-user-ref`](https://petsc.org/release/#citing-petsc), which has a DOI.

Once you've addressed those, you can proceed to:
