### 2024-03-02 73bac34
Update some code to `-std=f2018` (#184)
* removed labeled do statements

* remove obsolete common blocks in meshmod

* removed obsolete common blocks from frontal solver

* resolve some minor issues (f2018 warnings)
