+++ trunk/src/solver/frontal/base/real/frwcpR.F90
@@ -58,7 +58,7 @@
 !
 ! loop over the Elements
 ! **********************
-      do 200 iel=1,NUMELM
+      do iel=1,NUMELM
 !     ********************
 !
          negiel = -iel
@@ -67,7 +67,7 @@
 !
          call solin1 (iel, numdes, Aldest)
 !        ------------------------------
-         if(numdes .eq. 0) go to 200
+         if(numdes .eq. 0) goto 200
 !
 ! convert the nodal destination vectors to dof destination vectors (Amde
 !  also determine the number of dof to eliminate from the front (ne)
@@ -140,9 +140,9 @@
 ! ARDENT directives
 !$doit VBEST
 !
-            do 133 id = 1,NDOFM
+            do id = 1,NDOFM
                Amdest(id) = dabs(Amdest(id))
-  133       continue
+            enddo
 !wb <
 !
 ! symmetric
@@ -183,7 +183,7 @@
 ! eliminate those equations possible from the current front
 ! **********************************************************
 !
-         do 150 ie = 1,ne
+         do ie = 1,ne
 !        ****************
 !
 ! check if the equation Buffers are going to get full
@@ -341,11 +341,11 @@
 !$doit VBEST
 !
 !
-               do 110 j = 1,n
+               do j = 1,n
                   Buf(IL) = Frnt(m)
                   IL = IL - 1
                   m = m + KFW
-  110          continue
+               enddo
 !
             endif
 !
@@ -353,7 +353,7 @@
 !
             NFW = NFW - 1
 !           -------------
-  150    continue
+         enddo
 ! ***************
 ! store the final length of the current front (LFW) into Alelm
 !
@@ -373,7 +373,7 @@
             n = KFW
             m = NFW + 1
 !
-            do 170 i = 2,NFW
+            do i = 2,NFW
 !
 ! ALLIANT directives
 !vd$ select (vector)
@@ -381,17 +381,17 @@
 !$doit VBEST
 !
 !
-               do 160 j = 1,NFW
+               do j = 1,NFW
                   Frnt(m) = Frnt(n+j)
                   m = m + 1
-  160          continue
+               enddo
 !
                n = n + KFW
-  170       continue
+            enddo
 !
          endif
 !
-  200 continue
+  200 enddo
 ! ************
 !
 ! dump the final Buffers to diskfiles
