### 2024-03-02 ffe8045
some minor fixes
