+++ paper/paper.bib
@@ -2,7 +2,7 @@
 @article{henneking2021fiber,
   title = {Model and computational advancements to full vectorial {Maxwell} model for studying fiber amplifiers},
   author = {Henneking, Stefan and Grosek, Jacob and Demkowicz, Leszek},
-  journal = {Comput. Math. Appl.},
+  journal = {Computers \& Mathematics with Applications},
   volume = {85},
   pages = {30--41},
   year = {2021},
@@ -12,7 +12,7 @@
 @article{henneking2021pollution,
   Title = {A numerical study of the pollution error and {DPG} adaptivity for long waveguide simulations},
   Author = {Henneking, Stefan and Demkowicz, Leszek},
-  Journal = {Comput. Math. Appl.},
+  Journal = {Computers \& Mathematics with Applications},
   Publisher = {Elsevier},
   Volume = {95},
   Pages = {85--100},
@@ -38,7 +38,7 @@
 @article{badger2023scalable,
   title = {Scalable {DPG} Multigrid Solver for {Helmholtz} Problems: A Study on Convergence},
   author = {Badger, Jacob and Henneking, Stefan and Petrides, Socratis and Demkowicz, Leszek},
-  journal = {Comput. Math. Appl.},
+  journal = {Computers \& Mathematics with Applications},
   volume = {148},
   pages = {81--92},
   year = {2023},
@@ -82,7 +82,7 @@
 @article{petrides2021adaptive,
   title={An adaptive multigrid solver for {DPG} methods with applications in linear acoustics and electromagnetics},
   author={Petrides, Socratis and Demkowicz, Leszek},
-  journal={Comput. Math. Appl.},
+  journal={Computers \& Mathematics with Applications},
   volume={87},
   pages={12-26},
   year={2021},
@@ -91,7 +91,7 @@
 @article{amestoy2001mumps,
   Title = {A fully asynchronous multifrontal solver using distributed dynamic scheduling},
   Author = {Amestoy, P.R. and Duff, I.S. and L'Excellent, J.Y. and Koster, J.},
-  Journal = {SIAM J. Matrix Anal. Appl.},
+  Journal = {SIAM Journal on Matrix Analysis and Applications},
   Number = {1},
   Pages = {15--41},
   Publisher = {SIAM},
@@ -102,7 +102,7 @@
 @article{pt-scotch,
   title={{PT-Scotch}: a tool for efficient parallel graph ordering},
   author={Chevalier, C. and Pellegrini, F.},
-  journal={Parallel Comput.},
+  journal={Parallel Computing},
   volume={34},
   number={6-8},
   pages={318--331},
@@ -113,7 +113,7 @@
 @article{parmetis,
   title={A fast and high quality multilevel scheme for partitioning irregular graphs},
   author={Karypis, G. and Kumar, V.},
-  journal={SIAM J. Sci. Comput.},
+  journal={SIAM Journal on Scientific Computing},
   volume={20},
   number={1},
   pages={359--392},
@@ -127,13 +127,21 @@
   title = {Hierarchical data format version 5},
   url = {http://www.hdfgroup.org/HDF5}}
 
-@techreport{petsc-user-ref,
-  Author = {S. Balay and S. Abhyankar and M. Adams and J. Brown and P. Brune and K. Buschelman and L. Dalcin and A. Dener and V. Eijkhout and W. Gropp and D. Karpeyev and D. Kaushik and M. Knepley and D. May and L. McInnes and R. Mills and T. Munson and K. Rupp and P. Sanan and B. Smith and S. Zampini and H. Zhang and H. Zhang},
-  Institution = {Argonne National Laboratory},
[…]
