### 2024-03-02 11140f3
update CITATION file to release version 1.0
