+++ paper/paper.bib
@@ -52,6 +52,12 @@
   year = {2023},
   doi = {10.48550/arXiv.2309.00726}}
 
+@software{hp3D-v1.0,
+  Author = {Henneking, Stefan, and Petrides, Socratis, and Fuentes, Federico, and Badger, Jacob, and Demkowicz, Leszek},
+  Title = {hp3D: A Scalable MPI/OpenMP hp-Adaptive Finite Element Software Library for Complex Multiphysics Applications (v1.0)},
+  Year = {2024},
+  doi = {10.5281/zenodo.10763375}}
+
 @article{hpUserManual,
   Author = {Henneking, Stefan and Demkowicz, Leszek},
   Title = {{$hp$3D User Manual}},
