### 2024-03-02 #5946 @stefanhenneking
@jedbrown

I updated the paper as you suggested, tagged the current release version (tagged `v1.0` at https://github.com/Oden-EAG/hp3d/releases/tag/v1.0), and archived it on Zenodo.

The Zenodo DOI for the hp3D release `v1.0` is [10.5281/zenodo.10763375](https://zenodo.org/doi/10.5281/zenodo.10763375).

The all-versions DOI is [10.5281/zenodo.10763374](https://zenodo.org/doi/10.5281/zenodo.10763374).

I also added the Zenodo archive with DOI as a reference to the paper.
